Module ATDGenerated.Jsonbug_j

type taint_extra = Jsonbug_t.taint_extra = {

taint_source : string option;

taint_sink : string option;

taint_policy_privacy_effect : string option;

tainted_expression : string option;

}

val equal_taint_extra : taint_extra -> taint_extra -> bool

type redundant_fixme_info = Jsonbug_t.redundant_fixme_info = {

present_issue_types : string list;

comment_start_line : int;

comment_end_line : int;

}

val equal_redundant_fixme_info : redundant_fixme_info -> redundant_fixme_info -> bool

type issue_method = Jsonbug_t.issue_method = {

name : string;

params : string list;

}

val equal_issue_method : issue_method -> issue_method -> bool

type parameter_not_nullable_info = Jsonbug_t.parameter_not_nullable_info = {

class_name : string;

package_name : string option;

method_info : issue_method;

param_index : int;

}

val equal_parameter_not_nullable_info : parameter_not_nullable_info -> parameter_not_nullable_info -> bool

type nullsafe_mode = Jsonbug_t.nullsafe_mode

val equal_nullsafe_mode : nullsafe_mode -> nullsafe_mode -> bool

type nullsafe_meta_issue_info = Jsonbug_t.nullsafe_meta_issue_info = {

num_issues : int;

num_fixmes : int;

curr_nullsafe_mode : nullsafe_mode;

can_be_promoted_to : nullsafe_mode option;

}

val equal_nullsafe_meta_issue_info : nullsafe_meta_issue_info -> nullsafe_meta_issue_info -> bool

type method_info = Jsonbug_t.method_info = {

class_name : string;

package : string;

method_name : string;

call_line : int;

}

val equal_method_info : method_info -> method_info -> bool

type field_name = Jsonbug_t.field_name = {

class_name : string;

package_name : string option;

field : string;

}

val equal_field_name : field_name -> field_name -> bool

type access_level = Jsonbug_t.access_level

val equal_access_level : access_level -> access_level -> bool

type annotation_point_method = Jsonbug_t.annotation_point_method = {

method_name : string;

params : string list;

access_level : access_level;

}

val equal_annotation_point_method : annotation_point_method -> annotation_point_method -> bool

type annotation_point_kind = Jsonbug_t.annotation_point_kind

val equal_annotation_point_kind : annotation_point_kind -> annotation_point_kind -> bool

type annotation_point = Jsonbug_t.annotation_point = {

id : string;

kind : annotation_point_kind;

method_info : annotation_point_method option;

field_name : string option;

param_num : int option;

num_violations : int;

dependent_point_ids : string list;

}

val equal_annotation_point : annotation_point -> annotation_point -> bool

type nullsafe_extra = Jsonbug_t.nullsafe_extra = {

class_name : string;

package : string option;

method_info : issue_method option;

field : field_name option;

inconsistent_param_index : int option;

parameter_not_nullable_info : parameter_not_nullable_info option;

nullable_methods : method_info list option;

unvetted_3rd_party : string list option;

meta_issue_info : nullsafe_meta_issue_info option;

annotation_graph : annotation_point list option;

redundant_fixme_info : redundant_fixme_info option;

}

val equal_nullsafe_extra : nullsafe_extra -> nullsafe_extra -> bool

type loc = Jsonbug_t.loc = {

file : string;

lnum : int;

cnum : int;

enum : int;

}

val equal_loc : loc -> loc -> bool

type json_trace_item = Jsonbug_t.json_trace_item = {

level : int;

filename : string;

line_number : int;

column_number : int;

description : string;

}

val equal_json_trace_item : json_trace_item -> json_trace_item -> bool

type config_usage_extra = Jsonbug_t.config_usage_extra = {

config_name : string;

function_name : string;

filename : string;

line_number : int;

}

val equal_config_usage_extra : config_usage_extra -> config_usage_extra -> bool

type extra = Jsonbug_t.extra = {

cost_polynomial : string option;

cost_degree : int option;

nullsafe_extra : nullsafe_extra option;

copy_type : string option;

config_usage_extra : config_usage_extra option;

taint_extra : taint_extra option;

}

val equal_extra : extra -> extra -> bool

type jsonbug = Jsonbug_t.jsonbug = {

bug_type : string;

qualifier : string;

severity : string;

line : int;

column : int;

procedure : string;

procedure_start_line : int;

file : string;

bug_trace : json_trace_item list;

key : string;

node_key : string option;

hash : string;

dotty : string option;

infer_source_loc : loc option;

bug_type_hum : string;

traceview_id : int option;

censored_reason : string option;

access : string option;

extras : extra option;

}

val equal_jsonbug : jsonbug -> jsonbug -> bool

type report = Jsonbug_t.report

val equal_report : report -> report -> bool

val write_taint_extra : Stdlib.Buffer.t -> taint_extra -> unit

Output a JSON value of type taint_extra.

val string_of_taint_extra : ?len:int -> taint_extra -> string

Serialize a value of type taint_extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_extra

Input JSON data of type taint_extra.

val taint_extra_of_string : string -> taint_extra

Deserialize JSON data of type taint_extra.

val write_redundant_fixme_info : Stdlib.Buffer.t -> redundant_fixme_info -> unit

Output a JSON value of type redundant_fixme_info.

val string_of_redundant_fixme_info : ?len:int -> redundant_fixme_info -> string

Serialize a value of type redundant_fixme_info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_redundant_fixme_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> redundant_fixme_info

Input JSON data of type redundant_fixme_info.

val redundant_fixme_info_of_string : string -> redundant_fixme_info

Deserialize JSON data of type redundant_fixme_info.

val write_issue_method : Stdlib.Buffer.t -> issue_method -> unit

Output a JSON value of type issue_method.

val string_of_issue_method : ?len:int -> issue_method -> string

Serialize a value of type issue_method into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_issue_method : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> issue_method

Input JSON data of type issue_method.

val issue_method_of_string : string -> issue_method

Deserialize JSON data of type issue_method.

val write_parameter_not_nullable_info : Stdlib.Buffer.t -> parameter_not_nullable_info -> unit

Output a JSON value of type parameter_not_nullable_info.

val string_of_parameter_not_nullable_info : ?len:int -> parameter_not_nullable_info -> string

Serialize a value of type parameter_not_nullable_info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_parameter_not_nullable_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> parameter_not_nullable_info

Input JSON data of type parameter_not_nullable_info.

val parameter_not_nullable_info_of_string : string -> parameter_not_nullable_info

Deserialize JSON data of type parameter_not_nullable_info.

val write_nullsafe_mode : Stdlib.Buffer.t -> nullsafe_mode -> unit

Output a JSON value of type nullsafe_mode.

val string_of_nullsafe_mode : ?len:int -> nullsafe_mode -> string

Serialize a value of type nullsafe_mode into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_nullsafe_mode : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> nullsafe_mode

Input JSON data of type nullsafe_mode.

val nullsafe_mode_of_string : string -> nullsafe_mode

Deserialize JSON data of type nullsafe_mode.

val write_nullsafe_meta_issue_info : Stdlib.Buffer.t -> nullsafe_meta_issue_info -> unit

Output a JSON value of type nullsafe_meta_issue_info.

val string_of_nullsafe_meta_issue_info : ?len:int -> nullsafe_meta_issue_info -> string

Serialize a value of type nullsafe_meta_issue_info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_nullsafe_meta_issue_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> nullsafe_meta_issue_info

Input JSON data of type nullsafe_meta_issue_info.

val nullsafe_meta_issue_info_of_string : string -> nullsafe_meta_issue_info

Deserialize JSON data of type nullsafe_meta_issue_info.

val write_method_info : Stdlib.Buffer.t -> method_info -> unit

Output a JSON value of type method_info.

val string_of_method_info : ?len:int -> method_info -> string

Serialize a value of type method_info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_method_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> method_info

Input JSON data of type method_info.

val method_info_of_string : string -> method_info

Deserialize JSON data of type method_info.

val write_field_name : Stdlib.Buffer.t -> field_name -> unit

Output a JSON value of type field_name.

val string_of_field_name : ?len:int -> field_name -> string

Serialize a value of type field_name into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_field_name : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> field_name

Input JSON data of type field_name.

val field_name_of_string : string -> field_name

Deserialize JSON data of type field_name.

val write_access_level : Stdlib.Buffer.t -> access_level -> unit

Output a JSON value of type access_level.

val string_of_access_level : ?len:int -> access_level -> string

Serialize a value of type access_level into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_access_level : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> access_level

Input JSON data of type access_level.

val access_level_of_string : string -> access_level

Deserialize JSON data of type access_level.

val write_annotation_point_method : Stdlib.Buffer.t -> annotation_point_method -> unit

Output a JSON value of type annotation_point_method.

val string_of_annotation_point_method : ?len:int -> annotation_point_method -> string

Serialize a value of type annotation_point_method into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_annotation_point_method : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> annotation_point_method

Input JSON data of type annotation_point_method.

val annotation_point_method_of_string : string -> annotation_point_method

Deserialize JSON data of type annotation_point_method.

val write_annotation_point_kind : Stdlib.Buffer.t -> annotation_point_kind -> unit

Output a JSON value of type annotation_point_kind.

val string_of_annotation_point_kind : ?len:int -> annotation_point_kind -> string

Serialize a value of type annotation_point_kind into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_annotation_point_kind : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> annotation_point_kind

Input JSON data of type annotation_point_kind.

val annotation_point_kind_of_string : string -> annotation_point_kind

Deserialize JSON data of type annotation_point_kind.

val write_annotation_point : Stdlib.Buffer.t -> annotation_point -> unit

Output a JSON value of type annotation_point.

val string_of_annotation_point : ?len:int -> annotation_point -> string

Serialize a value of type annotation_point into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_annotation_point : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> annotation_point

Input JSON data of type annotation_point.

val annotation_point_of_string : string -> annotation_point

Deserialize JSON data of type annotation_point.

val write_nullsafe_extra : Stdlib.Buffer.t -> nullsafe_extra -> unit

Output a JSON value of type nullsafe_extra.

val string_of_nullsafe_extra : ?len:int -> nullsafe_extra -> string

Serialize a value of type nullsafe_extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_nullsafe_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> nullsafe_extra

Input JSON data of type nullsafe_extra.

val nullsafe_extra_of_string : string -> nullsafe_extra

Deserialize JSON data of type nullsafe_extra.

val write_loc : Stdlib.Buffer.t -> loc -> unit

Output a JSON value of type loc.

val string_of_loc : ?len:int -> loc -> string

Serialize a value of type loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> loc

Input JSON data of type loc.

val loc_of_string : string -> loc

Deserialize JSON data of type loc.

val write_json_trace_item : Stdlib.Buffer.t -> json_trace_item -> unit

Output a JSON value of type json_trace_item.

val string_of_json_trace_item : ?len:int -> json_trace_item -> string

Serialize a value of type json_trace_item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_json_trace_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> json_trace_item

Input JSON data of type json_trace_item.

val json_trace_item_of_string : string -> json_trace_item

Deserialize JSON data of type json_trace_item.

val write_config_usage_extra : Stdlib.Buffer.t -> config_usage_extra -> unit

Output a JSON value of type config_usage_extra.

val string_of_config_usage_extra : ?len:int -> config_usage_extra -> string

Serialize a value of type config_usage_extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_config_usage_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> config_usage_extra

Input JSON data of type config_usage_extra.

val config_usage_extra_of_string : string -> config_usage_extra

Deserialize JSON data of type config_usage_extra.

val write_extra : Stdlib.Buffer.t -> extra -> unit

Output a JSON value of type extra.

val string_of_extra : ?len:int -> extra -> string

Serialize a value of type extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> extra

Input JSON data of type extra.

val extra_of_string : string -> extra

Deserialize JSON data of type extra.

val write_jsonbug : Stdlib.Buffer.t -> jsonbug -> unit

Output a JSON value of type jsonbug.

val string_of_jsonbug : ?len:int -> jsonbug -> string

Serialize a value of type jsonbug into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_jsonbug : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> jsonbug

Input JSON data of type jsonbug.

val jsonbug_of_string : string -> jsonbug

Deserialize JSON data of type jsonbug.

val write_report : Stdlib.Buffer.t -> report -> unit

Output a JSON value of type report.

val string_of_report : ?len:int -> report -> string

Serialize a value of type report into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report

Input JSON data of type report.

val report_of_string : string -> report

Deserialize JSON data of type report.