Module ATDGenerated.Jsonconfigimpact_j

type loc = Jsonbug_t.loc

val equal_loc : loc -> loc -> bool

type sub_item = Jsoncost_t.sub_item = {

hash : string;

loc : loc;

procedure_name : string;

procedure_id : string;

}

val equal_sub_item : sub_item -> sub_item -> bool

type config_impact_mode = Jsonconfigimpact_t.config_impact_mode

val equal_config_impact_mode : config_impact_mode -> config_impact_mode -> bool

type item = Jsonconfigimpact_t.item = {

hash : string;

loc : loc;

procedure_name : string;

procedure_id : string;

unchecked_callees : string;

mode : config_impact_mode;

}

val equal_item : item -> item -> bool

type report = Jsonconfigimpact_t.report

val equal_report : report -> report -> bool

val write_loc : Bi_outbuf.t -> loc -> unit

Output a JSON value of type loc.

val string_of_loc : ?len:int -> loc -> string

Serialize a value of type loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> loc

Input JSON data of type loc.

val loc_of_string : string -> loc

Deserialize JSON data of type loc.

val write_sub_item : Bi_outbuf.t -> sub_item -> unit

Output a JSON value of type sub_item.

val string_of_sub_item : ?len:int -> sub_item -> string

Serialize a value of type sub_item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_sub_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> sub_item

Input JSON data of type sub_item.

val sub_item_of_string : string -> sub_item

Deserialize JSON data of type sub_item.

val write_config_impact_mode : Bi_outbuf.t -> config_impact_mode -> unit

Output a JSON value of type config_impact_mode.

val string_of_config_impact_mode : ?len:int -> config_impact_mode -> string

Serialize a value of type config_impact_mode into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_config_impact_mode : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> config_impact_mode

Input JSON data of type config_impact_mode.

val config_impact_mode_of_string : string -> config_impact_mode

Deserialize JSON data of type config_impact_mode.

val write_item : Bi_outbuf.t -> item -> unit

Output a JSON value of type item.

val string_of_item : ?len:int -> item -> string

Serialize a value of type item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> item

Input JSON data of type item.

val item_of_string : string -> item

Deserialize JSON data of type item.

val write_report : Bi_outbuf.t -> report -> unit

Output a JSON value of type report.

val string_of_report : ?len:int -> report -> string

Serialize a value of type report into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report

Input JSON data of type report.

val report_of_string : string -> report

Deserialize JSON data of type report.