Module ATDGenerated.Jsoncost_j

type loc = Jsonbug_t.loc

val equal_loc : loc -> loc -> bool

type sub_item = Jsoncost_t.sub_item = {

hash : string;

loc : loc;

procedure_name : string;

procedure_id : string;

}

val equal_sub_item : sub_item -> sub_item -> bool

type json_trace_item = Jsonbug_t.json_trace_item

val equal_json_trace_item : json_trace_item -> json_trace_item -> bool

type hum_info = Jsoncost_t.hum_info = {

hum_polynomial : string;

hum_degree : string;

big_o : string;

}

val equal_hum_info : hum_info -> hum_info -> bool

type info = Jsoncost_t.info = {

polynomial_version : int;

polynomial : string;

degree : int option;

hum : hum_info;

trace : json_trace_item list;

}

val equal_info : info -> info -> bool

type item = Jsoncost_t.item = {

hash : string;

loc : loc;

procedure_name : string;

procedure_id : string;

is_on_ui_thread : bool;

exec_cost : info;

autoreleasepool_size : info;

}

val equal_item : item -> item -> bool

type report = Jsoncost_t.report

val equal_report : report -> report -> bool

val write_loc : Bi_outbuf.t -> loc -> unit

Output a JSON value of type loc.

val string_of_loc : ?len:int -> loc -> string

Serialize a value of type loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> loc

Input JSON data of type loc.

val loc_of_string : string -> loc

Deserialize JSON data of type loc.

val write_sub_item : Bi_outbuf.t -> sub_item -> unit

Output a JSON value of type sub_item.

val string_of_sub_item : ?len:int -> sub_item -> string

Serialize a value of type sub_item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_sub_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> sub_item

Input JSON data of type sub_item.

val sub_item_of_string : string -> sub_item

Deserialize JSON data of type sub_item.

val write_json_trace_item : Bi_outbuf.t -> json_trace_item -> unit

Output a JSON value of type json_trace_item.

val string_of_json_trace_item : ?len:int -> json_trace_item -> string

Serialize a value of type json_trace_item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_json_trace_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> json_trace_item

Input JSON data of type json_trace_item.

val json_trace_item_of_string : string -> json_trace_item

Deserialize JSON data of type json_trace_item.

val write_hum_info : Bi_outbuf.t -> hum_info -> unit

Output a JSON value of type hum_info.

val string_of_hum_info : ?len:int -> hum_info -> string

Serialize a value of type hum_info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_hum_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> hum_info

Input JSON data of type hum_info.

val hum_info_of_string : string -> hum_info

Deserialize JSON data of type hum_info.

val write_info : Bi_outbuf.t -> info -> unit

Output a JSON value of type info.

val string_of_info : ?len:int -> info -> string

Serialize a value of type info into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> info

Input JSON data of type info.

val info_of_string : string -> info

Deserialize JSON data of type info.

val write_item : Bi_outbuf.t -> item -> unit

Output a JSON value of type item.

val string_of_item : ?len:int -> item -> string

Serialize a value of type item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> item

Input JSON data of type item.

val item_of_string : string -> item

Deserialize JSON data of type item.

val write_report : Bi_outbuf.t -> report -> unit

Output a JSON value of type report.

val string_of_report : ?len:int -> report -> string

Serialize a value of type report into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report

Input JSON data of type report.

val report_of_string : string -> report

Deserialize JSON data of type report.