Module ATDGenerated.Report_block_list_spec_j

type report_block_list_spec = Report_block_list_spec_t.report_block_list_spec = {

bug_type : string;

class_name : string option;

procedure_name : string;

file : string;

comment : string option;

}

type report_block_list_specs = Report_block_list_spec_t.report_block_list_specs

val write_report_block_list_spec : Stdlib.Buffer.t -> report_block_list_spec -> unit

Output a JSON value of type report_block_list_spec.

val string_of_report_block_list_spec : ?len:int -> report_block_list_spec -> string

Serialize a value of type report_block_list_spec into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report_block_list_spec : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report_block_list_spec

Input JSON data of type report_block_list_spec.

val report_block_list_spec_of_string : string -> report_block_list_spec

Deserialize JSON data of type report_block_list_spec.

val write_report_block_list_specs : Stdlib.Buffer.t -> report_block_list_specs -> unit

Output a JSON value of type report_block_list_specs.

val string_of_report_block_list_specs : ?len:int -> report_block_list_specs -> string

Serialize a value of type report_block_list_specs into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report_block_list_specs : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report_block_list_specs

Input JSON data of type report_block_list_specs.

val report_block_list_specs_of_string : string -> report_block_list_specs

Deserialize JSON data of type report_block_list_specs.