Module ATDGenerated.Semdiff_batch_j

type json_ast = Yojson.Safe.t

type semdiff_file = Semdiff_batch_t.semdiff_file = {

filename : string;

source : string;

ast : json_ast;

}

type semdiff_pair = Semdiff_batch_t.semdiff_pair = {

previous : semdiff_file;

current : semdiff_file;

}

type semdiff_input = Semdiff_batch_t.semdiff_input = {

language : string;

pairs : semdiff_pair list;

}

val write_json_ast : Stdlib.Buffer.t -> json_ast -> unit

Output a JSON value of type json_ast.

val string_of_json_ast : ?len:int -> json_ast -> string

Serialize a value of type json_ast into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_json_ast : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> json_ast

Input JSON data of type json_ast.

val json_ast_of_string : string -> json_ast

Deserialize JSON data of type json_ast.

val write_semdiff_file : Stdlib.Buffer.t -> semdiff_file -> unit

Output a JSON value of type semdiff_file.

val string_of_semdiff_file : ?len:int -> semdiff_file -> string

Serialize a value of type semdiff_file into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_semdiff_file : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> semdiff_file

Input JSON data of type semdiff_file.

val semdiff_file_of_string : string -> semdiff_file

Deserialize JSON data of type semdiff_file.

val write_semdiff_pair : Stdlib.Buffer.t -> semdiff_pair -> unit

Output a JSON value of type semdiff_pair.

val string_of_semdiff_pair : ?len:int -> semdiff_pair -> string

Serialize a value of type semdiff_pair into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_semdiff_pair : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> semdiff_pair

Input JSON data of type semdiff_pair.

val semdiff_pair_of_string : string -> semdiff_pair

Deserialize JSON data of type semdiff_pair.

val write_semdiff_input : Stdlib.Buffer.t -> semdiff_input -> unit

Output a JSON value of type semdiff_input.

val string_of_semdiff_input : ?len:int -> semdiff_input -> string

Serialize a value of type semdiff_input into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_semdiff_input : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> semdiff_input

Input JSON data of type semdiff_input.

val semdiff_input_of_string : string -> semdiff_input

Deserialize JSON data of type semdiff_input.