Module ATDGenerated.Specialized_call_graph_j

type node = Specialized_call_graph_t.node = {

name : string;

context : int;

}

type location = Specialized_call_graph_t.location = {

file : string;

line : int;

col : int;

}

type callee = Specialized_call_graph_t.callee = {

call_location : location;

callee : int;

}

type edge = Specialized_call_graph_t.edge = {

caller : node;

callees : callee list;

}

type call_graph = Specialized_call_graph_t.call_graph = {

edges : edge list;

contexts : string list;

}

val write_node : Stdlib.Buffer.t -> node -> unit

Output a JSON value of type node.

val string_of_node : ?len:int -> node -> string

Serialize a value of type node into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_node : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> node

Input JSON data of type node.

val node_of_string : string -> node

Deserialize JSON data of type node.

val write_location : Stdlib.Buffer.t -> location -> unit

Output a JSON value of type location.

val string_of_location : ?len:int -> location -> string

Serialize a value of type location into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_location : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> location

Input JSON data of type location.

val location_of_string : string -> location

Deserialize JSON data of type location.

val write_callee : Stdlib.Buffer.t -> callee -> unit

Output a JSON value of type callee.

val string_of_callee : ?len:int -> callee -> string

Serialize a value of type callee into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_callee : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> callee

Input JSON data of type callee.

val callee_of_string : string -> callee

Deserialize JSON data of type callee.

val write_edge : Stdlib.Buffer.t -> edge -> unit

Output a JSON value of type edge.

val string_of_edge : ?len:int -> edge -> string

Serialize a value of type edge into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_edge : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> edge

Input JSON data of type edge.

val edge_of_string : string -> edge

Deserialize JSON data of type edge.

val write_call_graph : Stdlib.Buffer.t -> call_graph -> unit

Output a JSON value of type call_graph.

val string_of_call_graph : ?len:int -> call_graph -> string

Serialize a value of type call_graph into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_call_graph : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> call_graph

Input JSON data of type call_graph.

val call_graph_of_string : string -> call_graph

Deserialize JSON data of type call_graph.