Module AbstractDomain.CountDomain

Domain keeping a non-negative count with a bounded maximum value. The count can be only incremented and decremented.

Parameters

MaxCount : MaxCount

Signature

include WithBottom with type t = private int

include S

include NoJoin

include IStdlib.PrettyPrintable.PrintableType

type t

val pp : IStdlib.PrettyPrintable.F.formatter -> t -> unit

val leq : lhs:t -> rhs:t -> bool

the implication relation: lhs <= rhs means lhs |- rhs

val join : t -> t -> t

val widen : prev:t -> next:t -> num_iters:int -> t

val bottom : t

The bottom value of the domain.

val is_bottom : t -> bool

Return true if this is the bottom value

top is maximum value

include WithTop with type t := t

include S

include NoJoin

include IStdlib.PrettyPrintable.PrintableType

type t

val pp : IStdlib.PrettyPrintable.F.formatter -> t -> unit

val leq : lhs:t -> rhs:t -> bool

the implication relation: lhs <= rhs means lhs |- rhs

val join : t -> t -> t

val widen : prev:t -> next:t -> num_iters:int -> t

val top : t

val is_top : t -> bool

val increment : t -> t

bump the count by one if it is less than the max

val decrement : t -> t

descrease the count by one if it is greater than 0

val add : t -> t -> t

capped sum of two states