Module Absint.AnalysisState

State of symbolic execution

val get_instr : unit -> IR.Sil.instr option

Get last instruction seen in symbolic execution

val get_loc_exn : unit -> IBase.Location.t

Get last location seen in symbolic execution

val get_loc : unit -> IBase.Location.t option

Get last location seen in symbolic execution

val get_node_exn : unit -> IR.Procdesc.Node.t

Get last node seen in symbolic execution

val get_node : unit -> IR.Procdesc.Node.t option

Get last node seen in symbolic execution

val get_remaining_disjuncts : unit -> int option

Get number of remaining disjuncts in the transfer function for Pulse

val get_session : unit -> int

Get last session seen in symbolic execution

val set_instr : IR.Sil.instr -> unit

Set last instruction seen in symbolic execution

val set_node : IR.Procdesc.Node.t -> unit

Set last node seen in symbolic execution

val set_remaining_disjuncts : int -> unit

Set number of remaining disjuncts in the transfer function for Pulse

val set_session : int -> unit

Set last session seen in symbolic execution