Module Absint.Annotations

Annotations.

val auto_cleanup : string

val expensive : string

val inject_prop : string

val immutable : string

val lockless : string

val no_allocation : string

val nullable : string

val nonnull : string

val performance_critical : string

val prop : string

val for_non_ui_thread : string

val for_ui_thread : string

val guarded_by : string

val suppress_lint : string

val thread_safe : string

val visibleForTesting : string

val annot_ends_with : IR.Annot.t -> string -> bool

annot_ends_with annot ann_name returns true if the class name of annot, without the package, is equal to ann_name

val ia_ends_with : IR.Annot.Item.t -> string -> bool

Check if there is an annotation in ia which ends with the given name

val ia_has_annotation_with : IR.Annot.Item.t -> (IR.Annot.t -> bool) -> bool

val ia_is_initializer : IR.Annot.Item.t -> bool

val ia_is_nonnull : IR.Annot.Item.t -> bool

val ia_is_jetbrains_notnull : IR.Annot.Item.t -> bool

val ia_is_nullable : IR.Annot.Item.t -> bool

val ia_is_nullsafe_strict : IR.Annot.Item.t -> bool

val ia_find_nullsafe : IR.Annot.Item.t -> IR.Annot.t option

val ia_is_expensive : IR.Annot.Item.t -> bool

val ia_is_functional : IR.Annot.Item.t -> bool

val ia_is_ignore_allocations : IR.Annot.Item.t -> bool

val ia_is_inject : IR.Annot.Item.t -> bool

val ia_is_suppress_lint : IR.Annot.Item.t -> bool

val ia_is_not_thread_safe : IR.Annot.Item.t -> bool

val ia_is_nonblocking : IR.Annot.Item.t -> bool

val ia_is_returns_ownership : IR.Annot.Item.t -> bool

val ia_is_synchronized_collection : IR.Annot.Item.t -> bool

val ia_is_thread_safe : IR.Annot.Item.t -> bool

val ia_is_thread_confined : IR.Annot.Item.t -> bool

val ia_is_thrift_service : IR.Annot.Item.t -> bool

val ia_is_volatile : IR.Annot.Item.t -> bool

val ia_is_worker_thread : IR.Annot.Item.t -> bool

val ia_is_uithread_equivalent : IR.Annot.Item.t -> bool

val pdesc_has_return_annot : IR.Procdesc.t -> (IR.Annot.Item.t -> bool) -> bool

return true if the given predicate evaluates to true on the annotation of pdesc's return value

val pname_has_return_annot : IR.Procname.t -> (IR.Annot.Item.t -> bool) -> bool

return true if the given predicate evaluates to true on the annotation of pname's return value

val attrs_return_annot_ends_with : IR.ProcAttributes.t -> string -> bool

return true if return value in attrs is annotated with a value ending with the given string

val method_has_annotation_with : IR.Annot.Item.t -> IR.Annot.Item.t list -> (IR.Annot.t -> bool) -> bool

val field_has_annot : IR.Fieldname.t -> IR.Struct.t -> (IR.Annot.Item.t -> bool) -> bool

val struct_typ_has_annot : IR.Struct.t -> (IR.Annot.Item.t -> bool) -> bool

return true if the given predicate evaluates to true on some annotation of struct_typ