Module Absint.DataFlow

type throws =

| DontKnow	(*
May or may not throw an exception.
*)

| Throws	(*
Definitely throws an exception.
*)

| DoesNotThrow	(*
Does not throw an exception.
*)

module type DFStateType = sig ... end

Module type used to define the state component for a dataflow algorithm.

module type DF = sig ... end

Type for the dataflow API.

module MakeDF (St : DFStateType) : DF with type state = St.t

Functor to create an instance of a dataflow analysis.