Module Absint.LowerHil

module type HilConfig = sig ... end

module DefaultConfig : HilConfig

module Make (TransferFunctions : TransferFunctions.HIL) (HilConfig : HilConfig) : sig ... end

Functor for turning HIL transfer functions into SIL transfer functions

module type S = sig ... end

module MakeAbstractInterpreterWithConfig (MakeAbstractInterpreter : AbstractInterpreter.Make) (HilConfig : HilConfig) (TransferFunctions : TransferFunctions.HIL) : S with type domain = TransferFunctions.Domain.t and module Interpreter = MakeAbstractInterpreter(Make(TransferFunctions)(HilConfig))

Wrapper around Interpreter to prevent clients from having to deal with IdAccessPathMapDomain.

module MakeAbstractInterpreter (TransferFunctions : TransferFunctions.HIL) : sig ... end

Simpler version of the above wrapper that uses the default HIL config