Module LowerHil.MakeAbstractInterpreter

Simpler version of the above wrapper that uses the default HIL config

Parameters

module TransferFunctions : TransferFunctions.HIL

Signature

include sig ... end

module Interpreter : sig ... end

type domain = TransferFunctions.Domain.t

val compute_post : Interpreter.TransferFunctions.analysis_data -> initial:domain -> IR.Procdesc.t -> domain option

compute and return the postcondition for the given procedure starting from initial.