Module S.Sources

module Known : sig ... end

Set of sources returned by callees of the current function

module FootprintConfig : AccessTree.Config

module Footprint : sig ... end

Set of access paths representing the sources that may flow in from the caller

module Sanitizers : sig ... end

Set of sanitizers that have been applied to these sources

type t = {

known : Known.t;

footprint : Footprint.t;

sanitizers : Sanitizers.t;

}

val empty : t

val is_empty : t -> bool

val of_source : Source.t -> t

val of_footprint : AccessPath.Abs.t -> t

val add : Source.t -> t -> t

val get_footprint_indexes : t -> IStdlib.IntSet.t