Module Absint.TaskSchedulerTypes

type target =

| Procname of {

proc_name : IR.Procname.t;

specialization : IR.Specialization.t option;

}

| File of IBase.SourceFile.t

val equal_target : target -> target -> bool

val hash_fold_target : Ppx_hash_lib.Std.Hash.state -> target -> Ppx_hash_lib.Std.Hash.state

val hash_target : target -> Ppx_hash_lib.Std.Hash.hash_value

type analysis_result =

| Ok
(*
Analysis finished normally.
*)

| RaceOn of {

dependency_filenames : string list;

}
(*
Analysis stopped when trying to access the summary of a callee and that callee is being analyzed by another worker. dependency_filenames are in the path to the callee's lock file.
*)