Module Absint.TransferFunctions

Transfer functions that push abstract states across instructions. A typical client should implement the Make signature to allow the transfer functions to be used with any kind of CFG.

module type S = sig ... end

module type SIL = sig ... end

module type HIL = sig ... end

module type DisjunctiveConfig = sig ... end

module type DisjReady = sig ... end