Module Absint__Decompile

val find_boolean_assignment : IR.Procdesc.Node.t -> IR.Pvar.t -> bool -> IR.Procdesc.Node.t option

Find a boolean assignment to a temporary variable holding a boolean condition. The boolean parameter indicates whether the true or false branch is required.

val find_normal_variable_funcall : IR.Procdesc.Node.t -> IR.Ident.t -> (IR.Exp.t * IR.Exp.t list * IBase.Location.t * IR.CallFlags.t) option

Find the function call instruction used to initialize normal variable id, and return the function name and arguments

val find_program_variable_assignment : IR.Procdesc.Node.t -> IR.Pvar.t -> (IR.Procdesc.Node.t * IR.Ident.t) option

Find a program variable assignment in the current node or straightline predecessor.