Module Absint__LowerHil

module type HilConfig = sig ... end

module DefaultConfig : HilConfig

module Make : functor (TransferFunctions : Absint.TransferFunctions.HIL) -> functor (HilConfig : HilConfig) -> sig ... end

Functor for turning HIL transfer functions into SIL transfer functions

module type S = sig ... end

module MakeAbstractInterpreterWithConfig : functor (MakeAbstractInterpreter : Absint.AbstractInterpreter.Make) -> functor (HilConfig : HilConfig) -> functor (TransferFunctions : Absint.TransferFunctions.HIL) -> S with type domain = TransferFunctions.Domain.t and module Interpreter = MakeAbstractInterpreter(Make(TransferFunctions)(HilConfig))

Wrapper around Interpreter to prevent clients from having to deal with IdAccessPathMapDomain.

module MakeAbstractInterpreter : functor (TransferFunctions : Absint.TransferFunctions.HIL) -> sig ... end

Simpler version of the above wrapper that uses the default HIL config