Module Absint__Reporting

type log_t = ?⁠ltr:Absint.Errlog.loc_trace -> ?⁠extras:ATDGenerated.Jsonbug_t.extra -> IBase.Checker.t -> IBase.IssueType.t -> string -> unit

val log_issue_from_summary : ?⁠severity_override:IBase.IssueType.severity -> IR.Procdesc.t -> Absint.Errlog.t -> node:Absint.Errlog.node -> session:int -> loc:IBase.Location.t -> ltr:Absint.Errlog.loc_trace -> ?⁠extras:ATDGenerated.Jsonbug_t.extra -> IBase.Checker.t -> Absint.IssueToReport.t -> unit

val log_frontend_issue : Absint.Errlog.t -> loc:IBase.Location.t -> node_key:IR.Procdesc.NodeKey.t -> ltr:Absint.Errlog.loc_trace -> Absint.IssueToReport.t -> unit

Report a frontend issue of a given kind in the given error log.

val log_issue : IR.Procdesc.t -> Absint.Errlog.t -> loc:IBase.Location.t -> log_t

Add an issue to the given error log.

val log_issue_external : IR.Procname.t -> issue_log:Absint.IssueLog.t -> ?⁠severity_override:IBase.IssueType.severity -> loc:IBase.Location.t -> ltr:Absint.Errlog.loc_trace -> ?⁠access:string -> ?⁠extras:ATDGenerated.Jsonbug_t.extra -> IBase.Checker.t -> IBase.IssueType.t -> string -> Absint.IssueLog.t

Log an issue to the error log in IssueLog associated with the given procname.

val is_suppressed : ?⁠field_name:IR.Fieldname.t option -> IR.Tenv.t -> IR.ProcAttributes.t -> IBase.IssueType.t -> bool

should an issue report be suppressed due to a @SuppressLint("issue") annotation?