Module BO.Boolean

type t =

| Bottom

| False

| True

| Top

val of_bool : bool -> t

val equal : t -> t -> bool

val is_false : t -> bool

val is_true : t -> bool

val not_ : t -> t

val and_ : t -> t -> t

val or_ : t -> t -> t

module EqualOrder : sig ... end

This module provides abstract transfer functions for comparisons on unordered values, only based on equality, like abstract locations.