Module Boolean.EqualOrder

This module provides abstract transfer functions for comparisons on unordered values, only based on equality, like abstract locations.

type b = t

type t = {

on_equal : b;

on_not_equal : b;

}

val eq : t

val ne : t

val strict_cmp : t

> or <

val loose_cmp : t

>= or <=

val top : t

val of_equal : t -> b -> b