Module BufferOverrunDomain.CoreVal

CoreVal is similar to Val, but its compare function is defined only on core parts, interval, pointers, and array blocks, of abstract value. This domain is to keep pruned values, where we do not need to care about the other fields in the abstract values.

type t = Val.t