Module Backend

module CCallSpecializedWithClosures : sig ... end

This preanalysis aims to create specialized clones of methods that have blocks as arguments and that are called with concrete closures, and then it calls these clone methods instead of the original ones. One complication is with the captured variables in the closure: we add them to the formals of the cloned method and pass them through to the concrete blocks. We do this transformation in two steps:

module CallGraph : sig ... end

module CallbackOfChecker : sig ... end

Conversions from checkers taking "functional" Interprocedural.t et al. payloads to Callbacks.proc_callback_t and friends.

module Callbacks : sig ... end

Module to register and invoke checkers' callbacks.

module ClosureSubstSpecializedMethod : sig ... end

Documentation in CCallSpecializedWithClosures.mli

module ClosuresSubstitution : sig ... end

module Devirtualizer : sig ... end

module FileScheduler : sig ... end

module GCStats : sig ... end

module InferAnalyze : sig ... end

Main module for the analysis after the capture phase

module InferAnalyzeJson : sig ... end

Main module for the analyzejson analysis after the capture phase

module Issue : sig ... end

module MergeCapture : sig ... end

module NodePrinter : sig ... end

Simplified html node printer for checkers

module Ondemand : sig ... end

Module for on-demand analysis.

module Payloads : sig ... end

module Preanal : sig ... end

module Printer : sig ... end

Printers for the analysis results

module ProcLocker : sig ... end

module Procedures : sig ... end

module RegisterCheckers : sig ... end

module RestartScheduler : sig ... end

module ReverseAnalysisCallGraph : sig ... end

module StarvationGlobalAnalysis : sig ... end

module Summary : sig ... end

Procedure summaries: the results of the capture and all the analysis for a single procedure, plus some statistics

module SyntacticCallGraph : sig ... end

module Tasks : sig ... end