Module Backend.AnalysisDependencyGraph

val invalidate : changed_files:IBase.SourceFile.Set.t option -> unit

Build a dependency graph from the summaries in the specs table, flag any transitive dependencies of the changed_files, then invalidate any flagged transitive dependencies.

val build_for_analysis_replay : unit -> CallGraph.t

the callgraph obtained by adding edges from the summary_loads dependencies in the Summary.t.dependencies field of every procedure; used to replay the previous analysis inter-procedural schedule