Module Backend.NodePrinter

Simplified html node printer for checkers

val with_session : ?kind:[< `ComputePre | `ExecNode | `ExecNodeNarrowing | `WTO ] -> pp_name:(Stdlib.Format.formatter -> unit) -> IR.Procdesc.Node.t -> f:(unit -> 'a) -> 'a

Wraps f in an html debug session

val print_html : bool IStdlib.IStd.ref

Controls whether the HTML printing is enabled for all analyses except pre-analysis. Set to false in preanalysis to prevent printing and turned back again.