Module Backend.SpecsFiles

val iter : f:(Summary.t -> unit) -> unit

Iterates over all summaries from the .specs files

val iter_from_config : f:(Summary.t -> unit) -> unit

Iterates over all sumaries from the .specs files unless a list of specs files has been passed on the command line

val delete : IR.Procname.t -> unit

Delete the .specs file associated with a summary and remove the summary from the caches in Summary.ml and ondemand.ml

val pp_from_config : Stdlib.Format.formatter -> unit