Module Summary.OnDisk

val clear_cache : unit -> unit

Remove all the elements from the cache of summaries

val get : lazy_payloads:bool -> IR.Procname.t -> t option

Return the summary option for the procedure name

val reset : IR.Procdesc.t -> t

Reset a summary rebuilding the dependents and preserving the proc attributes if present.

val store : t -> unit

Save summary for the procedure into the spec database

val delete : IR.Procname.t -> unit

Delete the .specs file corresponding to the procname and remove its summary from the Summary cache

val delete_all : procedures:IR.Procname.t list -> unit

Similar to delete, but delete all summaries for a list of procedures

val iter_specs : f:( t -> unit ) -> unit

Iterates over all stored summaries

val iter_report_summaries_from_config : f: ( IR.Procname.t -> IBase.Location.t -> Costlib.CostDomain.summary option -> Costlib.ConfigImpactAnalysis.Summary.t option -> Absint.Errlog.t -> unit ) -> unit

Iterates over all analysis artefacts listed above, for each procedure

val get_model_proc_desc : IR.Procname.t -> IR.Procdesc.t option

val get_count : unit -> int

Counts the summaries currently stored on disk.