Module Backend__Ondemand

val get_proc_desc : IR.Procname.t -> IR.Procdesc.t option

Find a proc desc for the procedure, perhaps loading it from disk.

val analyze_proc_name : Absint.Exe_env.t -> caller_summary:Backend.Summary.t -> IR.Procname.t -> Backend.Summary.t option

analyze_proc_name exe_env ~caller_summary callee_pname performs an on-demand analysis of callee_pname triggered during the analysis of caller_summary

val analyze_proc_name_no_caller : Absint.Exe_env.t -> IR.Procname.t -> Backend.Summary.t option

analyze_proc_name_no_caller exe_env callee_pname performs an on-demand analysis of callee_pname triggered by the top-level of a file-level checker. This must not be used in any other context, as this will break incremental analysis.

module LocalCache : sig ... end

val analyze_file : Absint.Exe_env.t -> IBase.SourceFile.t -> unit

Invoke all the callbacks registered in Callbacks on the given file.

val analyze_proc_name_toplevel : Absint.Exe_env.t -> IR.Procname.t -> unit

Invoke all the callbacks registered in Callbacks on the given procedure.