Module Biabduction.Exceptions

Biabduction uses exceptions to store issues in summaries

exception Abduction_case_not_implemented of IBase.Logging.ocaml_pos

exception Analysis_stops of Absint.Localise.error_desc * IBase.Logging.ocaml_pos option

exception Array_of_pointsto of IBase.Logging.ocaml_pos

exception Array_out_of_bounds_l1 of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Array_out_of_bounds_l2 of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Array_out_of_bounds_l3 of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Bad_footprint of IBase.Logging.ocaml_pos

exception Cannot_star of IBase.Logging.ocaml_pos

exception Class_cast_exception of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Custom_error of string * IBase.IssueType.severity * Absint.Localise.error_desc

exception Dangling_pointer_dereference of bool * Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Divide_by_zero of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Empty_vector_access of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Inherently_dangerous_function of Absint.Localise.error_desc

exception Internal_error of Absint.Localise.error_desc

exception Leak of bool * bool * Absint.Localise.error_desc * bool * IR.PredSymb.resource * IBase.Logging.ocaml_pos

exception Missing_fld of IR.Fieldname.t * IBase.Logging.ocaml_pos

exception Premature_nil_termination of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Null_dereference of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Precondition_not_found of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Precondition_not_met of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Retain_cycle of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Registered_observer_being_deallocated of Absint.Localise.error_desc * IBase.Logging.ocaml_pos

exception Skip_function of Absint.Localise.error_desc

exception Symexec_memory_error of IBase.Logging.ocaml_pos

exception Wrong_argument_number of IBase.Logging.ocaml_pos

val handle_exception : exn -> bool

Return true if the exception is not serious and should be handled in timeout mode

val print_exception_html : string -> exn -> unit

print a description of the exception to the html output

val recognize_exception : exn -> Absint.IssueToReport.t