Module Biabduction__PropUtil

val remove_locals_ret : IR.Tenv.t -> IR.Procdesc.t -> Biabduction.Prop.normal Biabduction.Prop.t -> Biabduction.Prop.normal Biabduction.Prop.t

remove locals and return variable from the prop

val remove_locals_formals : IR.Tenv.t -> IR.Procdesc.t -> Biabduction.Prop.normal Biabduction.Prop.t -> IR.Pvar.t list * Biabduction.Prop.normal Biabduction.Prop.t

Deallocate the stack variables in pvars, and replace them by normal variables. Return the list of stack variables whose address was still present after deallocation.

val remove_seed_vars : IR.Tenv.t -> 'a Biabduction.Prop.t -> Biabduction.Prop.normal Biabduction.Prop.t

remove seed vars from a prop