Module Biabduction__Tabulation

val remove_constant_string_class : IR.Tenv.t -> 'a Biabduction.Prop.t -> Biabduction.Prop.normal Biabduction.Prop.t

Remove constant string or class from a prop

val check_attr_dealloc_mismatch : IR.PredSymb.t -> IR.PredSymb.t -> unit

Check if the attribute change is a mismatch between a kind of allocation and a different kind of deallocation

val find_dereference_without_null_check_in_sexp : Biabduction.Predicates.strexp -> (int * IR.PredSymb.path_pos) option

Check whether a sexp contains a dereference without null check, and return the line number and path position

val create_cast_exception : IR.Tenv.t -> IBase.Logging.ocaml_pos -> IR.Procname.t option -> IR.Exp.t -> IR.Exp.t -> IR.Exp.t -> exn

raise a cast exception

val prop_is_exn : IR.Procname.t -> 'a Biabduction.Prop.t -> bool

check if a prop is an exception

val prop_get_exn_name : IR.Procname.t -> 'a Biabduction.Prop.t -> IR.Typ.Name.t option

when prop is an exception, return the exception name

val lookup_custom_errors : 'a Biabduction.Prop.t -> string option

search in prop contains an error state

val exe_function_call : Biabduction.BiabductionSummary.t Absint.InterproceduralAnalysis.t -> callee_attributes:IR.ProcAttributes.t -> callee_pname:IR.Procname.t -> callee_summary:Biabduction.BiabductionSummary.t -> ret_id:IR.Ident.t -> IBase.Location.t -> actuals:(IR.Exp.t * IR.Typ.t) list -> Biabduction.Prop.normal Biabduction.Prop.t -> Biabduction.Paths.Path.t -> (Biabduction.Prop.normal Biabduction.Prop.t * Biabduction.Paths.Path.t) list

Execute the function call and return the list of results with return value