Module Edge.Kind

type t =

| Direct
(*
Immediate copy; e.g., assigment or passing an argument
*)

| Call
(*
Target is ArgumentOf
*)

| Return
(*
Source is ReturnOf
*)

| Capture
(*
X=1, F=fun()->X end has Capture edge from X to F
*)

| Builtin
(*
Edge coming from a suppressed builtin call, ultimately exported as a Copy
*)

| Summary of {

callee : IR.Procname.t;

shape_is_preserved : bool;

}
(*
Summarizes the effect of a procedure call
*)

| DynamicCallFunction

| DynamicCallModule