Module Checkers.LineageShape

module FieldLabel : sig ... end

module FieldPath : sig ... end

module Cell : sig ... end

Cells correspond to variable fields for which no subfield will be considered by the analysis, either because they semantically have none, or because the abstract domain decides that considering them would lead to too deep or too wide field structures.

module Summary : sig ... end

val checker : Summary.t Absint.InterproceduralAnalysis.t -> Summary.t option