Module LineageShape.Cell

Cells correspond to variable fields for which no subfield will be considered by the analysis, either because they semantically have none, or because the abstract domain decides that considering them would lead to too deep or too wide field structures.

A field path of a cell:

Cannot have a length greater than Config.lineage_field_depth. For instance, if that limit is 2, the only cell under both X#foo#bar#baz and X#foo#bar#ham will be X#foo#bar.

Cannot cross a field whose field table is wider than Config.lineage_field_width. For instance, of the variable X has a huge number of fields, the only cell under X#field1 will be X itself.

Cannot go twice on same-typed fields if Config.lineage_prevent_cycles is set. For instance, if X has fields X#head and X#tail, and X#tail has the same shape as X, then the cells under X will be X#head and X#tail.

If one of these conditions happens, then the cell field path will be truncated before reaching a "scalar" field and the cell will be "abstract".

To make sure that the aforementioned properties hold, one cannot construct cells directly, but should access them through the Summary.fold_cells and Summary.fold_cell_pairs functions.

The lineage graph is built on cells.

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val pp : t Fmt.t

val var : t -> IR.Var.t

val field_path : t -> FieldPath.t

val is_abstract : t -> bool

val var_appears_in_source_code : t -> bool