Module LineageShape.FieldPath

A module to help manipulating lists of (nested) fields.

type t = FieldLabel.t list

The fields are listed in syntactic order: [a; b] for x#a#b.

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Equal.S with type t := t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

include IStdlib.IStd.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__Comparator.comparator

val validate_lbound : min:t Core__.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core__.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core__.Maybe_bound.t -> max:t Core__.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end

val pp : t Fmt.t

val first_field_is : FieldLabel.t -> t -> bool

Returns true iff the field path is not empty and its head is equal to the field label.