Module SimpleShape.Summary

type t

val pp : Stdlib.Format.formatter -> t -> unit

val fold_terminal_fields : t -> (IR.Var.t * IR.Fieldname.t list) -> max_width:int -> max_depth:int -> prevent_cycles:bool -> init:'accum -> f:( 'accum -> IR.Fieldname.t list -> 'accum ) -> 'accum

Folds over all terminal fields of a variable with fields. A field list is "terminal" if its length (that includes the prefixed fields given as parameters) is equal to max_depth, or no more field can be subscripted from its corresponding type, or it has strictly more than max_width immediate subfields.

The parameter fields are expected in syntactic order: [a; b] for x#a#b.

The result will not cross any shape whose field table is wider than max_width, even if one of the parameter does. For instance, if some variable X has a huge number of fields, the only terminal field of X#field1 will be X itself.

val concat_map_terminal_fields : t -> (IR.Var.t * IR.Fieldname.t list) -> max_width:int -> max_depth:int -> prevent_cycles:bool -> f:( IR.Fieldname.t list -> 'a list ) -> 'a list

A specialised version of fold_terminal_fields for when you want to accumulate in a list some results obtained from each terminal field.

val fold_terminal_fields_2 : t -> (IR.Var.t * IR.Fieldname.t list) -> (IR.Var.t * IR.Fieldname.t list) -> max_width:int -> max_depth:int -> prevent_cycles:bool -> init:'accum -> f:( 'accum -> IR.Fieldname.t list -> IR.Fieldname.t list -> 'accum ) -> 'accum

Folds over all terminal fields of two same-shape variables with their fields. See fold_terminal_fields.

Dies if the parameters do not have the same shape.

The folding function f will always be called on pairs of terminal fields.

Note that, since the parameters field lists may not have the same length, f may be called on field lists of different lengths. For instance, if the parameters var#field1 and var'#field2#field3 don't have any subfield, then f will be called with #field1 and #field2#field3.

Also, since the parameters lengths are taken into account for determining the depth of the result, the arguments of f may not have the same suffix when extracting these parameters: for instance, if max_depth = 1 and var and var'#field have a subfield foo, then f will be called on #foo and #field.

The same width limitation as the function fold_terminal_fields is ensured for both parameters (even if these parameters cross wide shapes, the result will not).