Module ClangFrontend.Ast_expressions

This module creates extra ast constructs that are needed for the translation

val create_pointer_qual_type : ?quals:IR.Typ.type_quals -> ATDGenerated.Clang_ast_t.qual_type -> ATDGenerated.Clang_ast_t.qual_type

val create_reference_qual_type : ?quals:IR.Typ.type_quals -> ATDGenerated.Clang_ast_t.qual_type -> ATDGenerated.Clang_ast_t.qual_type

val create_void_type : ATDGenerated.Clang_ast_t.qual_type

val create_char_star_type : ?quals:IR.Typ.type_quals -> unit -> ATDGenerated.Clang_ast_t.qual_type

val create_class_pointer_qual_type : ?quals:IR.Typ.type_quals -> IR.Typ.Name.t -> ATDGenerated.Clang_ast_t.qual_type

val create_nil : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.stmt

val create_implicit_cast_expr : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.stmt list -> ATDGenerated.Clang_ast_t.qual_type -> ATDGenerated.Clang_ast_t.cast_kind -> ATDGenerated.Clang_ast_t.stmt

val create_decl_info : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.pointer -> ATDGenerated.Clang_ast_t.decl_info

val default_var_decl_info : ATDGenerated.Clang_ast_t.var_decl_info

val create_named_decl_info : string -> ATDGenerated.Clang_ast_t.named_decl_info

val create_decl_ref_expr : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.pointer -> ATDGenerated.Clang_ast_t.named_decl_info -> ATDGenerated.Clang_ast_t.qual_type -> ATDGenerated.Clang_ast_t.stmt

val create_obj_c_message_expr : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.qual_type -> ATDGenerated.Clang_ast_t.selector -> ATDGenerated.Clang_ast_t.stmt list -> ATDGenerated.Clang_ast_t.stmt

val make_obj_c_message_expr_info_class : string -> IR.Typ.Name.t -> ATDGenerated.Clang_ast_t.pointer option -> ATDGenerated.Clang_ast_t.obj_c_message_expr_info

val trans_with_conditional : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.expr_info -> ATDGenerated.Clang_ast_t.stmt list -> ATDGenerated.Clang_ast_t.stmt

We translate an expression with a conditional x <=> x?1:0

val trans_negation_with_conditional : ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.expr_info -> ATDGenerated.Clang_ast_t.stmt list -> ATDGenerated.Clang_ast_t.stmt

We translate the logical negation of an expression with a conditional !x <=> x?0:1