Module ClangFrontend.CEnum_decl

Translate an enumeration declaration by adding it to the tenv and

translating the code and adding it to a fake procdesc

val enum_decl : ATDGenerated.Clang_ast_t.decl -> IR.Typ.desc