Module ClangFrontend.CType_decl

module CProcname : sig ... end

Processes types and record declarations by adding them to the tenv

val get_record_typename : ?tenv:IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> IR.Typ.Name.t

val add_types_from_decl_to_tenv : IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> IR.Typ.desc

val add_predefined_types : IR.Tenv.t -> unit

Add the predefined types objc_class which is a struct, and Class, which is a pointer to objc_class.

val qual_type_to_sil_type : IR.Tenv.t -> ATDGenerated.Clang_ast_t.qual_type -> IR.Typ.t

val class_from_pointer_type : IR.Tenv.t -> ATDGenerated.Clang_ast_t.qual_type -> IR.Typ.Name.t

val get_type_from_expr_info : ATDGenerated.Clang_ast_t.expr_info -> IR.Tenv.t -> IR.Typ.t

val get_template_args : IR.Tenv.t -> ATDGenerated.Clang_ast_t.template_instantiation_arg_info list -> IR.Typ.template_arg list

val method_signature_of_decl : IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> ?block_return_type:ATDGenerated.Clang_ast_t.qual_type -> ?block_as_arg_attributes:IR.ProcAttributes.block_as_arg_attributes option -> IR.Procname.t -> CMethodSignature.t

val method_signature_body_of_decl : IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> ?block_return_type:ATDGenerated.Clang_ast_t.qual_type -> ?block_as_arg_attributes:IR.ProcAttributes.block_as_arg_attributes option -> IR.Procname.t -> CMethodSignature.t * ATDGenerated.Clang_ast_t.stmt option * CFrontend_config.instr_type list

val should_add_return_param : IR.Typ.t -> bool

val type_of_captured_var : IR.Tenv.t -> is_block_inside_objc_class_method:bool -> ATDGenerated.Clang_ast_t.decl_ref -> IR.Typ.t option