Module CType_decl.CProcname

val from_decl : ?tenv:IR.Tenv.t -> ?block_return_type:ATDGenerated.Clang_ast_t.qual_type -> ?outer_proc:IR.Procname.t -> ATDGenerated.Clang_ast_t.decl -> IR.Procname.t

Given decl, return its procname. This function should be used for all procedures present in original AST

module NoAstDecl : sig ... end

WARNING: functions from this module should not be used if full decl is available in AST