Module ClangFrontend.CVar_decl

Process variable declarations by saving them as local or global variables.

val sil_var_of_decl : CContext.t -> ATDGenerated.Clang_ast_t.decl -> IR.Procname.t -> IR.Pvar.t

Computes the local variables of a function or method to be added to the procdesc

val sil_var_of_decl_ref : CContext.t -> ATDGenerated.Clang_ast_t.source_range -> ATDGenerated.Clang_ast_t.decl_ref -> IR.Procname.t -> IR.Pvar.t

val add_var_to_locals : IR.Procdesc.t -> ATDGenerated.Clang_ast_t.decl -> IR.Typ.t -> IR.Pvar.t -> unit

val sil_var_of_captured_var : CContext.t -> ATDGenerated.Clang_ast_t.source_range -> IR.Procname.t -> ATDGenerated.Clang_ast_t.decl_ref -> (IR.Pvar.t * IR.Typ.t * bool) option

val captured_vars_from_block_info : CContext.t -> ATDGenerated.Clang_ast_t.source_range -> ATDGenerated.Clang_ast_t.block_captured_variable list -> (IR.Pvar.t * IR.Typ.t * bool) list

val mk_temp_sil_var : IR.Procdesc.t -> name:string -> IR.Pvar.t

val mk_temp_sil_var_for_expr : CContext.t -> name:string -> clang_pointer:int -> ATDGenerated.Clang_ast_t.expr_info -> IR.Pvar.t * IR.Typ.t

val materialize_cpp_temporary : CContext.t -> ATDGenerated.Clang_ast_t.stmt_info -> ATDGenerated.Clang_ast_t.expr_info -> IR.Pvar.t * IR.Typ.t

val mk_sil_global_var : IR.Tenv.t -> CFrontend_config.translation_unit_context -> ?mk_name:(string -> IR.Mangled.t -> IR.Mangled.t) -> ATDGenerated.Clang_ast_t.decl_info -> ATDGenerated.Clang_ast_t.named_decl_info -> ATDGenerated.Clang_ast_t.var_decl_info -> ATDGenerated.Clang_ast_t.template_instantiation_arg_info list option -> ATDGenerated.Clang_ast_t.qual_type -> IR.Pvar.t