Module ClangFrontend.ClangPointers

module Map : sig ... end

val pointer_decl_table : ATDGenerated.Clang_ast_t.decl IStdlib.IStd.Int.Table.t

maps decl pointer to its decl record

val pointer_stmt_table : ATDGenerated.Clang_ast_t.stmt IStdlib.IStd.Int.Table.t

maps stmt pointer to its stmt record

val pointer_type_table : ATDGenerated.Clang_ast_t.c_type IStdlib.IStd.Int.Table.t

map pointer to its type

val populate_all_tables : ATDGenerated.Clang_ast_t.decl -> unit

discover what pointers should point to in the tables above; should be run once for the current toplevel decl