Module ClangFrontend.ClangWrapper

val exe : prog:string -> args:string list -> unit

Given a clang command, normalize it via clang -### if needed to get a clear view of what work is being done and which source files are being compiled, if any. Pass the resulting files to compile to Capture to be captured, i.e., parsed and translated into Infer's IR IR.Sil.