Module ClangFrontend.ObjcInterface_decl

In this module an ObjC interface declaration is processed. The class is saved in the tenv as a struct with the corresponding fields, potential superclass and list of defined methods

val interface_declaration : CAst_utils.qual_type_to_sil_type -> CAst_utils.procname_from_decl -> IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> IR.Typ.desc

val interface_impl_declaration : CAst_utils.qual_type_to_sil_type -> CAst_utils.procname_from_decl -> IR.Tenv.t -> ATDGenerated.Clang_ast_t.decl -> IR.Typ.desc