Module ClangFrontend__CArithmetic_trans

val bin_op_to_string : ATDGenerated.Clang_ast_t.binary_operator_info -> string

val binary_operation_instruction : ATDGenerated.Clang_ast_t.source_range -> ATDGenerated.Clang_ast_t.binary_operator_info -> (IR.Exp.t * IR.Typ.t) -> IR.Typ.t -> (IR.Exp.t * IR.Typ.t) -> IBase.Location.t -> IR.Exp.t * IR.Sil.instr list

Returns a pair (binary_expression, instructions). "binary_expression" is returned when we are calculating an expression "instructions" is not empty when the binary operator is actually a statement like an assignment.

val unary_operation_instruction : ClangFrontend.CFrontend_config.translation_unit_context -> ATDGenerated.Clang_ast_t.unary_operator_info -> IR.Exp.t -> IR.Typ.t -> IBase.Location.t -> IR.Exp.t * IR.Sil.instr list

val sil_const_plus_one : IR.Exp.t -> IR.Exp.t