Parameter CFrontend_decl_funct.1-T

val instructions_trans : ClangFrontend.CContext.t -> ATDGenerated.Clang_ast_t.stmt -> ClangFrontend.CFrontend_config.instr_type list -> IR.Procdesc.Node.t -> is_destructor_wrapper:bool -> IR.Procdesc.Node.t list

It receives the context, a list of statements from clang ast, list of custom statments to be added before clang statements and the exit node and it returns a list of cfg nodes that represent the translation of the stmts into sil.