Module ClangFrontend__CScope

type var_to_destroy = {

pvar : IR.Pvar.t;

typ : IR.Typ.t;

qual_type : ATDGenerated.Clang_ast_t.qual_type;

marker : IR.Pvar.t option;
Some m means that creating pvar should also set m to 1 so that we know whether pvar needs to be destroyed after the current full-expression

}

val breaks_control_flow : ATDGenerated.Clang_ast_t.stmt -> bool

module Variables : sig ... end

module CXXTemporaries : sig ... end