Module ClangFrontend__CTrans_utils.Self

This module handles the translation of the variable self which is challenging because self is used both as a variable in instance method calls and also as a type in class method calls.

val add_self_parameter_for_super_instance : ATDGenerated.Clang_ast_t.stmt_info -> ClangFrontend.CContext.t -> IR.Procname.t -> IBase.Location.t -> ATDGenerated.Clang_ast_t.obj_c_message_expr_info -> trans_result option

val is_var_self : IR.Pvar.t -> bool -> bool