Module RacerDDomain.AttributeMapDomain

include Absint.AbstractDomain.InvertedMapS with type key = AccessExpression.t and type value = Attribute.t

include IStdlib.PrettyPrintable.PPMonoMap with type key = AccessExpression.t with type value = Attribute.t

include IStdlib.PrettyPrintable.MonoMap with type key = AccessExpression.t with type value = Attribute.t

type key = AccessExpression.t

type value = Attribute.t

type t

val empty : t

val is_empty : t -> bool

val mem : key -> t -> bool

val add : key -> value -> t -> t

val update : key -> (value option -> value option) -> t -> t

val singleton : key -> value -> t

val remove : key -> t -> t

val merge : (key -> value option -> value option -> value option) -> t -> t -> t

val union : (key -> value -> value -> value option) -> t -> t -> t

val compare : (value -> value -> int) -> t -> t -> int

val equal : (value -> value -> bool) -> t -> t -> bool

val iter : (key -> value -> unit) -> t -> unit

val fold : (key -> value -> 'a -> 'a) -> t -> 'a -> 'a

val for_all : (key -> value -> bool) -> t -> bool

val exists : (key -> value -> bool) -> t -> bool

val filter : (key -> value -> bool) -> t -> t

val filter_map : (key -> value -> value option) -> t -> t

val partition : (key -> value -> bool) -> t -> t * t

val cardinal : t -> int

val bindings : t -> (key * value) list

val min_binding : t -> key * value

val min_binding_opt : t -> (key * value) option

val max_binding : t -> key * value

val max_binding_opt : t -> (key * value) option

val choose : t -> key * value

val choose_opt : t -> (key * value) option

val split : key -> t -> t * value option * t

val find : key -> t -> value

val find_opt : key -> t -> value option

val find_first : (key -> bool) -> t -> key * value

val find_first_opt : (key -> bool) -> t -> (key * value) option

val find_last : (key -> bool) -> t -> key * value

val find_last_opt : (key -> bool) -> t -> (key * value) option

val map : (value -> value) -> t -> t

val mapi : (key -> value -> value) -> t -> t

val is_singleton_or_more : t -> (key * value) IStdlib.IContainer.singleton_or_more

val fold_map : t -> init:'a -> f:('a -> value -> 'a * value) -> 'a * t

val fold_mapi : t -> init:'a -> f:(key -> 'a -> value -> 'a * value) -> 'a * t

val of_seq : (key * value) Stdlib.Seq.t -> t

val to_seq : t -> (key * value) Stdlib.Seq.t

include IStdlib.PrettyPrintable.PrintableType with type t := t

val pp_key : IStdlib.PrettyPrintable.F.formatter -> key -> unit

include Absint.AbstractDomain.WithTop with type t := t

include Absint.AbstractDomain.S with type t := t

include Absint.AbstractDomain.Comparable with type t := t

include IStdlib.PrettyPrintable.PrintableType with type t := t

val pp : IStdlib.PrettyPrintable.F.formatter -> t -> unit

val leq : lhs:t -> rhs:t -> bool

the implication relation: lhs <= rhs means lhs |- rhs

val join : t -> t -> t

val widen : prev:t -> next:t -> num_iters:int -> t

val top : t

val is_top : t -> bool

val get : AccessExpression.t -> t -> Attribute.t

find the Attribute.t associated with a given access expression or return Attribute.bottom

val is_functional : t -> AccessExpression.t -> bool

val propagate_assignment : AccessExpression.t -> Absint.HilExp.t -> t -> t

propagate attributes from the leaves to the root of an RHS Hil expression