Module RacerDDomain.OwnershipAbstractValue

type t = private

| OwnedIf of IStdlib.IntSet.t
(*
Owned if the formals at the given indexes are owned in the caller; unconditionally owned if the set of formals is empty = bottom of the lattice
*)

| Unowned
(*
Unowned value; top of the lattice
*)

val owned : t

val make_owned_if : int -> t