Module Costlib__ConstraintSolver

module F = Stdlib.Format

module BasicCost = Costlib.CostDomain.BasicCost

module Node = Absint.ProcCfg.DefaultNode

type debug = {

f : a. ('a, F.formatter, unit, unit) IStdlib.IStd.format4 -> 'a;

}

module Equalities : sig ... end

val compute_costs : debug:debug -> BasicCost.t Node.IdMap.t -> Equalities.t -> unit

repeatedly improve the costs given the constraints

val get_node_nb_exec : Equalities.t -> Node.t -> BasicCost.t

compute the number of times a node is executed by taking into account the program structural (e.g. control-flow) constraints

val collect_constraints : debug:debug -> IR.Procdesc.t -> Equalities.t

collect initial constraints for a CFG