Module Costlib__ControlFlowCost

module F = Stdlib.Format

module BasicCost = Costlib.CostDomain.BasicCost

module Node = Absint.ProcCfg.DefaultNode

module Item : sig ... end

module Sum : sig ... end

type t = [

| `Edge of Node.id * Node.id

| `Node of Node.id

| `Sum of int * Item.t list

]

A Control-flow cost represents the number of times the flow of control can go through a certain CFG item (a node or an edge), or a sum of such things

val make_node : Node.id -> t

val make_pred_edge : 'a -> 'b -> [> `Edge of 'b * 'a ]

val make_succ_edge : 'a -> 'b -> [> `Edge of 'a * 'b ]

val pp : F.formatter -> t -> unit

val sum : Item.t list -> t

module Set : sig ... end