Module Datalog.Fact

type t =

| Reachable of {

proc_name : IR.Procname.t;

}

| Extends of {

typ : IR.Typ.Name.t;

typ_super : IR.Typ.Name.t;

}

| Cast of {

proc_name : IR.Procname.t;

dest : IR.Ident.t;

src : IR.Ident.t;

dest_typ : IR.Typ.t;

}

| Alloc of {

proc_name : IR.Procname.t;

return : IR.Ident.t;

allocation_site : string;

typ : IR.Typ.t;

}

| VirtualCall of {

proc_name : IR.Procname.t;

call_site : string;

receiver : IR.Ident.t;

proc_signature : string;

}

| StaticCall of {

proc_name : IR.Procname.t;

call_site : string;

call_proc : IR.Procname.t;

}

| ActualArg of {

proc_name : IR.Procname.t;

call_site : string;

n_arg : int;

arg : IR.Ident.t;

}

| FormalArg of {

proc_name : IR.Procname.t;

n_arg : int;

arg : IR.Ident.t;

}

| ActualReturn of {

proc_name : IR.Procname.t;

call_site : string;

return : IR.Ident.t;

}

| FormalReturn of {

proc_name : IR.Procname.t;

return : IR.Ident.t;

}

Type of a Datalog fact. In all facts, "proc_name" is the procedure inside of which the fact is emitted. The allocation and call sites are uniquely identified by the string "class:line:assigned_variable" (note that in SIL there is always an assigned variable, even for procedures with a void return type). The assigned variable is used instead of the column because the Java frontend does not provide information about columns, and variables have unique names.

val to_string : t -> string

val iter_fact_types : ( string -> unit ) -> unit

val reachable : IR.Procname.t -> t

val extends : IR.Typ.Name.t -> IR.Typ.Name.t -> t

val cast : IR.Procname.t -> IR.Ident.t -> IR.Ident.t -> IR.Typ.t -> t

val alloc : IR.Procname.t -> IR.Ident.t -> IBase.Location.t -> IR.Typ.t -> t

val virtual_call : IR.Procname.t -> IBase.Location.t -> IR.Ident.t -> IR.Procname.t -> IR.Ident.t -> t

val static_call : IR.Procname.t -> IBase.Location.t -> IR.Ident.t -> IR.Procname.t -> t

val actual_arg : IR.Procname.t -> IBase.Location.t -> IR.Ident.t -> int -> IR.Ident.t -> t

val formal_arg : IR.Procname.t -> int -> IR.Ident.t -> t

val actual_return : IR.Procname.t -> IBase.Location.t -> IR.Ident.t -> t

val formal_return : IR.Procname.t -> IR.Ident.t -> t