Module IBase

module BuckMode : sig ... end

module Checker : sig ... end

module ClangProc : sig ... end

module CommandDoc : sig ... end

module CommandLineOption : sig ... end

Definition and parsing of command line arguments

module Config : sig ... end

Configuration values: either constant, determined at compile time, or set at startup time by system calls, environment variables, or command line options

module CostIssues : sig ... end

module CostIssuesTestField : sig ... end

module CostKind : sig ... end

module DB : sig ... end

Database of analysis results

module DBWriter : sig ... end

module Die : sig ... end

module Epilogues : sig ... end

module Exception : sig ... end

The restart scheduler and biabduction use exceptions for control flow (restarts/timeouts respectively). Functions here abstract away the semantics of when an exception can be ignored.

module ExecutionDuration : sig ... end

module FileDiff : sig ... end

module ForkUtils : sig ... end

module IssueType : sig ... end

module IssuesTestField : sig ... end

module Language : sig ... end

module LineReader : sig ... end

Module to read specific lines from files. The data from any file will stay in memory until the handle is collected by the gc

module Location : sig ... end

module LogEntry : sig ... end

Log entry data model, global log entry store and functions to manipulate it. Direct access to the store is not exposed.

module Logging : sig ... end

log messages at different levels of verbosity

module MarkupFormatter : sig ... end

module PerfEvent : sig ... end

module Process : sig ... end

module ProcessPool : sig ... end

module ProcessPoolState : sig ... end

module RestartSchedulerException : sig ... end

module ResultsDatabase : sig ... end

module ResultsDir : sig ... end

module ResultsDirEntryName : sig ... end

Entries in the results directory (infer-out/). Unless you want to specify a custom results directory you probably want to use ResultsDir.Entry instead of this module.

module Scuba : sig ... end

Low-level Scuba logging functionality. Provides functionality to log anything to any scuba table. (Note that Scuba is a schema-free storage, so it won't require any changes). Don't use this module directly for logging to tables with known structure. Use high-level functions that are aware of the table structure.

module ScubaLogging : sig ... end

Functionality for logging into "infer_events" Scuba table. The table is organized in form of key-value pairs. Two most important fields are "event" and "value". Other fields in the table correspond to things common for this particular run of Infer.

module Serialization : sig ... end

Serialization of data stuctures

module SourceFile : sig ... end

module SqliteUtils : sig ... end

module Stats : sig ... end

collect statistics about backend/analysis operations

module TaskBar : sig ... end

module Utils : sig ... end

module Version : sig ... end