Module IBase.AnalysisRequest

type checker_without_payload =

| LoopHoisting

type t = private

| All
(*
Request the analyses for all payloads enabled
*)

| One of PayloadId.t
(*
Request an analysis for one payload
*)

| CheckerWithoutPayload of checker_without_payload
(*
Request an analysis of a checker that has no payload in DB
*)

This is given to the callback iteration, so that only a subset of them are triggered when necessary.

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val pp : Ppx_show_runtime.Format.formatter -> t -> unit

val show : t -> string

val all : t

val one : PayloadId.t -> t

val checker_without_payload : checker_without_payload -> t