Module IBase.LineReader

Module to read specific lines from files. The data from any file will stay in memory until the handle is collected by the gc

type t

val create : unit -> t

val from_loc : t -> Location.t -> string option

get the line from a location looking for the copy of the file in the results dir

val iteri : t -> SourceFile.t -> f:(int -> string -> unit) -> unit

iterate on the lines of the file, with line numbers