Module IBase.ProcessPoolState

val in_child : bool IStdlib.IStd.ref

Keep track of whether the current execution is in a child process

val update_status : ( Mtime.t -> string -> unit ) IStdlib.IStd.ref

Ping the task bar whenever a new task is started with the start time and a description for the task

val update_heap_words : ( unit -> unit ) IStdlib.IStd.ref

val get_pid : unit -> IStdlib.IStd.Pid.t

val reset_pid : unit -> unit

val has_running_children : bool IStdlib.IStd.ref