Module IBase.Scuba

type table =

| InferEvents

A scuba table

type sample

A sample to be added to Scuba

val new_sample : time:int option -> sample

Create an empty sample with given creation timestamp. If time is not specified, corresponds to current timestamp.

val add_int : name:string -> value:int -> sample -> sample

Set a new integer field and its value to the sample. Overwrites if a field with this name was already set.

val add_normal : name:string -> value:string -> sample -> sample

Set a new string (normal in Scuba terminology) field and its value to the sample. Overwrites if a field with this name was already set.

val add_tagset : name:string -> value:string list -> sample -> sample

Set a new set of strings (tagset in Scuba terminology) field and its value to the sample. Overwrites if a field with this name was already set.

val log : table -> sample list -> unit

The main function. Log a collection of samples to the given table.