Module IBase.Timeable

module F = Stdlib.Format

type t =

| Checker of Checker.t

| Preanalysis
(*
the "pre-analysis" phase of a procedure, before we run the checkers
*)

types of everything we might possibly want to time

val to_string : t -> string

val pp : F.formatter -> t -> unit

module Map : IStdlib.IStd.Caml.Map.S with type key = t

val mk_map_of_all : init:'a -> 'a Map.t

mk_map_of_all ~init is a map timeable -> init for all timeable keys