Module IBase.WorkerPoolState

type worker_id =

| Pid of IStdlib.IStd.Pid.t

| Domain of int

val set_in_child : int option -> unit

Keep track of whether the current execution is in a child process/domain

val get_in_child : unit -> int option

Get whether the current execution is in a child process/domain

val update_status : (Mtime.t option -> string -> unit) IStdlib.IStd.ref

Ping the task bar whenever a new task is started with the start time (or just rename the current task if the start time is None) and a description for the task

val update_heap_words : (unit -> unit) IStdlib.IStd.ref

val get_pid : unit -> IStdlib.IStd.Pid.t

val reset_pid : unit -> unit