Module IBase__Checker

type t =

| AnnotationReachability

| Biabduction

| BufferOverrun

| ClassLoads

| Cost

| Eradicate

| FragmentRetainsView

| ImmutableCast

| Impurity

| InefficientKeysetIterator

| Linters

| LithoRequiredProps

| Liveness

| LoopHoisting

| NullsafeDeprecated

| PrintfArgs

| Pulse

| Purity

| Quandary

| RacerD

| ResourceLeak

| SIOF

| SelfInBlock

| Starvation

| Uninit

val equal : t -> t -> bool

val all : t list

type support =

| NoSupport
checker does not run at all for this language

| Support
checker is expected to give reasonable results

| ExperimentalSupport
checker runs but is not expected to give reasonable results

| ToySupport
the checker is for teaching purposes only (like experimental but with no plans to improve it)

per-language support for each checker

type config = {

support : IBase.Language.t -> support;

short_documentation : string;

cli_flag : string;
the flag to enable this option on the command line, without the leading "--" (like the ~long argument of CommandLineOption functions)

show_in_help : bool;

enabled_by_default : bool;

cli_deprecated_flags : string list;
more command-line flags, similar to ~deprecated arguments

}

val config : t -> config