Module IBase__LineReader

type t

val create : unit -> t

val from_loc : t -> IBase.Location.t -> string option

get the line from a location looking for the copy of the file in the results dir

val iteri : t -> IBase.SourceFile.t -> f:(int -> string -> unit) -> unit

iterate on the lines of the file, with line numbers