Module IBase__MarkupFormatter

val wrap_monospaced : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit

used to combine pp together, wrap content into a monospaced block

val pp_monospaced : Stdlib.Format.formatter -> string -> unit

pp to wrap into a monospaced block

val monospaced_to_string : string -> string

wrap into a monospaced block

val wrap_code : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit

used to combine pp together, wrap content into a code block

val pp_code : Stdlib.Format.formatter -> string -> unit

pp to wrap into a code block

val wrap_bold : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a -> unit

used to combine pp together, wrap content into a bold block

val pp_bold : Stdlib.Format.formatter -> string -> unit

pp to wrap into a bold block

val bold_to_string : string -> string

wrap into a bold block