Module IBase__RestartSchedulerException

exception ProcnameAlreadyLocked of {

dependency_filename : string;

}

for the Restart scheduler: raise when a worker tries to analyze a procedure already being analyzed by another process

val is_not_restart_exception : exn -> bool

check if the exception passed is the one defined above