Module IBase__Scribe

type category =

| InferEvents

Scribe category. Refer to Scribe documentation for details

val log : category -> string list -> unit

Immediately log a list of messages to a given Scribe category. Refer to Scribe documentation to learn about message format.