Module IR

module AnalysisGlobalState : sig ... end

module Annot : sig ... end

The Smallfoot Intermediate Language: Annotations

module Attributes : sig ... end

Module to manage the table of attributes.

module BUILTINS : sig ... end

module BiabductionModels : sig ... end

module Binop : sig ... end

The Smallfoot Intermediate Language: Binary Operators

module BuiltinDecl : sig ... end

Procnames for the builtin functions supported

module CSharpClassName : sig ... end

module CallFlags : sig ... end

The Smallfoot Intermediate Language: Call Flags

module CapturedVar : sig ... end

module Cfg : sig ... end

Control Flow Graph for Interprocedural Analysis

module ClangMethodKind : sig ... end

module Const : sig ... end

The Smallfoot Intermediate Language: Constants

module DecompiledExp : sig ... end

The Smallfoot Intermediate Language: Decompiled Expressions

module Dependencies : sig ... end

module DotCfg : sig ... end

module ErlangTypeName : sig ... end

module Exp : sig ... end

The Smallfoot Intermediate Language: Expressions

module Fieldname : sig ... end

module Filtering : sig ... end

module HackClassName : sig ... end

module Ident : sig ... end

Identifiers: program variables and logical variables

module Inferconfig : sig ... end

module Instrs : sig ... end

Manipulate possibly-reversed lists of instructions efficiently

module IntLit : sig ... end

module IntegerWidths : sig ... end

The Smallfoot Intermediate Language: Types

module Io_infer : sig ... end

Module to handle IO. Includes html and xml modules.

module JConfig : sig ... end

module JavaClassName : sig ... end

module Mangled : sig ... end

Module for Mangled Names

module MissingDependencies : sig ... end

Stateful store of procedure and sourcefile dependencies found to be missing from the capture database during analysis.

module ObjCDispatchModels : sig ... end

module PredSymb : sig ... end

The Smallfoot Intermediate Language: Predicate Symbols

module ProcAttributes : sig ... end

module Procdesc : sig ... end

module Procname : sig ... end

module Pvar : sig ... end

Program variables.

module PythonClassName : sig ... end

module QualifiedCppName : sig ... end

module Sil : sig ... end

The Smallfoot Intermediate Language

module SourceFiles : sig ... end

module Specialization : sig ... end

module SpecializeProcdesc : sig ... end

module StdTyp : sig ... end

module Struct : sig ... end

module Subtype : sig ... end

The Smallfoot Intermediate Language: Subtypes

module Tenv : sig ... end

Module for Type Environments.

module Typ : sig ... end

The Smallfoot Intermediate Language: Types

module Unop : sig ... end

The Smallfoot Intermediate Language: Unary Operators

module Var : sig ... end

Single abstraction for all the kinds of variables in SIL

module WeakTopologicalOrder : sig ... end