Module Procname.CSharp

Type of csharp procedure names.

type kind =

| Non_Static	(*
in Java, procedures called with invokevirtual, invokespecial, and invokeinterface
*)

| Static	(*
in Java, procedures called with invokestatic
*)

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val constructor_method_name : string

val get_method : t -> string

Return the method name of a csharp procedure name.

val get_class_type_name : t -> Typ.Name.t

Return the class name as a typename of a java procedure name.

val get_return_typ : t -> Typ.t

Return the return type of pname_csharp. return Tvoid if there's no return type

val get_class_name : t -> string

Return the class name of a java procedure name.

val is_generated : t -> bool

Check if the proc name comes from generated code