Module Procname.Hack

type t = private {

class_name : HackClassName.t option;

function_name : string;

arity : int option;

}

Hack procedure is identified by the class and function names and its arity. The arity can be absent for external function declarations of the form declare F.f(...): ...

TODO(arr): consider making the arity non-optional if we remove function declarations from the Tenv

val get_class_name_as_a_string : t -> string option