Module IR.PythonClassName

module F = Stdlib.Format

type builtin_type =

| PyBool

| PyDict

| PyInt

| PyNone

| PyObject

| PyString

| PyTuple

val compare_builtin_type : builtin_type -> builtin_type -> int

val equal_builtin_type : builtin_type -> builtin_type -> bool

val yojson_of_builtin_type : builtin_type -> Ppx_yojson_conv_lib.Yojson.Safe.t

val sexp_of_builtin_type : builtin_type -> Sexplib0.Sexp.t

val builtin_type_of_sexp : Sexplib0.Sexp.t -> builtin_type

val hash_fold_builtin_type : Ppx_hash_lib.Std.Hash.state -> builtin_type -> Ppx_hash_lib.Std.Hash.state

val hash_builtin_type : builtin_type -> Ppx_hash_lib.Std.Hash.hash_value

val hash_normalize_builtin_type : builtin_type -> builtin_type

val hash_normalize_builtin_type_opt : builtin_type option -> builtin_type option

val hash_normalize_builtin_type_list : builtin_type list -> builtin_type list

type builtin_closure =

| DictFun

| IntFun

| StrFun

| TypeFun

val compare_builtin_closure : builtin_closure -> builtin_closure -> int

val equal_builtin_closure : builtin_closure -> builtin_closure -> bool

val yojson_of_builtin_closure : builtin_closure -> Ppx_yojson_conv_lib.Yojson.Safe.t

val sexp_of_builtin_closure : builtin_closure -> Sexplib0.Sexp.t

val builtin_closure_of_sexp : Sexplib0.Sexp.t -> builtin_closure

val hash_fold_builtin_closure : Ppx_hash_lib.Std.Hash.state -> builtin_closure -> Ppx_hash_lib.Std.Hash.state

val hash_builtin_closure : builtin_closure -> Ppx_hash_lib.Std.Hash.hash_value

val hash_normalize_builtin_closure : builtin_closure -> builtin_closure

val hash_normalize_builtin_closure_opt : builtin_closure option -> builtin_closure option

val hash_normalize_builtin_closure_list : builtin_closure list -> builtin_closure list

type t =

| Builtin of builtin_type

| Globals of string

| Closure of string

| BuiltinClosure of builtin_closure

| ClassCompanion of {

module_name : string;

class_name : string;

}

| ClassInstance of {

module_name : string;

class_name : string;

}

| ModuleAttribute of {

module_name : string;

attr_name : string;

}

| Filename of string

| Package of string

| Wildcard

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val hash_normalize : t -> t

val hash_normalize_opt : t option -> t option

val hash_normalize_list : t list -> t list

val classname : t -> string

val components : t -> string list

val pp : F.formatter -> t -> unit

val to_string : t -> string

val is_final : t -> bool

val is_singleton : t -> bool

val split_module_attr : t -> (string * string) option

if the argument is a closure type or a module type, split the corresponding name into a pair (module_name, function_name)

val concatenate_package_name_and_file_name : t -> string -> t option

val get_builtin_closure_from_builtin_type : builtin_type -> builtin_closure option