Module Struct.ClassInfo

type t =

| NoInfo

| CppClassInfo of {

is_trivially_copyable : bool;

}
(*
class kind in C++
*)

| JavaClassInfo of {

kind : java_class_kind;
(*
class kind in Java
*)

loc : IBase.Location.t option;
(*
None should correspond to rare cases when it was impossible to fetch the location in source file
*)

}

| HackClassInfo of hack_class_kind

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val pp : Ppx_show_runtime.Format.formatter -> t -> unit

val show : t -> string