Module MethodInfo.Hack

type kind = private

| IsClass
(*
Normal method call
*)

| IsTrait of {

used : Typ.Name.t;

is_direct : bool;

}
(*
Trait method call: used is the name of the class uses the trait. If it is a direct trait method call, e.g. Trait::foo, used is the name of the trait.
*)