Module IR__Attributes

Module to manage the table of attributes.

type attributes_kind

val deserialize_attributes_kind : Sqlite3.Data.t -> attributes_kind

val store : proc_desc:IR.Procdesc.t option -> IR.ProcAttributes.t -> unit

Save .attr file for the procedure into the attributes database.

val load : IR.Procname.t -> IR.ProcAttributes.t option

Load the attributes for the procedure from the attributes database.

val is_no_return : IR.Procname.t -> bool

val pp_attributes_kind : Stdlib.Format.formatter -> attributes_kind -> unit