Module IR__DecompiledExp

The Smallfoot Intermediate Language: Decompiled Expressions

module F = Stdlib.Format

type t =

| Darray of t * t

| Dbinop of IR.Binop.t * t * t

| Dconst of IR.Const.t

| Dsizeof of IR.Typ.t * t option * IR.Subtype.t

| Dderef of t

| Dfcall of t * t list * IBase.Location.t * IR.CallFlags.t

| Darrow of t * IR.Fieldname.t

| Ddot of t * IR.Fieldname.t

| Dpvar of IR.Pvar.t

| Dpvaraddr of IR.Pvar.t

| Dunop of IR.Unop.t * t

| Dunknown

| Dretcall of t * t list * IBase.Location.t * IR.CallFlags.t

expression representing the result of decompilation

type vpath = t option

Value paths: identify an occurrence of a value in a symbolic heap each expression represents a path, with Dpvar being the simplest one

val pp : F.formatter -> t -> unit

val to_string : t -> string

convert to a string

val pp_vpath : IStdlib.Pp.env -> F.formatter -> vpath -> unit

Pretty print a value path

val has_tmp_var : t -> bool

return true if dexp contains a temporary pvar