Module IR__Procname.ObjC_Cpp

type kind =

| CPPMethod of {

mangled : string option;

}

| CPPConstructor of {

mangled : string option;

}

| CPPDestructor of {

mangled : string option;

}

| ObjCClassMethod

| ObjCInstanceMethod

| ObjCInternalMethod

val compare_kind : kind -> kind -> int

type t = {

class_name : IR.Typ.Name.t;

kind : kind;

method_name : string;

parameters : Parameter.clang_parameter list;

template_args : IR.Typ.template_spec_info;

}

Type of Objective C and C++ procedure names: method signatures.

val compare : t -> t -> int

val make : IR.Typ.Name.t -> string -> kind -> IR.Typ.template_spec_info -> Parameter.clang_parameter list -> t

Create an objc procedure name from a class_name and method_name.

val get_class_name : t -> string

val get_class_type_name : t -> IR.Typ.Name.t

val get_class_qualifiers : t -> IR.QualifiedCppName.t

val objc_method_kind_of_bool : bool -> kind

Create ObjC method type from a bool is_instance.

val is_objc_constructor : string -> bool

Check if this is a constructor method in Objective-C.

val is_objc_dealloc : string -> bool

Check if this is a dealloc method in Objective-C.

val is_destructor : t -> bool

Check if this is a dealloc method.

val is_inner_destructor : t -> bool

Check if this is a frontend-generated "inner" destructor (see D5834555/D7189239)