Module IR__Typ.Name

type t = name

Named types.

val compare : t -> t -> int

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val loose_compare : t -> t -> int

Similar to compare, but addresses CStruct x and CppClass x as equal.

val compare_name : t -> t -> int

Similar to compare, but compares only names, except template arguments.

val equal : t -> t -> bool

Equality for typenames

val hash : t -> int

val to_string : t -> string

convert the typename to a string

val pp : Stdlib.Format.formatter -> t -> unit

val is_class : t -> bool

is_class name holds if name names CPP/Objc/Java class

val is_union : t -> bool

is_union name holds if name names C/CPP union

val is_same_type : t -> t -> bool

is_class name1 name2 holds if name1 and name2 name same kind of type

val name : t -> string

name of the typename without qualifier

val qual_name : t -> IR.QualifiedCppName.t

qualified name of the type, may return nonsense for Java classes

val unqualified_name : t -> IR.QualifiedCppName.t

val get_template_spec_info : t -> template_spec_info option

val is_objc_protocol : t -> bool

module C : sig ... end

module CSharp : sig ... end

module Java : sig ... end

module Cpp : sig ... end

module Objc : sig ... end

module Set : IStdlib.PrettyPrintable.PPSet with type PPSet.elt = t

module Map : IStdlib.PrettyPrintable.PPMap with type PPMap.key = t

module Normalizer : IStdlib.HashNormalizer.S with type t = t