Module Name.Java

val from_string : string -> t

Create a typename from a Java classname in the form "package.class"

val is_class : t -> bool

is_class name holds if name names a Java class

val get_java_class_name_exn : t -> IR.JavaClassName.t

Ensure name is a java class name and return underlying JavaClassName

val is_external : t -> bool

return true if the typename is in the .inferconfig list of external classes

val is_anonymous_inner_class_name_exn : t -> bool

Throws if it is not a Java class

val is_anonymous_inner_class_name_opt : t -> bool option

return None if it is not a Java class