Module IR__Var

type t = private

| LogicalVar of IR.Ident.t

| ProgramVar of IR.Pvar.t

val compare : t -> t -> int

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val equal : t -> t -> bool

val compare_modulo_this : t -> t -> int

val of_id : IR.Ident.t -> t

val of_pvar : IR.Pvar.t -> t

val of_formal_index : int -> t

Create a variable representing the ith formal of the current procedure

val get_all_vars_in_exp : IR.Exp.t -> t IStdlib.IStd.Sequence.t

Get all free and program vars

val to_exp : t -> IR.Exp.t

val get_ident : t -> IR.Ident.t option

val get_pvar : t -> IR.Pvar.t option

val is_pvar : t -> bool

val is_global : t -> bool

val is_local_to_procedure : IR.Procname.t -> t -> bool

val is_return : t -> bool

val is_footprint : t -> bool

val is_none : t -> bool

val is_this : t -> bool

val appears_in_source_code : t -> bool

return true if this variable appears in source code (i.e., is not a LogicalVar or a frontend-generated ProgramVar)

val is_cpp_temporary : t -> bool

val get_footprint_index : t -> int option

val pp : Stdlib.Format.formatter -> t -> unit

module Map : IStdlib.PrettyPrintable.PPMap with type PPMap.key = t