Module IStdlib

module ARList : sig ... end

module Escape : sig ... end

Escape a string, eg replace reserved characters with escape sequences

module GOption : sig ... end

module HashNormalizer : sig ... end

module ICompare : sig ... end

module IContainer : sig ... end

module IExn : sig ... end

module IList : sig ... end

module IOption : sig ... end

module IResult : sig ... end

module ISequence : sig ... end

Utility functions for sequences

module IStd : sig ... end

module ISys : sig ... end

module ImperativeUnionFind : sig ... end

module IntSet : sig ... end

module LRUHashtbl : sig ... end

Hash table the size of which is limited by LRU

module PartialOrder : sig ... end

module PhysEqual : sig ... end

Helpers function to enforce physical equality.

module PolyVariantEqual : sig ... end

Open to bring equality (=) for polymorphic variants into scope

module Pp : sig ... end

module PrettyPrintable : sig ... end

module RecencyMap : sig ... end

module RevList : sig ... end

module SpecialChars : sig ... end

module UnionFind : sig ... end