Module IStdlib.IExn

val reraise_after : f:(unit -> unit) -> exn -> 'a

Reraise the exception after doing f. Always reraise immediately after catching the exception, otherwise the backtrace can be wrong.

val reraise_if : f:(unit -> bool) -> exn -> unit

Reraise the exception if f returns true. Always reraise immediately after catching the exception, otherwise the backtrace can be wrong.