Module IStdlib.IFilename

val open_temp_file : ?perm:int -> ?in_dir:string -> string -> string -> string * IStdlib.IStd.Out_channel.t

escapes prefix and suffix to avoid failures related to invalid characters in filenames

val temp_file : ?perm:int -> ?in_dir:string -> string -> string -> string

escapes prefix and suffix to avoid failures related to invalid characters in filenames

val temp_dir : ?perm:int -> ?in_dir:string -> string -> string -> string