Module IStdlib.IFmt

type 'a t := 'a Fmt.t

Low-level combinators built on Fmt. Sophisticated pretty-printing functions should rather go in Pp.

val colon_sp : 'a t

colon_sp is Fmt.any ":@ ". It prints a colon and a space break hint.

val if' : bool -> 'a t -> 'a t

Should be directly in Fmt >=0.9.1.

if' bool pp is pp if bool is true and Fmt.nop otherwise.

module Labelled : sig ... end

Adapters of Fmt functions using labelled interfaces