Module IStdlib.RecencyMap

module F = Stdlib.Format

module type Config = sig ... end

module type S = sig ... end

A functional map interface where only the N most recently-accessed elements are guaranteed to be persisted, similarly to an LRU cache. The map stores at most 2*N elements.

module Make (Key : PrettyPrintable.PrintableEquatableOrderedType) (Value : PrettyPrintable.PrintableOrderedType) (Config : Config) : S with type key = Key.t and type value = Value.t