Module IStdlib__PrettyPrintable.MakePPUniqRankSet

Parameters

Rank : PrintableEquatableOrderedType

Val : PrintableRankedType with type rank = Rank.t

Signature

type t

val compare : t -> t -> int

val equal : t -> t -> bool

type rank = Rank.t

type elt = Val.t

val add : t -> elt -> t

val empty : t

val find_rank : t -> rank -> elt option

val fold : t -> init:'accum -> f:('accum -> elt -> 'accum) -> 'accum

val fold_map : t -> init:'accum -> f:('accum -> elt -> 'accum * elt) -> 'accum * t

val for_all : f:(elt -> bool) -> t -> bool

val is_empty : t -> bool

val is_singleton : t -> bool

val is_subset : t -> of_:t -> bool

val map : t -> f:(elt -> elt) -> t

val singleton : elt -> t

val elements : t -> elt list

val remove : elt -> t -> t

val mem : elt -> t -> bool

val union_prefer_left : t -> t -> t

in case an element with the same rank is present both in lhs and rhs, keep the one from lhs in union_prefer_left lhs rhs

val pp : ?⁠print_rank:bool -> F.formatter -> t -> unit