Module InferModules.AccessPathDomains

module Set : sig ... end

add-only set of access paths. To make common operations efficient (namely, add, join, and widen), the set is allowed to contain elements whose concretization is redundant (e.g., x* and x.f). these redundancies can be eliminated by expressing the set in its canonical form via a call to normalize. however, normalize is quadratic in the size of the set, so it should be used sparingly (recommendation: only before computing a summary based on the access path set)