Module InferModules.OndemandCapture

module L = InferBase.Logging

val compilation_db : CompilationDatabase.t lazy_t

val try_capture : InferIR.ProcAttributes.t -> InferIR.ProcAttributes.t option

Given proc_attributes try to produce proc_attributes' where proc_attributes'.is_defined = true It may trigger capture of extra files to do so and when it does, it waits for frontend to finish before returning