Module InferModules.Printer

module LineReader : sig ... end

Module to read specific lines from files. The data from any file will stay in memory until the handle is collected by the gc

val curr_html_formatter : Stdlib.Format.formatter InferStdlib.IStd.ref

Current html formatter

val force_delayed_prints : unit -> unit

Execute the delayed print actions

val node_finish_session : InferIR.Procdesc.Node.t -> unit

Finish a session, and perform delayed print actions if required

val node_start_session : pp_name:(Stdlib.Format.formatter -> unit) -> InferIR.Procdesc.Node.t -> int -> unit

Start a session, and create a new html fine for the node if it does not exist yet

val write_proc_html : InferIR.Procdesc.t -> unit

Write html file for the procedure.

val write_all_html_files : InferBase.SourceFile.t -> unit

Create filename.ext.html for each file in the cluster.