Module Printer.LineReader

Module to read specific lines from files. The data from any file will stay in memory until the handle is collected by the gc

type t

val create : unit -> t

create a line reader

val from_file_linenum : t -> InferBase.SourceFile.t -> int -> string option

get the line from a source file and line number

val from_loc : t -> InferBase.Location.t -> string option

get the line from a location looking for the copy of the file in the results dir