Module StarvationDomain.Lock

Abstraction of a path that represents a lock, special-casing equality and comparisons to work over type, base variable modulo this and access list

include ExplicitTrace.Element with type Element.t = InferIR.AccessPath.t

type t = InferIR.AccessPath.t

val compare : t -> t -> int

val pp : InferStdlib.PrettyPrintable.F.formatter -> t -> unit

val pp_human : Stdlib.Format.formatter -> t -> unit

val pp_call : Stdlib.Format.formatter -> InferIR.CallSite.t -> unit

val owner_class : t -> InferIR.Typ.name option

Class of the root variable of the path representing the lock

val equal : t -> t -> bool

val pp_locks : F.formatter -> t -> unit