Module Summary.SummaryServer

module Value = SummaryValue

val get : key:string -> Value.t option

get a value, None means no key exists

val set : key:string -> Value.t -> unit

set a key/value pair. NB we swallow failures due to exceeding max value size currently. This will need to be changed before memcached is used as a primary store.