Module SummaryPayload.Make

Parameters

P : Payload

Signature

type t = P.t

val update_summary : t -> Summary.t -> Summary.t

Update the corresponding part of the payload in the procedure summary

val of_summary : Summary.t -> t option

Read the corresponding part of the payload from the procedure summary

val read_full : caller_summary:Summary.t -> callee_pname:InferIR.Typ.Procname.t -> (InferIR.Procdesc.t * t) option

Return the proc desc and payload for the given procedure. Runs the analysis on-demand if necessary.

val read : caller_summary:Summary.t -> callee_pname:InferIR.Typ.Procname.t -> t option

Return the payload for the given procedure. Runs the analysis on-demand if necessary.

val read_toplevel_procedure : InferIR.Typ.Procname.t -> t option