Module InferModules__AbstractInterpreter

module VisitCount : sig ... end

module State : sig ... end

module type S = sig ... end

type of an intraprocedural abstract interpreter

module type Make = functor (TransferFunctions : InferModules.TransferFunctions.SIL) -> S with module TransferFunctions = TransferFunctions

module MakeRPO : Make

create an intraprocedural abstract interpreter from transfer functions using the reverse post-order scheduler

module MakeWTO : Make

create an intraprocedural abstract interpreter from transfer functions using Bourdoncle's strongly connected component weak topological order