Module InferModules__AndroidFramework

Android lifecycle types and their lifecycle methods that are called by the framework

val drawable_prefix : string

prefix for Drawable fields in generated resources

val is_autocloseable : InferIR.Tenv.t -> InferIR.Typ.Name.t -> bool

val is_view : InferIR.Tenv.t -> InferIR.Typ.Name.t -> bool

return true if typename <: android.view.View

val is_fragment : InferIR.Tenv.t -> InferIR.Typ.Name.t -> bool

val is_destroy_method : InferIR.Typ.Procname.t -> bool

return true if procname is a special lifecycle cleanup method