Module InferModules__Exe_env

Execution environments: basically a cache of where procedures are and what is their type environment

type file_data

type t = private {

proc_map : file_data InferIR.Typ.Procname.Hash.t;
map from procedure name to file data

file_map : file_data InferBase.SourceFile.Hash.t;
map from source files to file data

}

val mk : unit -> t

Create a new cache

val get_tenv : t -> InferIR.Typ.Procname.t -> InferIR.Tenv.t

return the type environment associated with the procedure

val get_integer_type_widths : t -> InferIR.Typ.Procname.t -> InferIR.Typ.IntegerWidths.t

return the integer type widths associated with the procedure