Module InferModules__Models

module Hashtbl = InferStdlib.IStd.Caml.Hashtbl

val match_method_name : InferIR.Typ.Procname.t -> InferStdlib.IStd.String.t -> bool

val table_has_procedure : (string, 'a) Hashtbl.t -> InferIR.Typ.Procname.t -> bool

val get_modelled_annotated_signature : InferIR.ProcAttributes.t -> InferModules.AnnotatedSignature.t

Return the annotated signature of the procedure, taking into account models.

val is_modelled_nullable : InferIR.Typ.Procname.t -> bool

Return true when the procedure has been modelled for nullable.

val is_check_not_null : InferIR.Typ.Procname.t -> bool

Check if the procedure is one of the known Preconditions.checkNotNull.

val get_check_not_null_parameter : InferIR.Typ.Procname.t -> int option

Parameter number for a procedure known to be a checkNotNull

val is_check_state : InferIR.Typ.Procname.t -> bool

Check if the procedure is one of the known Preconditions.checkState.

val is_check_argument : InferIR.Typ.Procname.t -> bool

Check if the procedure is one of the known Preconditions.checkArgument.

val is_noreturn : InferIR.Typ.Procname.t -> bool

Check if the procedure does not return.

val is_optional_get : InferIR.Typ.Procname.t -> bool

Check if the procedure is Optional.get().

val is_optional_isPresent : InferIR.Typ.Procname.t -> bool

Check if the procedure is Optional.isPresent().

val is_true_on_null : InferIR.Typ.Procname.t -> bool

Check if the procedure returns true on null.

val is_containsKey : InferIR.Typ.Procname.t -> bool

Check if the procedure is Map.containsKey().

val is_mapPut : InferIR.Typ.Procname.t -> bool

Check if the procedure is Map.put().