Module InferModules__ObjcInterface_decl

In this module an ObjC interface declaration is processed. The class is saved in the tenv as a struct with the corresponding fields, potential superclass and list of defined methods

val interface_declaration : InferModules.CAst_utils.qual_type_to_sil_type -> InferModules.CAst_utils.procname_from_decl -> InferIR.Tenv.t -> InferGenerated.Clang_ast_t.decl -> InferIR.Typ.desc

val interface_impl_declaration : InferModules.CAst_utils.qual_type_to_sil_type -> InferModules.CAst_utils.procname_from_decl -> InferIR.Tenv.t -> InferGenerated.Clang_ast_t.decl -> InferIR.Typ.desc