Module InferModules__Ondemand

Module for on-demand analysis.

val get_proc_desc : InferIR.Typ.Procname.t -> InferIR.Procdesc.t option

Find a proc desc for the procedure, perhaps loading it from disk.

val analyze_proc_desc : caller_summary:InferModules.Summary.t -> InferIR.Procdesc.t -> InferModules.Summary.t option

analyze_proc_desc ~caller_summary callee_pdesc performs an on-demand analysis of callee_pdesc triggered during the analysis of caller_summary

val analyze_proc_name : caller_summary:InferModules.Summary.t -> InferIR.Typ.Procname.t -> InferModules.Summary.t option

analyze_proc_name ~caller_summary callee_pname performs an on-demand analysis of callee_pname triggered during the analysis of caller_summary

val analyze_proc_name_no_caller : InferIR.Typ.Procname.t -> InferModules.Summary.t option

analyze_proc_name_no_caller callee_pname performs an on-demand analysis of callee_pname triggered by the top-level of a cluster checker

val set_exe_env : InferModules.Exe_env.t -> unit

Set the execution enviroment used during on-demand analysis.

module LocalCache : sig ... end

val analyze_file : InferModules.Exe_env.t -> InferBase.SourceFile.t -> unit

Invoke all the callbacks registered in Callbacks on the given file.

val analyze_proc_name_toplevel : InferModules.Exe_env.t -> InferIR.Typ.Procname.t -> unit

Invoke all the callbacks registered in Callbacks on the given procedure.