Module InferModules__PulseAbductiveDomain.Memory

stack operations like PulseDomain.Heap but that also take care of propagating facts to the precondition

module Access = InferModules.PulseDomain.Memory.Access

module Edges = InferModules.PulseDomain.Memory.Edges

val add_attributes : AbstractAddress.t -> Attributes.t -> t -> t

val add_edge : AbstractAddress.t -> Access.t -> InferModules.PulseDomain.AddrTracePair.t -> t -> t

val check_valid : unit InferModules.PulseDomain.InterprocAction.t -> AbstractAddress.t -> t -> (t, InferModules.PulseDomain.Invalidation.t InferModules.PulseDomain.Trace.t) InferStdlib.IStd.result

val find_opt : AbstractAddress.t -> t -> InferModules.PulseDomain.Memory.cell option

val find_edge_opt : AbstractAddress.t -> Access.t -> t -> InferModules.PulseDomain.AddrTracePair.t option

val set_cell : AbstractAddress.t -> InferModules.PulseDomain.Memory.cell -> t -> t

val invalidate : (AbstractAddress.t * InferModules.PulseDomain.ValueHistory.t) -> InferModules.PulseDomain.Invalidation.t InferModules.PulseDomain.InterprocAction.t -> t -> t

val get_closure_proc_name : AbstractAddress.t -> t -> InferIR.Typ.Procname.t option

val is_std_vector_reserved : AbstractAddress.t -> t -> bool

val std_vector_reserve : AbstractAddress.t -> t -> t

val eval_edge : AbstractAddress.t -> Access.t -> t -> t * InferModules.PulseDomain.AddrTracePair.t

eval_edge addr access astate follows the edge addr --access--> . in memory and returns what it points to or creates a fresh value if that edge didn't exist.