Module InferModules__PulseDomain.Attributes

include InferStdlib.PrettyPrintable.PPUniqRankSet with type elt = Attribute.t

type t

type elt = Attribute.t

val add : t -> elt -> t

val empty : t

val equal : t -> t -> bool

val find_rank : t -> int -> elt option

val fold : t -> init:'accum -> f:('accum -> elt -> 'accum) -> 'accum

val is_empty : t -> bool

val is_singleton : t -> bool

val is_subset : t -> of_:t -> bool

val map : t -> f:(elt -> elt) -> t

val singleton : elt -> t

val union_prefer_left : t -> t -> t

in case an element with the same rank is present both in lhs and rhs, keep the one from lhs in union_prefer_left lhs rhs

val pp : InferStdlib.PrettyPrintable.F.formatter -> t -> unit

val get_must_be_valid : t -> unit InterprocAction.t option

val get_address_of_stack_variable : t -> (InferIR.Var.t * ValueHistory.t * InferBase.Location.t) option

val is_modified : t -> bool