Module InferModules__RacerDDomain.AttributeMapDomain

type t

val find : InferIR.AccessPath.t -> t -> AttributeSetDomain.t

val add : InferIR.AccessPath.t -> AttributeSetDomain.t -> t -> t

val has_attribute : InferIR.AccessPath.t -> Attribute.t -> t -> bool

val get_choices : InferIR.AccessPath.t -> t -> Choice.t list

get the choice attributes associated with the given access path

val add_attribute : InferIR.AccessPath.t -> Attribute.t -> t -> t

val propagate_assignment : InferIR.AccessPath.t -> InferIR.HilExp.t -> t -> t

propagate attributes from the leaves to the root of an RHS Hil expression