Module InferModules__RacerDDomain.TraceElem

include InferModules.ExplicitTrace.TraceElem with type elem_t = Access.t

type elem_t = Access.t

type t = private {

elem : elem_t;

loc : InferBase.Location.t;

trace : InferIR.CallSite.t list;

}

An elem which occured at loc, after the chain of steps (usually calls) in trace.

Both pp and pp_human simply call the same function on the trace element.

include InferModules.ExplicitTrace.Element with type Element.t := t

include InferStdlib.PrettyPrintable.PrintableOrderedType

include InferStdlib.IStd.Caml.Set.OrderedType

type t

val compare : t -> t -> int

include InferStdlib.PrettyPrintable.PrintableType with type t := t

type t

val pp : InferStdlib.PrettyPrintable.F.formatter -> t -> unit

val pp_human : Stdlib.Format.formatter -> t -> unit

Pretty printer used for trace construction; pp is used for debug output.

val pp_call : Stdlib.Format.formatter -> InferIR.CallSite.t -> unit

val make : elem_t -> InferBase.Location.t -> t

val map : f:(elem_t -> elem_t) -> t -> t

val get_loc : t -> InferBase.Location.t

Starting location of the trace: this is either loc if trace==[], or the head of trace.

val make_loc_trace : ?⁠nesting:int -> t -> InferIR.Errlog.loc_trace

val with_callsite : t -> InferIR.CallSite.t -> t

Push given callsite onto trace, extending the call chain by one.

module FiniteSet : InferModules.ExplicitTrace.FiniteSet with type FiniteSet.elt = t

A powerset of traces.

val is_write : t -> bool

is it a write OR a container write

val is_container_write : t -> bool

val map : f:(InferIR.AccessPath.t -> InferIR.AccessPath.t) -> t -> t

val make_container_access : InferIR.AccessPath.t -> InferIR.Typ.Procname.t -> is_write:bool -> InferBase.Location.t -> t

val make_field_access : InferIR.AccessPath.t -> is_write:bool -> InferBase.Location.t -> t