Module InferModules__SymExecBlocks

val resolve_method_with_block_args_and_analyze : caller_summary:InferModules.Summary.t -> InferIR.Typ.Procname.t -> (InferIR.Exp.t * InferIR.Typ.t) list -> (InferModules.Summary.t * (InferIR.Exp.t * InferIR.Typ.t) list) option

resolve_method_with_block_args_and_analyze caller_pdesc pname args create a copy of the method pname if it is defined and it's called with the correct number of arguments, and some arguments are block closures. The copy is created by adding extra formals for each captured variable, and by swapping the calls to the block arguments to the calls to the concrete blocks. The new procedure is analyzed and the possibly computed summary is returned together with the list of arguments where the closures where swapped by their captured variables.