Module InferModules__ToplMonitor

val generate : InferModules.ToplAutomaton.t -> InferIR.Typ.Procname.t -> InferIR.Procdesc.t option

generate automaton proc_name returns a CFG, provided that proc_name is a recognized procedure name

val sourcefile : unit -> InferBase.SourceFile.t

For debug.

val cfg : unit -> InferIR.Cfg.t

For debug. This datastructure accumulates all the procedures that were synthesized by the current process. If the implementation is correct, then different processes synthesize the same procedures, given the same set of Topl properties. However, for debug, we print the datastructure in a filename that contains the PID, which is why sourcefile is exposed.