Module Integration.CaptureSILJson

Capture module for the json analysis in the capture phase

val capture : cfg_json:string -> tenv_json:string -> unit

Run the capture of the files for which we have cfg in cfg_json, type environment tenv_json in and changed_files, if specified.