Module Integration.ConfigImpactPostProcess

val instantiate_unchecked_callees_cond : Costlib.ConfigImpactAnalysis.Summary.t -> Costlib.ConfigImpactAnalysis.Summary.t

Instantiate conditional unchecked callees with collected config fields. WARN: This function must be called after the summaries for all procedures completed, because it evaluates a lazy value internally.

val is_in_gated_classes : IR.Procname.t -> bool

Check if a procedure is in a gated class. WARN: This function must be called after the summaries for all procedures completed, because it evaluates a lazy value internally.