Module JavaFrontend.JFrontend

val path_of_cached_classname : Javalib_pack.JBasics.class_name -> string

path_of_cached_classname cn returns the path of a cached classname

val cache_classname : Javalib_pack.JBasics.class_name -> unit

cache_classname cn stores the classname to the disk

val is_classname_cached : Javalib_pack.JBasics.class_name -> bool

is_classname_cached cn

val compute_source_icfg : JProgramDesc.t -> IR.Tenv.t -> string -> string option -> IBase.SourceFile.t -> IR.Cfg.t

compute_cfg linereader program tenv source_basename source_file create the control flow graph for the file source_file by translating all the classes in program originating from source_file

val compute_class_icfg : IBase.SourceFile.t -> JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JCode.jcode Javalib_pack.Javalib.interface_or_class -> IR.Cfg.t

Compute the CFG for a class