Module JavaFrontend.JTransType

val create_fieldname : Javalib_pack.JBasics.class_name -> Javalib_pack.JBasics.field_signature -> IR.Fieldname.t

translate the name of the field

val get_method_kind : Javalib_pack.JCode.jcode Javalib_pack.Javalib.jmethod -> IR.Procname.Java.kind

val get_method_procname : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.class_name -> Javalib_pack.JBasics.method_signature -> IR.Procname.Java.kind -> IR.Procname.t

returns a procedure name based on the class name and the method's signature

val translate_method_name : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JCode.jcode Javalib_pack.Javalib.jmethod -> IR.Procname.t

translate the SIL procedure name of the Java method

val get_class_struct_typ : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.class_name -> IR.Struct.t

get_class_struct_typ program tenv cn returns the struct_typ representation of the class

val get_class_type_no_pointer : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.class_name -> IR.Typ.t

get_class_type_no_pointer program tenv cn returns the sil type representation of the class without the pointer part

val get_class_type : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.class_name -> IR.Typ.t

get_class_type program tenv cn returns the sil type representation of the class

val is_autogenerated_assert_field : IR.Fieldname.t -> bool

return true if field_name is the autogenerated C.$assertionsDisabled field for class C

val sizeof_of_object_type : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.object_type -> IR.Subtype.t -> IR.Exp.t

create sizeof expressions from the object type and the list of subtypes

val value_type : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.value_type -> IR.Typ.t

transforms a Java type to a Typ.t.

val param_type : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.class_name -> Sawja_pack.JBir.var -> Javalib_pack.JBasics.value_type -> IR.Typ.t

return the type of a formal parameter, looking up the class name in case of "this"

val return_type : JProgramDesc.t -> IR.Tenv.t -> Javalib_pack.JBasics.method_signature -> IR.Typ.t

Returns the return type of the method based on the return type specified in ms.

val expr_type : JContext.t -> Sawja_pack.JBir.expr -> IR.Typ.t

translates the type of an expression

val cast_type : Sawja_pack.JBir.conv -> IR.Typ.t

translates a conversion type from Java to Sil.

val create_array_type : IR.Typ.t -> int -> IR.Typ.t

create_array_type typ dim creates an array type with dimension dim and content typ

val object_type_to_string : Javalib_pack.JBasics.object_type -> string

returns a string representation of an object Java type