Module Global.Set

include NS.Set.S with type elt := t

type compare_elt

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

type ('a, 'compare_a) comparer := 'a -> 'a -> int

type compare

compare types are equipped with functions to support use of @@deriving compare, equal, sexp on types parameterized by such singleton types for compare functions. These derived functions are never actually called, since the compare type parameters are phantom.

val compare_compare : compare -> compare -> int

val equal_compare : compare -> compare -> bool

val sexp_of_compare : compare -> Sexplib0.Sexp.t

val compare_of_sexp : Sexplib0.Sexp.t -> compare

val comparer : (t, compare) comparer

module Provide_of_sexp (_ : sig ... end) : sig ... end

Pretty-print

val pp_full : ?pre:(unit, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> ?suf:(unit, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> ?sep:(unit, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> (Stdlib.Format.formatter -> t -> unit) -> Stdlib.Format.formatter -> t -> unit

module Provide_pp (_ : sig ... end) : sig ... end

Construct

val empty : t

val of_ : t -> t

val of_option : t option -> t

val of_list : t list -> t

val add : t -> t -> t

val add_option : t option -> t -> t

val add_list : t list -> t -> t

val remove : t -> t -> t

val diff : t -> t -> t

val inter : t -> t -> t

val union : t -> t -> t

val diff_inter : t -> t -> t * t

val diff_inter_diff : t -> t -> t * t * t

val union_list : t list -> t

Query

val is_empty : t -> bool

val cardinal : t -> int

val mem : t -> t -> bool

val subset : t -> of_:t -> bool

val disjoint : t -> t -> bool

val max_elt : t -> t option

val only_elt : t -> t option

val classify : t -> t NS__.NS0.zero_one_many

val choose : t -> t option

Find an unspecified element. Different elements may be chosen for equivalent sets. O(1).

val choose_exn : t -> t

Find an unspecified element. Different elements may be chosen for equivalent sets. O(1).

val pop : t -> (t * t) option

Find and remove an unspecified element. O(1).

val pop_exn : t -> t * t

Find and remove an unspecified element. O(1).

val pop_min_elt_exn : t -> t * t

Find and remove the minimum element. O(1).

val split : t -> t -> t * bool * t

val find : t -> t -> t

Transform

val map : t -> f:(t -> t) -> t

val flat_map : t -> f:(t -> t) -> t

val filter : t -> f:(t -> bool) -> t

val filter_map : t -> f:(t -> t option) -> t

Map f over a set, removing any element e where f e = None.

val partition : t -> f:(t -> bool) -> t * t

Traverse

val iter : t -> f:(t -> unit) -> unit

val exists : t -> f:(t -> bool) -> bool

val for_all : t -> f:(t -> bool) -> bool

val fold : t -> 's -> f:(t -> 's -> 's) -> 's

val reduce : t -> f:(t -> t -> t) -> t option

val fold_map : t -> 's -> f:(t -> 's -> t * 's) -> t * 's

Map f over a set while threading an accumulator through calls to f.

Convert

val to_list : t -> t list

val to_iter : t -> (t -> unit) -> unit

val of_iter : ((t -> unit) -> unit) -> t

include sig ... end

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val pp : t NS.pp