Module Llair.Func

type t = func

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val pp : t NS.pp

val pp_call : t call NS.pp

include NS.Invariant.S with type t := t

val invariant : t -> unit

val mk : name:FuncName.t -> formals:Reg.t NS.iarray -> freturn:Reg.t option -> fthrow:Reg.t -> entry:block -> cfg:block NS.iarray -> loc:Loc.t -> t

val mk_undefined : name:FuncName.t -> formals:Reg.t NS.iarray -> freturn:Reg.t option -> fthrow:Reg.t -> loc:Loc.t -> t

val find : string -> functions -> t option

Look up a function of the given name in the given functions.

val fold_cfg : func -> 'a -> f:(block -> 'a -> 'a) -> 'a

Fold over the blocks of the control-flow graph of a function.

val is_undefined : t -> bool

Holds of functions that are declared but not defined.