Module Iter.Infix

val (--) : int -> int -> int t

a -- b is the range of integers from a to b, both included, in increasing order. It will therefore be empty if a > b.

val (--^) : int -> int -> int t

a --^ b is the range of integers from b to a, both included, in decreasing order (starts from a). It will therefore be empty if a < b.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

Monadic bind (infix version of flat_map

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

Infix version of map

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

Applicative operator (product+application)

val (<+>) : 'a t -> 'a t -> 'a t

Concatenation of iterators