Parameter Make_from_Comparer.Key

type t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

type ('a, 'compare_a) comparer := 'a -> 'a -> int

type compare

compare types are equipped with functions to support use of @@deriving compare, equal, sexp on types parameterized by such singleton types for compare functions. These derived functions are never actually called, since the compare type parameters are phantom.

val compare_compare : compare -> compare -> int

val equal_compare : compare -> compare -> bool

val sexp_of_compare : compare -> Sexplib0.Sexp.t

val compare_of_sexp : Sexplib0.Sexp.t -> compare

val comparer : (t, compare) comparer