Module Nullsafe

module AggregatedSummaries : sig ... end

Results of proc-level analysis, grouped by Java class they belong to. Useful for class-level analysis, when summaries for all methods are calculated and need to be combined together.

module AnnotatedField : sig ... end

Representation of a declared class field with nullsafe-specific data

module AnnotatedNullability : sig ... end

Nullability of a type in Java program (e.g. in a function or field declaration). It might come from explicit annotations (or lack of annotation), or from other sources, including conventions about defaults, models, or the mode nullsafe runs in. NOTE: This is complementary to InferredNullability.t. InferredNullability contains info about _actual_ nullability (what did nullsafe infer according to its flow-sensitive rules.). In contrast, AnnotatedNullability represents _formal_ type as it appears in the program code. NOTE: Nullsafe disregards user-provided annotations for local types, so annotated nullability applies only for types declared at methods and field level.

module AnnotatedSignature : sig ... end

Method signature with annotations.

module AnnotatedType : sig ... end

Formal type in program together with its nullability information.

module AnnotationGraph : sig ... end

Annotation graph for a Java class comprises of: * Nodes (annotation points): potential places in the class declaration to be annotated as @Nullable * For each annotation point the graph stores what will happen if this annotation point becomes real @Nullable (new violations that will arise) * Edges: if annotating a point A will require annotating another point B as @Nullable, A -> B are connected in the graph.

module AssignmentRule : sig ... end

Assignment rule should be checked when a value is assigned to a location. Assignment can be explicit (lhs = rhs) or implicit (e.g. returning from a function). This rule checks if null can be passed to a place that does not expect null.

module ClassLevelAnalysis : sig ... end

This stage is called for each Java class after all corresponding proc summaries are calculated.

module DereferenceRule : sig ... end

Dereference rule should be checked every type an object is dereferenced. The rule checks if the reference is nullable.

module Eradicate : sig ... end

The main entry point for Nullsafe typechecker.

module EradicateChecks : sig ... end

Module for the checks called by Eradicate.

module EradicateReporting : sig ... end

module ErrorRenderingUtils : sig ... end

Helper tools for nicer rendering nullsafe error.

module FileLevelAnalysis : sig ... end

module IDEnv : sig ... end

Environment for temporary identifiers used in instructions. Lazy implementation: only created when actually used.

module ImmutableChecker : sig ... end

module InferredNullability : sig ... end

Module to represent nullability of expressions inferred during flow-sensitive symbolic execution. NOTE: This is complementaty to InferredNullability.t. InferredNullability contains info about _formal_ nullability (what does the code say about nullability of a given type, according to explicit annotations and implicit agreements (e.g. models)). In contrast, InferredNullability represents what Nullsafe thinks about such and such expression according to its type inference rules.

module InheritanceRule : sig ... end

Inheritance rule:

module Initializers : sig ... end

Functionality for evaluating typestates after class initialization. Nullsafe's notion of initialization augments Java's one in the following way:

module ModelTables : sig ... end

module Models : sig ... end

Methods dealing with specific knowledge about code in important third libraries, standard libraries, etc

module Nullability : sig ... end

module NullsafeInit : sig ... end

module NullsafeIssue : sig ... end

Information about the nullsafe issue to be reported to the user / put into the result json

module NullsafeMode : sig ... end

Represents a type-checking mode of nullsafe.

module NullsafeSummary : sig ... end

Summary of per-procedure nullsafe result analysis

module OverAnnotatedRule : sig ... end

Checks if a type in signature (e.g. return value) can be made more specific. NOTE: This rule is complementatary to assignment rule. While assignment rule checks a single assignment `lhs = rhs`, this rule checks checks ALL assignments to `lhs` in the program. NOTE: Violation of this rule is not a type violation, hence it should never be surfaced as error: `lhs`'s type can be intentionally made broad by code author (e.g. to anticipate future changes in the implementation). Heuristits are required to correctly surface overannotated rule to the user. This rule is useful for some scenarios, especially for nullability code conversions when it is expected that some signatures were annotated with @Nullable defensively, so surfacing such cases can improve API and make migration smooth.

module ProvisionalAnnotation : sig ... end

Provisional annotation is an imaginary @Nullable annotation that is added to a class element. * (It implies the corresponding element is NOT annotated as @Nullable in the source code.) * The purpose of provisional annotation is to compute Nullability graph: what would happen if such and such * element was annotated as @Nullable.

module ProvisionalViolation : sig ... end

A simple wrapper over provisional violations for Rules

module ThirdPartyAnnotationGlobalRepo : sig ... end

A singleton. Should be initialized once prior to start of nullsafe.

module ThirdPartyAnnotationInfo : sig ... end

In-memory storage the information about nullability annotation of third-party methods.

module ThirdPartyAnnotationInfoLoader : sig ... end

module ThirdPartyMethod : sig ... end

A helper module responsible for representing nullability information for a single 3rd party method, as well with functionality to read this information from the 3rd party nullability repository.

module TypeCheck : sig ... end

Module type for the type checking functions.

module TypeErr : sig ... end

Module for Type Error messages.

module TypeOrigin : sig ... end

module TypeState : sig ... end

Module for typestates: maps from expressions to annotated types, with extensions.