Module Nullsafe.AggregatedSummaries

Results of proc-level analysis, grouped by Java class they belong to. Useful for class-level analysis, when summaries for all methods are calculated and need to be combined together.

module ClassInfo : sig ... end

Aggregated information for each user defined (not anonymous) Java class

val aggregate : (IR.JavaClassName.t * NullsafeSummary.t) list -> ClassInfo.t list

Given a list of all summaries and their classes, group them by names and aggregate in a list of top-level classes.