Module Nullsafe.AnnotationGraph

Annotation graph for a Java class comprises of: * Nodes (annotation points): potential places in the class declaration to be annotated as @Nullable * For each annotation point the graph stores what will happen if this annotation point becomes real @Nullable (new violations that will arise) * Edges: if annotating a point A will require annotating another point B as @Nullable, A -> B are connected in the graph.

val build_graph : IR.Tenv.t -> IR.Struct.t -> IR.Typ.name -> ProvisionalViolation.t list -> ATDGenerated.Jsonbug_t.annotation_point list

Given a Java class and the list of all provisional violations found in that class, build the annotation graph