Module Nullsafe.AssignmentRule

Assignment rule should be checked when a value is assigned to a location. Assignment can be explicit (lhs = rhs) or implicit (e.g. returning from a function). This rule checks if null can be passed to a place that does not expect null.

type violation

val compare_violation : violation -> violation -> int

val check : lhs:AnnotatedNullability.t -> rhs:InferredNullability.t -> (unit, violation) IStdlib.IStd.result

If `null` can leak from a "less strict" type to "more strict" type, this is an Assignment Rule violation.

module ProvisionalViolation : sig ... end

Violation that will occur if the provisional annotation becomes real @Nullable

module ReportableViolation : sig ... end

Violation that needs to be reported to the user.