Module AssignmentRule.ReportableViolation

Violation that needs to be reported to the user.

type t

val from : NullsafeMode.t -> violation -> t option

Depending on the mode, violation might or might not be important enough to be reported to the user. If it should NOT be reported for that mode, this function will return None.

type assignment_type =

| PassingParamToFunction of function_info

| AssigningToField of IR.Fieldname.t

| ReturningFromFunction of IR.Procname.Java.t

and function_info = {

param_signature : AnnotatedSignature.param_signature;

actual_param_expression : string;

param_index : int;

annotated_signature : AnnotatedSignature.t;

procname : IR.Procname.Java.t;

}

val compare_assignment_type : assignment_type -> assignment_type -> int

val compare_function_info : function_info -> function_info -> int

val make_nullsafe_issue : assignment_location:IBase.Location.t -> assignment_type -> t -> NullsafeIssue.t

Given context around violation, return error message together with the info where to put this message