Module Nullsafe.DereferenceRule

Dereference rule should be checked every type an object is dereferenced. The rule checks if the reference is nullable.

type violation

val compare_violation : violation -> violation -> int

val check : InferredNullability.t -> ( unit, violation ) IStdlib.IStd.result

violation of Dereference rule reflects possibility of dereferencing of `null`. Note that this might or might not be severe enough to be reported to the user, depending on the mode agreements.

module ProvisionalViolation : sig ... end

Violation that will occur if the provisional annotation becomes real @Nullable

module ReportableViolation : sig ... end

Violation that needs to be reported to the user.