Module DereferenceRule.ReportableViolation

Violation that needs to be reported to the user.

type t

type dereference_type =

| MethodCall of IR.Procname.Java.t

| AccessToField of IR.Fieldname.t

| AccessByIndex of {

index_desc : string;

}

| ArrayLengthAccess

val compare_dereference_type : dereference_type -> dereference_type -> int

val from : NullsafeMode.t -> violation -> t option

Depending on the mode, violation might or might not be important enough to be reported to the user. If it should NOT be reported for that mode, this function will return None.

val make_nullsafe_issue : t -> dereference_location:IBase.Location.t -> dereference_type -> nullable_object_descr:string option -> NullsafeIssue.t

Given context around violation, return error message together with the info where to put this message