Module Nullsafe.InferredNullability

Module to represent nullability of expressions inferred during flow-sensitive symbolic execution. NOTE: This is complementaty to InferredNullability.t. InferredNullability contains info about _formal_ nullability (what does the code say about nullability of a given type, according to explicit annotations and implicit agreements (e.g. models)). In contrast, InferredNullability represents what Nullsafe thinks about such and such expression according to its type inference rules.

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val get_nullability : t -> Nullability.t

val create : TypeOrigin.t -> t

val is_nonnullish : t -> bool

Check whether corresponding Nullability is Nullability.is_nonnullish

val get_simple_origin : t -> TypeOrigin.t

The simple explanation of how was nullability inferred.

val get_provisional_annotations : t -> ProvisionalAnnotation.t list

val join : t -> t -> t

This is what happens with nullability when we join two flows in CFG, e.g.

if(something) {
  a = e1;
} else {
  a = e2;
}
// what is nullability of `a` at this point?

val origin_is_fun_defined : t -> bool

val pp : Stdlib.Format.formatter -> t -> unit