Module Nullsafe.ProvisionalAnnotation

Provisional annotation is an imaginary @Nullable annotation that is added to a class element. * (It implies the corresponding element is NOT annotated as @Nullable in the source code.) * The purpose of provisional annotation is to compute Nullability graph: what would happen if such and such * element was annotated as @Nullable.

type t =

| Field of {

field_name : IR.Fieldname.t;

}

| Method of IR.Procname.Java.t

| Param of {

method_info : IR.Procname.Java.t;

num : int;

}

val compare : t -> t -> int

val pp : Stdlib.Format.formatter -> t -> unit