Module Nullsafe__AggregatedSummaries

module ClassInfo : sig ... end

Aggregated information for each user defined (not anonymous) Java class

val aggregate : (IR.JavaClassName.t * Nullsafe.NullsafeSummary.t) list -> ClassInfo.t list

Given a list of all summaries and their classes, group them by names and aggregate in a list of top-level classes.