Module Nullsafe__AggregatedSummaries.ClassInfo

Aggregated information for each user defined (not anonymous) Java class

type t

val get_class_name : t -> IR.JavaClassName.t

val get_summaries : t -> Nullsafe.NullsafeSummary.t list

val get_nested_classes_info : t -> t list

val get_nested_anonymous_summaries : t -> Nullsafe.NullsafeSummary.t list IR.JavaClassName.Map.t

List of all anonymous class summaries belonging to this class, together with name of anonymous class. This is a flattenned list, so we don't care if one anonymous class is nested inside the other.

val get_recursive_summaries : t -> (IR.JavaClassName.t * Nullsafe.NullsafeSummary.t) list

A flattened list of all summaries, user-level, nested, and anonymous, combined together

val pp : Stdlib.Format.formatter -> t -> unit