Module Nullsafe__AnnotationGraph

val build_graph : IR.Tenv.t -> IR.Struct.t -> IR.Typ.name -> Nullsafe.ProvisionalViolation.t list -> ATDGenerated.Jsonbug_t.annotation_point list

Given a Java class and the list of all provisional violations found in that class, build the annotation graph