Module Nullsafe__ClassLevelAnalysis

val analyze_class : IR.Tenv.t -> IBase.SourceFile.t -> Nullsafe.AggregatedSummaries.ClassInfo.t -> Absint.IssueLog.t -> Absint.IssueLog.t

Given aggregated summary for a class, analyze it, and return updated issue log, if necessary. This function will be called for each non-trivial1 anonymous class in the file, including nested classes. Order of calls is not specified.

1The class is non-trivial if it has at least one procedure, or contains at least one nested non-trivial class.

(Note that IssueLog.t is a mutable type so it can be actually mutated by this function: returning IssueLog.t is done for convenient chaining.)