Module Nullsafe__ThirdPartyAnnotationGlobalRepo

val initialize : absolute_path_to_repo:string option -> Nullsafe.ThirdPartyAnnotationInfo.storage -> unit

Should be initialized exactly once before access. Path to repo is None if there is no repo on the disk.

val get_repo : unit -> Nullsafe.ThirdPartyAnnotationInfo.storage

Can be accessed only when initialization was done

val get_user_friendly_third_party_sig_file_name : filename:string -> string

Given filename containing the repo signatures, render it in a way convenient for the user