Module Pulselib.PulseAbductiveDomain

module F = Stdlib.Format

module BaseDomain = PulseBaseDomain

module BaseMemory = PulseBaseMemory

module BaseStack = PulseBaseStack

module Decompiler = PulseDecompiler

module DecompilerExpr = PulseDecompilerExpr

module PathContext = PulsePathContext

Layer on top of BaseDomain to propagate operations on the current state to the pre-condition when necessary

The abstract type t is a pre/post pair in the style of biabduction.

module type BaseDomainSig = sig ... end

signature common to the "normal" Domain, representing the post at the current program point, and the inverted PreDomain, representing the inferred pre-condition

module PostDomain : BaseDomainSig

The post abstract state at each program point, or current state.

module PreDomain : BaseDomainSig

The inferred pre-condition at each program point, biabduction style.

type t = private {

post : PostDomain.t;	(*
state at the current program point
*)

pre : PreDomain.t;	(*
inferred procedure pre-condition leading to the current program point
*)

path_condition : PulseBasicInterface.Formula.t;	(*
arithmetic facts true along the path (holding for both pre and post since abstract values are immutable)
*)

decompiler : Decompiler.t;

topl : PulseTopl.state;	(*
state at of the Topl monitor at the current program point, when Topl is enabled
*)

need_specialization : bool;	(*
a call that could be resolved via analysis-time specialization has been skipped
*)

skipped_calls : PulseBasicInterface.SkippedCalls.t;	(*
metadata: procedure calls for which no summary was found
*)

}

pre/post on a single program path

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val leq : lhs:t -> rhs:t -> bool

val pp : Stdlib.Format.formatter -> t -> unit

val mk_initial : IR.Tenv.t -> IR.Procname.t -> IR.ProcAttributes.t -> t

module Stack : sig ... end

stack operations like BaseStack but that also take care of propagating facts to the precondition

module Memory : sig ... end

memory operations like BaseMemory but that also take care of propagating facts to the precondition

module AddressAttributes : sig ... end

attribute operations like PulseBaseAddressAttributes but that also take care of propagating facts to the precondition

val should_havoc_if_unknown : unit -> [> `ShouldHavoc | `ShouldOnlyHavocResources ]

val apply_unknown_effect : ?havoc_filter: ( PulseBasicInterface.AbstractValue.t -> BaseMemory.Access.t -> BaseMemory.AddrTrace.t -> bool ) -> PulseBasicInterface.ValueHistory.t -> PulseBasicInterface.AbstractValue.t -> t -> t

do as much as possible to assume "the best" happened at that location: deallocate and initialize everything reachable from the address, then havoc all the edges starting from the address passing havoc_filter (by default everything passes the filter)

val is_local : IR.Var.t -> t -> bool

val find_post_cell_opt : PulseBasicInterface.AbstractValue.t -> t -> BaseDomain.cell option

val get_unreachable_attributes : t -> PulseBasicInterface.AbstractValue.t list

collect the addresses that have attributes but are unreachable in the current post-condition

val add_skipped_call : IR.Procname.t -> PulseBasicInterface.Trace.t -> t -> t

val add_skipped_calls : PulseBasicInterface.SkippedCalls.t -> t -> t

val set_path_condition : PulseBasicInterface.Formula.t -> t -> t

val set_need_specialization : t -> t

val unset_need_specialization : t -> t

val map_decompiler : t -> f:( Decompiler.t -> Decompiler.t ) -> t

val set_post_edges : PulseBasicInterface.AbstractValue.t -> BaseMemory.Edges.t -> t -> t

directly set the edges for the given address, bypassing abduction altogether

val set_post_cell : PathContext.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> BaseDomain.cell -> IBase.Location.t -> t -> t

directly set the edges and attributes for the given address, bypassing abduction altogether

val remove_from_post : PulseBasicInterface.AbstractValue.t -> t -> t

remove any association from the given address, bypassing abduction altogether

val incorporate_new_eqs : PulseBasicInterface.Formula.new_eqs -> t -> ( t, [> `PotentialInvalidAccess of t * PulseBasicInterface.AbstractValue.t * (PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option) ] ) IStdlib.IStd.result PulseBasicInterface.SatUnsat.t

Check that the new equalities discovered are compatible with the current pre and post heaps, e.g. x = 0 is not compatible with x being allocated, and x = y is not compatible with x and y being allocated separately. In those cases, the result is Unsat.

val incorporate_new_eqs_on_val : PulseBasicInterface.Formula.new_eqs -> PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.AbstractValue.t

Similar to incorporate_new_eqs, but apply to an abstract value.

val initialize : PulseBasicInterface.AbstractValue.t -> t -> t

Remove "Uninitialized" attribute of the given address

val set_uninitialized : IR.Tenv.t -> PathContext.t -> [ `LocalDecl of IR.Pvar.t * PulseBasicInterface.AbstractValue.t option | `Malloc of PulseBasicInterface.AbstractValue.t ] -> IR.Typ.t -> IBase.Location.t -> t -> t

Add "Uninitialized" attributes when a variable is declared or a memory is allocated by malloc.

module Summary : sig ... end

module Topl : sig ... end