Module PulseAbductiveDomain.AddressAttributes

Safe version of PulseBaseAddressAttributes

val abduce_attribute : Pulselib.PulseBasicInterface.AbstractValue.t -> Pulselib.PulseBasicInterface.Attribute.t -> t -> t

add the attribute to the pre, if the address is in pre

val abduce_and_add : Pulselib.PulseBasicInterface.AbstractValue.t -> Pulselib.PulseBasicInterface.Attributes.t -> t -> t

add the attributes to both the current state and, if meaningful, the pre

val add_one : Pulselib.PulseBasicInterface.AbstractValue.t -> Pulselib.PulseBasicInterface.Attribute.t -> t -> t

add the attribute only to the post

val add_attrs : Pulselib.PulseBasicInterface.AbstractValue.t -> Pulselib.PulseBasicInterface.Attributes.t -> t -> t

val find_opt : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.Attributes.t option

val check_valid : PathContext.t -> ?must_be_valid_reason: Pulselib.PulseBasicInterface.Invalidation.must_be_valid_reason -> Pulselib.PulseBasicInterface.Trace.t -> Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (t, Pulselib.PulseBasicInterface.Invalidation.t * Pulselib.PulseBasicInterface.Trace.t) IStdlib.IStd.result

val check_initialized : PathContext.t -> Pulselib.PulseBasicInterface.Trace.t -> Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (t, unit) IStdlib.IStd.result

val add_taint_sink : PathContext.t -> Pulselib.PulseBasicInterface.TaintItem.t -> Pulselib.PulseBasicInterface.Trace.t -> Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val invalidate : (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) -> Pulselib.PulseBasicInterface.Invalidation.t -> IBase.Location.t -> t -> t

val initialize : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

Remove "Uninitialized" attribute of the given address

val set_uninitialized : IR.Tenv.t -> PathContext.t -> [ `LocalDecl of IR.Pvar.t * Pulselib.PulseBasicInterface.AbstractValue.t option | `Malloc of Pulselib.PulseBasicInterface.AbstractValue.t ] -> IR.Typ.t -> IBase.Location.t -> t -> t

Add "Uninitialized" attributes when a variable is declared or a memory is allocated by malloc.

val always_reachable : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val allocate : Pulselib.PulseBasicInterface.Attribute.allocator -> Pulselib.PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val java_resource_release : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val hack_async_await : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val is_java_resource_released : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val csharp_resource_release : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val is_csharp_resource_released : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val add_dynamic_type : IR.Typ.t -> Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val add_dynamic_type_source_file : IR.Typ.t -> IBase.SourceFile.t -> Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val add_ref_counted : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val is_ref_counted : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val remove_allocation_attr : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val remove_taint_attrs : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val get_dynamic_type : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> IR.Typ.t option

val get_dynamic_type_source_file : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (IR.Typ.t * IBase.SourceFile.t option) option

val get_static_type : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> IR.Typ.Name.t option

val get_allocation : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (Pulselib.PulseBasicInterface.Attribute.allocator * Pulselib.PulseBasicInterface.Trace.t) option

val get_closure_proc_name : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> IR.Procname.t option

val get_copied_into : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.Attribute.CopiedInto.t option

val get_copied_return : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (Pulselib.PulseBasicInterface.AbstractValue.t * bool * Pulselib.PulseBasicInterface.Attribute.CopyOrigin.t * IBase.Location.t) option

val remove_copied_return : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val get_source_origin_of_copy : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.AbstractValue.t option

val get_taint_sources_and_sanitizers : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.Attribute.TaintedSet.t * Pulselib.PulseBasicInterface.Attribute.TaintSanitizedSet.t

val get_propagate_taint_from : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.Attribute.taint_in list option

val is_end_of_collection : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val mark_as_end_of_collection : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val is_std_vector_reserved : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val std_vector_reserve : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> t

val add_unreachable_at : Pulselib.PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val add_copied_return : Pulselib.PulseBasicInterface.AbstractValue.t -> source:Pulselib.PulseBasicInterface.AbstractValue.t -> is_const_ref:bool -> Pulselib.PulseBasicInterface.Attribute.CopyOrigin.t -> IBase.Location.t -> t -> t

val get_config_usage : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.Attribute.ConfigUsage.t option

val get_const_string : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> string option

val get_valid_returned_from_unknown : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> Pulselib.PulseBasicInterface.AbstractValue.t list option

val get_written_to : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (Pulselib.PulseBasicInterface.Timestamp.t * Pulselib.PulseBasicInterface.Trace.t) option

val is_copied_from_const_ref : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val is_std_moved : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> bool

val get_address_of_stack_variable : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> (IR.Var.t * IBase.Location.t * Pulselib.PulseBasicInterface.ValueHistory.t) option