Module PulseAbductiveDomain.Memory

Safe version of PulseBaseMemory

val add_edge : PathContext.t -> (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) -> PulseAccess.t -> (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) -> IBase.Location.t -> t -> t

val eval_edge : (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) -> PulseAccess.t -> t -> t * (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t)

eval_edge (addr,hist) access astate follows the edge addr --access--> . in memory and returns what it points to or creates a fresh value if that edge didn't exist.

val fold_edges : Pulselib.PulseBasicInterface.AbstractValue.t -> (t, PulseAccess.t * (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t), _) IStdlib.IStd.Container.fold

val find_edge_opt : Pulselib.PulseBasicInterface.AbstractValue.t -> PulseAccess.t -> t -> (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) option

val exists_edge : Pulselib.PulseBasicInterface.AbstractValue.t -> t -> f: ((PulseAccess.t * (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t)) -> bool) -> bool