Module PulseAbductiveDomain.PreDomain

The inferred pre-condition at each program point, biabduction style.

NOTE: PreDomain and Domain theoretically differ in that PreDomain should be the inverted lattice of Domain, but since we never actually join states or check implication the two collapse into one. *

type t = private BaseDomain.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t