Module PulseAbstractValue.Constants

val get_int : IR.IntLit.t -> t

Get or create an abstract value associated with a constant IntLit.t. The idea is that clients will record in the abstract state that the returned t is equal to the given integer. If the same integer is queried later on then this module will return the same abstract variable.