Module Pulselib.PulseAccessResult

module AbductiveDomain = PulseAbductiveDomain

module DecompilerExpr = PulseDecompilerExpr

module Diagnostic = PulseDiagnostic

type error =

| PotentialInvalidAccess of {

astate : AbductiveDomain.t;

address : DecompilerExpr.t;

must_be_valid : PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option;

}

| ReportableError of {

astate : AbductiveDomain.t;

diagnostic : Diagnostic.t;

}

| ISLError of {

astate : AbductiveDomain.t;

}

| WithSummary of error * AbductiveDomain.Summary.t	(*
An error for which the summary corresponding to the abstract state has already been computed, to avoid having to compute the summary for a given abstract state multiple times. For convenience this is part of the error datatype but the invariant that an error within a WithSummary (error, summary) cannot itself be a WithSummary _.
*)

type 'a t = ( 'a, error ) PulseResult.t

val with_summary : ( 'a, error * AbductiveDomain.Summary.t ) PulseResult.t -> 'a t

type abductive_error = [

| `ISLError of AbductiveDomain.t

| `PotentialInvalidAccess of AbductiveDomain.t * PulseBasicInterface.AbstractValue.t * (PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option)

]

Intermediate datatype since AbductiveDomain cannot refer to this module without creating a circular dependency.

Purposefully omits `MemoryLeak errors as it's a good idea to double-check if you really want to report a leak.

type abductive_summary_error = [

| `PotentialInvalidAccessSummary of AbductiveDomain.Summary.t * AbductiveDomain.t * DecompilerExpr.t * (PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option)

]

val of_result_f : ( 'a, error ) IStdlib.IStd.result -> f:( error -> 'a ) -> 'a t

val of_result : ( AbductiveDomain.t, error ) IStdlib.IStd.result -> AbductiveDomain.t t

val of_error_f : error -> f:( error -> 'a ) -> 'a t

val of_abductive_summary_error : [< abductive_summary_error ] -> error * AbductiveDomain.Summary.t

val of_abductive_result : ( 'a, [< abductive_error ] ) IStdlib.IStd.result -> 'a t

val of_abductive_summary_result : ( 'a, [< abductive_summary_error ] ) IStdlib.IStd.result -> ( 'a, error * AbductiveDomain.Summary.t ) PulseResult.t

val of_abductive_access_result : PulseBasicInterface.Trace.t -> ( 'a, [< `InvalidAccess of PulseBasicInterface.AbstractValue.t * PulseBasicInterface.Invalidation.t * PulseBasicInterface.Trace.t * AbductiveDomain.t | abductive_error ] ) IStdlib.IStd.result -> 'a t

val ignore_leaks : ( AbductiveDomain.Summary.t, [< `MemoryLeak of AbductiveDomain.Summary.t * AbductiveDomain.t * PulseBasicInterface.Attribute.allocator * PulseBasicInterface.Trace.t * IBase.Location.t | `ResourceLeak of AbductiveDomain.Summary.t * AbductiveDomain.t * IR.JavaClassName.t * PulseBasicInterface.Trace.t * IBase.Location.t | `RetainCycle of AbductiveDomain.Summary.t * AbductiveDomain.t * PulseBasicInterface.Trace.t list * DecompilerExpr.t * DecompilerExpr.t * IBase.Location.t | abductive_summary_error ] ) IStdlib.IStd.result -> ( AbductiveDomain.Summary.t, [> abductive_summary_error ] ) IStdlib.IStd.result