Module Pulselib.PulseAttribute

module F = Stdlib.Format

module AbstractValue = PulseAbstractValue

module CallEvent = PulseCallEvent

module ConfigName = OpenSource.FbPulseConfigName

module DecompilerExpr = PulseDecompilerExpr

module Invalidation = PulseInvalidation

module Taint = PulseTaint

module Timestamp = PulseTimestamp

module Trace = PulseTrace

module ValueHistory = PulseValueHistory

type allocator =

| CMalloc

| CustomMalloc of IR.Procname.t

| CRealloc

| CustomRealloc of IR.Procname.t

| CppNew

| CppNewArray

| JavaResource of IR.JavaClassName.t

| CSharpResource of IR.CSharpClassName.t

| ObjCAlloc

val equal_allocator : allocator -> allocator -> bool

val pp_allocator : F.formatter -> allocator -> unit

type taint_in = {

v : AbstractValue.t;

}

val compare_taint_in : taint_in -> taint_in -> int

val equal_taint_in : taint_in -> taint_in -> bool

module Tainted : sig ... end

module TaintedSet : IStdlib.PrettyPrintable.PPSet with type elt = Tainted.t

module TaintSink : sig ... end

module TaintSinkSet : IStdlib.PrettyPrintable.PPSet with type elt = TaintSink.t

module TaintSanitized : sig ... end

module TaintSanitizedSet : IStdlib.PrettyPrintable.PPSet with type elt = TaintSanitized.t

module CopyOrigin : sig ... end

module CopiedInto : sig ... end

module ConfigUsage : sig ... end

type t =

| AddressOfCppTemporary of IR.Var.t * ValueHistory.t

| AddressOfStackVariable of IR.Var.t * IBase.Location.t * ValueHistory.t

| Allocated of allocator * Trace.t

| AlwaysReachable

| Closure of IR.Procname.t

| ConfigUsage of ConfigUsage.t

| ConstString of string

| CopiedInto of CopiedInto.t	(*
records the copied var/field for each source address
*)

| CopiedReturn of {

source : AbstractValue.t;

is_const_ref : bool;

from : CopyOrigin.t;

copied_location : IBase.Location.t;

}	(*
records the copied value for the return address
*)

| DynamicType of IR.Typ.t * IBase.SourceFile.t option

| EndOfCollection

| Invalid of Invalidation.t * Trace.t

| MustBeInitialized of Timestamp.t * Trace.t

| MustBeValid of Timestamp.t * Trace.t * Invalidation.must_be_valid_reason option

| MustNotBeTainted of TaintSinkSet.t

| JavaResourceReleased

| CSharpResourceReleased

| PropagateTaintFrom of taint_in list

| RefCounted

| ReturnedFromUnknown of AbstractValue.t list

| SourceOriginOfCopy of {

source : PulseAbstractValue.t;

is_const_ref : bool;

}	(*
records the source value for a given copy to lookup the appropriate heap in non-disj domain
*)

| StdMoved

| StdVectorReserve

| Tainted of TaintedSet.t

| TaintSanitized of TaintSanitizedSet.t

| Uninitialized

| UnknownEffect of CallEvent.t * ValueHistory.t	(*
generated by calls to unknown functions to remember that a pointer has been passed to an unknown function and so everything reachable from it has potentially been affected in unknown ways
*)

| UnreachableAt of IBase.Location.t	(*
temporary marker to remember where a variable became unreachable; helps with accurately reporting leaks
*)

| UsedAsBranchCond of IR.Procname.t * IBase.Location.t * Trace.t

| WrittenTo of Timestamp.t * Trace.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val pp : F.formatter -> t -> unit

val is_suitable_for_pre : t -> bool

val is_suitable_for_post : t -> bool

val filter_unreachable : AbstractValue.Set.t AbstractValue.Map.t -> ( AbstractValue.t -> bool ) -> t -> t option

update an attribute to get rid of abstract values that do not satisfy the given predicate; the result is None if the attribute becomes meaningless as a result

module Attributes : sig ... end