Module PulseBaseDomain.GraphVisit

val fold : var_filter:(IR.Var.t -> bool) -> t -> init:'accum -> f: (IR.Var.t -> 'accum -> Pulselib.PulseBasicInterface.AbstractValue.t -> Memory.Access.t list -> ('accum, 'final) Base.Continue_or_stop.t) -> finish:('accum -> 'final) -> Pulselib.PulseBasicInterface.AbstractValue.Set.t * 'final

Generic graph traversal of the memory starting from each variable in the stack that pass var_filter, in order. Returns the result of folding over every address in the graph and the set of addresses that have been visited before f returned Stop or all reachable addresses were seen. f is passed each address together with the variable from which the address was reached and the access path from that variable to the address.

val fold_from_addresses : Pulselib.PulseBasicInterface.AbstractValue.t Stdlib.Seq.t -> t -> init:'accum -> already_visited:Pulselib.PulseBasicInterface.AbstractValue.Set.t -> f: ('accum -> Pulselib.PulseBasicInterface.AbstractValue.t -> Memory.Access.t list -> ('accum, 'final) Base.Continue_or_stop.t) -> finish:('accum -> 'final) -> Pulselib.PulseBasicInterface.AbstractValue.Set.t * 'final

Similar to fold, but start from given addresses, instead of stack variables. Use already_visited as initial set of visited values.