Module Pulselib.PulseBasicInterface

Basic Pulse modules that are safe to use in any module

module AbstractValue = PulseAbstractValue

module Attribute = PulseAttribute

module Attributes = PulseAttribute.Attributes

module CallEvent = PulseCallEvent

module Invalidation = PulseInvalidation

module PathCondition = PulsePathCondition

module SatUnsat = PulseSatUnsat

module SkippedCalls = PulseSkippedCalls

module Taint = PulseTaint

module Timestamp = PulseTimestamp

module Trace = PulseTrace

module ValueHistory = PulseValueHistory

include module type of struct include SatUnsat.Types end

for opening to get Sat and Unsat in the namespace

type nonrec 'a sat_unsat_t = 'a SatUnsat.t =

| Unsat

| Sat of 'a

Enforce short form usage

module PulseAbstractValue = PulseAbstractValue

module PulseAttribute = PulseAttribute

module PulseCallEvent = PulseCallEvent

module PulseInvalidation = PulseInvalidation

module PulsePathCondition = PulsePathCondition

module PulseTimestamp = PulseTimestamp

module PulseSkippedCalls = PulseSkippedCalls

module PulseTaint = PulseTaint

module PulseTrace = PulseTrace

module PulseValueHistory = PulseValueHistory

module PulseSatUnsat = PulseSatUnsat