Module Pulselib.PulseDiagnostic

module F = Stdlib.Format

module Attribute = PulseAttribute

module CallEvent = PulseCallEvent

module DecompilerExpr = PulseDecompilerExpr

module Invalidation = PulseInvalidation

module Taint = PulseTaint

module Trace = PulseTrace

module ValueHistory = PulseValueHistory

type calling_context = (CallEvent.t * IBase.Location.t) list

val compare_calling_context : calling_context -> calling_context -> int

val equal_calling_context : calling_context -> calling_context -> bool

type access_to_invalid_address = {

calling_context : calling_context;	(*
the list of function calls leading to the issue being realised, in outermost-to-innermost order, which is an additional common prefix to the traces in the record
*)

invalid_address : DecompilerExpr.t;

invalidation : Invalidation.t;

invalidation_trace : Trace.t;	(*
assuming we are in the calling context, the trace leads to invalidation without further assumptions
*)

access_trace : Trace.t;	(*
assuming we are in the calling context, the trace leads to an access to the value invalidated in invalidation_trace without further assumptions
*)

must_be_valid_reason : Invalidation.must_be_valid_reason option;

}

val compare_access_to_invalid_address : access_to_invalid_address -> access_to_invalid_address -> int

val equal_access_to_invalid_address : access_to_invalid_address -> access_to_invalid_address -> bool

val yojson_of_access_to_invalid_address : access_to_invalid_address -> Ppx_yojson_conv_lib.Yojson.Safe.t

module ErlangError : sig ... end

type read_uninitialized_value = {

calling_context : calling_context;	(*
the list of function calls leading to the issue being realised, which is an additional common prefix to the traces in the record
*)

trace : Trace.t;	(*
assuming we are in the calling context, the trace leads to read of the uninitialized value
*)

}

val compare_read_uninitialized_value : read_uninitialized_value -> read_uninitialized_value -> int

val equal_read_uninitialized_value : read_uninitialized_value -> read_uninitialized_value -> bool

val yojson_of_read_uninitialized_value : read_uninitialized_value -> Ppx_yojson_conv_lib.Yojson.Safe.t

type flow_kind =

| TaintedFlow

| FlowToSink

| FlowFromSource

val equal_flow_kind : flow_kind -> flow_kind -> bool

type t =

| AccessToInvalidAddress of access_to_invalid_address

| ConstRefableParameter of {

param : IR.Var.t;

typ : IR.Typ.t;

location : IBase.Location.t;

}

| MemoryLeak of {

allocator : Attribute.allocator;

allocation_trace : Trace.t;

location : IBase.Location.t;

}

| RetainCycle of {

assignment_traces : Trace.t list;

value : DecompilerExpr.t;

path : DecompilerExpr.t;

location : IBase.Location.t;

}

| ErlangError of ErlangError.t

| ReadUninitializedValue of read_uninitialized_value

| ResourceLeak of {

class_name : IR.JavaClassName.t;

allocation_trace : Trace.t;

location : IBase.Location.t;

}

| StackVariableAddressEscape of {

variable : IR.Var.t;

history : ValueHistory.t;

location : IBase.Location.t;

}

| TaintFlow of {

expr : DecompilerExpr.t;

source : Taint.t * ValueHistory.t;

sink : Taint.t * Trace.t;

location : IBase.Location.t;

flow_kind : flow_kind;

}

| UnnecessaryCopy of {

copied_into : PulseAttribute.CopiedInto.t;

typ : IR.Typ.t;

location : IBase.Location.t;

copied_location : (IR.Procname.t * IBase.Location.t) option;

from : PulseAttribute.CopyOrigin.t;

}

an error to report to the user

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val pp : F.formatter -> t -> unit

val aborts_execution : t -> bool

whether the presence of an error should abort the execution

val get_message : t -> string

val get_location : t -> IBase.Location.t

val get_copy_type : t -> IR.Typ.t option

val get_issue_type : latent:bool -> t -> IBase.IssueType.t

val get_trace : t -> Absint.Errlog.loc_trace