Module PulseDiagnostic.ReadUninitialized

type t = {

typ : Attribute.UninitializedTyp.t;

calling_context : calling_context;
(*
the list of function calls leading to the issue being realised, which is an additional common prefix to the traces in the record
*)

trace : Trace.t;
(*
assuming we are in the calling context, the trace leads to read of the uninitialized value
*)

}

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t