Module Pulselib.PulseExecutionDomain

module AbductiveDomain = PulseAbductiveDomain

module LatentIssue = PulseLatentIssue

type 'abductive_domain_t base_t =

| ContinueProgram of 'abductive_domain_t	(*
represents the state at the program point
*)

| ExceptionRaised of 'abductive_domain_t	(*
state after an exception has been thrown
*)

| ExitProgram of AbductiveDomain.summary	(*
represents the state originating at exit/divergence.
*)

| AbortProgram of AbductiveDomain.summary	(*
represents the state at the program point that caused an error
*)

| LatentAbortProgram of {

astate : AbductiveDomain.summary;

latent_issue : LatentIssue.t;

}	(*
this path leads to an error but we don't have conclusive enough data to report it yet
*)

| LatentInvalidAccess of {

astate : AbductiveDomain.summary;

address : PulseBasicInterface.AbstractValue.t;

must_be_valid : PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option;

calling_context : (PulseBasicInterface.CallEvent.t * IBase.Location.t) list;

}	(*
if address is ever observed to be invalid then there is an invalid access because it must_be_valid
*)

| ISLLatentMemoryError of AbductiveDomain.summary	(*
represents the state at the program point that might cause an error; used for Config.pulse_isl
*)

type t = AbductiveDomain.t base_t

include Absint.AbstractDomain.Disjunct with type t := t

include Absint.AbstractDomain.Comparable with type t := t

include IStdlib.PrettyPrintable.PrintableType with type t := t

val pp : IStdlib.PrettyPrintable.F.formatter -> t -> unit

val leq : lhs:t -> rhs:t -> bool

the implication relation: lhs <= rhs means lhs |- rhs

val equal_fast : t -> t -> bool

equal_fast x y must imply x <=> y; it's a good idea for this function to be "fast", e.g. not depend on the size of its input

val is_normal : t -> bool

test if the abstract state represents exactly concrete states

val is_exceptional : t -> bool

test if the abstract state represents exactly exceptional concrete states

val exceptional_to_normal : t -> t

convert all exceptional states into normal states (used when reaching a handler)

val continue : AbductiveDomain.t -> t

val is_unsat_cheap : t -> bool

see PulsePathCondition.is_unsat_cheap

type summary = AbductiveDomain.summary base_t

val compare_summary : summary -> summary -> int

val equal_summary : summary -> summary -> bool

val yojson_of_summary : summary -> Ppx_yojson_conv_lib.Yojson.Safe.t