Module PulseFormulaPhi.TermDomainOrRange

type domain_or_range =

| Domain

| Range

| DomainAndRange

for a binding t=v, describes whether a variable v' appears in t (domain), is v (range), or both; we could have a different representation of term occurrences using a map t -> Domain | Range | DomainAndRange to match what we are trying to do closer but we use a set like other occurrence maps, oh well.

val compare_domain_or_range : domain_or_range -> domain_or_range -> int

type t = Term.t * domain_or_range

val compare : t -> t -> int

module Set : sig ... end

val pp_set : (Term.F.formatter -> Term.LinArith.Var.t -> unit) -> F.formatter -> Set.t -> unit