Module Pulselib.PulseFormulaTerm

module F = Stdlib.Format

module L = IBase.Logging

module CItv = PulseCItv

module SatUnsat = PulseSatUnsat

module Debug = PulseFormulaDebug

module Var = PulseFormulaVar

module Q = QSafeCapped

module Z = ZSafe

module LinArith = PulseFormulaLinArit

type function_symbol =

| Unknown of Var.t

| Procname of IR.Procname.t

val compare_function_symbol : function_symbol -> function_symbol -> int

val equal_function_symbol : function_symbol -> function_symbol -> bool

val yojson_of_function_symbol : function_symbol -> Ppx_yojson_conv_lib.Yojson.Safe.t

val pp_function_symbol : Var.F.formatter -> function_symbol -> unit

type operand =

| AbstractValueOperand of Var.t

| ConstOperand of IR.Const.t

| FunctionApplicationOperand of {

f : function_symbol;

actuals : Var.t list;

}

val compare_operand : operand -> operand -> int

val equal_operand : operand -> operand -> bool

val pp_operand : Var.F.formatter -> operand -> unit

type 'term_t subst_target = 'term_t LinArith.subst_target =

| QSubst of Q.t

| ConstantSubst of 'term_t * Var.t option

| VarSubst of Var.t

| LinSubst of LinArith.t

| NonLinearTermSubst of 'term_t

val subst_f : Var.Map.key Var.Map.t -> Var.Map.key -> Var.Map.key

val targetted_subst_var : Var.Map.key Var.Map.t -> Var.Map.key -> 'a subst_target

Expressive term structure to be able to express all of SIL, but the main smarts of the formulas are for the equality between variables and linear arithmetic subsets. Terms (and atoms, below) are kept as a last-resort for when outside that fragment.

type t =

| Const of Q.t

| String of string

| Var of Var.t

| Procname of IR.Procname.t

| FunctionApplication of {

f : t;

actuals : t list;

}

| Linear of LinArith.t

| Add of t * t

| Minus of t

| LessThan of t * t

| LessEqual of t * t

| Equal of t * t

| NotEqual of t * t

| Mult of t * t

| DivI of t * t

| DivF of t * t

| And of t * t

| Or of t * t

| Not of t

| Mod of t * t

| BitAnd of t * t

| BitOr of t * t

| BitNot of t

| BitShiftLeft of t * t

| BitShiftRight of t * t

| BitXor of t * t

| StringConcat of t * t

| IsInstanceOf of {

var : Var.t;

typ : IR.Typ.t;

nullable : bool;

}

| IsInt of t

val compare : t -> t -> int

val equal : t -> t -> bool

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val equal_syntax : t -> t -> bool

val needs_paren : t -> bool

val pp_paren : (F.formatter -> LinArith.Var.t -> unit) -> needs_paren:(t -> bool) -> F.formatter -> t -> unit

val pp_no_paren : (F.formatter -> LinArith.Var.t -> unit) -> F.formatter -> t -> unit

val pp : (F.formatter -> LinArith.Var.t -> unit) -> F.formatter -> t -> unit

val of_q : Q.t -> t

val of_intlit : IR.IntLit.t -> t

val of_const : IR.Const.t -> t

val of_operand : operand -> t

val of_subst_target : t subst_target -> t

val one : t

val zero : t

val of_bool : bool -> t

val of_unop : IR.Unop.t -> t -> t

val of_binop : IR.Binop.t -> t -> t -> t

val is_zero : t -> bool

val is_non_zero_const : t -> bool

val has_known_non_boolean_type : t -> bool

val fold_map_direct_subterms : t -> init:'a -> f:('a -> t -> 'a * t) -> 'a * t

Fold f on the strict sub-terms of t, if any. Preserve physical equality if f does.

val map_direct_subterms : t -> f:(t -> t) -> t

val satunsat_map_direct_subterms : t -> f:(t -> t SatUnsat.Import.sat_unsat_t) -> t SatUnsat.Import.sat_unsat_t

val fold_subst_variables : init:'a -> f_subst:('a -> LinArith.Var.t -> 'a * t LinArith.subst_target) -> ?f_post:(prev:t -> 'a -> t -> 'a * t) -> t -> 'a * t

val fold_variables : init:'a -> f:('a -> LinArith.Var.t -> 'a) -> ?f_post:(prev:t -> 'a -> t -> 'a * t) -> t -> 'a

val iter_variables : t -> f:(LinArith.Var.t -> unit) -> unit

val subst_variables : f:(LinArith.Var.t -> t LinArith.subst_target) -> ?f_post:(prev:t -> unit -> t -> unit * t) -> t -> t

val has_var_notin : Var.Set.t -> t -> bool

val eval_const_shallow_ : t -> t option

reduce to a constant when the direct sub-terms are constants

val eval_const_shallow : t -> t SatUnsat.Import.sat_unsat_t

val simplify_shallow : t -> t SatUnsat.Import.sat_unsat_t

val linearize : t -> t

more or less syntactic attempt at detecting when an arbitrary term is a linear formula; call Atom.eval_term first for best results

val simplify_linear : t -> t

val get_as_isinstanceof : t -> (Var.t * IR.Typ.t * bool) option

val get_as_var : t -> Var.t option

val get_as_const : t -> Q.t option

val is_non_numeric_constant : t -> bool

val get_as_linear : t -> LinArith.t option

val to_subst_target : t -> t subst_target

module VarMap : sig ... end

module Set : sig ... end