Module Pulselib.PulseInterproc

type call_state

type contradiction = private

| Aliasing of {

addr_caller : PulseBasicInterface.AbstractValue.t;

addr_callee : PulseBasicInterface.AbstractValue.t;

addr_callee' : PulseBasicInterface.AbstractValue.t;

call_state : call_state;

}	(*
raised when the precondition and the current state disagree on the aliasing, i.e. some addresses callee_addr and callee_addr' that are distinct in the pre are aliased to a single address caller_addr in the caller's current state. Typically raised when calling foo(z,z) where the spec for foo(x,y) says that x and y are disjoint.
*)

| CapturedFormalActualLength of {

captured_formals : (IR.Var.t * IR.Typ.t) list;

captured_actuals : ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list;

}

| FormalActualLength of {

formals : (IR.Var.t * IR.Typ.t) list;

actuals : ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list;

}

| PathCondition

val apply_summary : PulseDomainInterface.PathContext.t -> IR.Procname.t -> IBase.Location.t -> callee_summary:PulseDomainInterface.AbductiveDomain.Summary.t -> captured_formals:(IR.Var.t * IR.Typ.t) list -> captured_actuals: ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list -> formals:(IR.Var.t * IR.Typ.t) list -> actuals: ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list -> PulseDomainInterface.AbductiveDomain.t -> (PulseDomainInterface.AbductiveDomain.t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) option * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) PulseBasicInterface.AbstractValue.Map.t) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t * contradiction option

result of applying one pre/post pair of a callee's summary:

PulseSatUnsat.t.Unsat if that path in the callee is infeasible

otherwise, there can be an error detected

otherwise, the result is a new abstract state, an optional return value, and a substitution callee_abstract_value -> caller_abstract_value mapping callee's abstract values to what they became in the new (caller) state