Module Pulselib.PulseMutualRecursion

module F = Stdlib.Format

module PathContext = PulsePathContext

type t = Pulselib.PulseBasicInterface.Trace.t

the trace represents the start of a cycle but isn't a cycle yet, it's a sequence of calls that end where ondemand returned no summary due to mutual recursion; these proto-cycles get bubbled the cyclic call stack until the cycle is closed again

val mk : PathContext.t -> IBase.Location.t -> IR.Procname.t -> t

a trace of length 1

val get_inner_call : t -> IR.Procname.t

the "initial", innermost call in the cycle

val pp : F.formatter -> t -> unit