Module Pulselib.PulseOperations

type t = PulseDomainInterface.AbductiveDomain.t

val check_addr_access : PulseDomainInterface.PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> PulseOperationResult.Import.access_mode -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

Check that the address is not known to be invalid

module Closures : sig ... end

val pulse_model_type : IR.Typ.name

Struct type name of "__infer_pulse_model"

module ModeledField : sig ... end

val conservatively_initialize_args : PulseBasicInterface.AbstractValue.t list -> t -> t

Set all reachable values from the given list as initialized conservatively.

val eval : PulseDomainInterface.PathContext.t -> PulseOperationResult.Import.access_mode -> IBase.Location.t -> IR.Exp.t -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t

Use the stack and heap to evaluate the given expression down to an abstract address representing its value.

Return an error state if it traverses some known invalid address or if the end destination is known to be invalid.

val eval_var : PulseDomainInterface.PathContext.t -> IBase.Location.t -> IR.Pvar.t -> t -> t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)

Similar to eval but for pvar only. Always succeeds.

val prune : PulseDomainInterface.PathContext.t -> IBase.Location.t -> condition:IR.Exp.t -> t -> (t * PulseBasicInterface.ValueHistory.t) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t

val eval_deref : PulseDomainInterface.PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> IBase.Location.t -> IR.Exp.t -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t

Like eval but evaluates *exp.

val eval_access : PulseDomainInterface.PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> PulseOperationResult.Import.access_mode -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseDomainInterface.BaseMemory.Access.t -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t

Like eval but starts from an address instead of an expression, checks that it is valid, and if so dereferences it according to the access.

val eval_deref_access : PulseDomainInterface.PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> PulseOperationResult.Import.access_mode -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseDomainInterface.BaseMemory.Access.t -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t

Like eval_access but does additional dereference.

val eval_proc_name : PulseDomainInterface.PathContext.t -> IBase.Location.t -> IR.Exp.t -> t -> (t * IR.Procname.t option) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t

val havoc_id : IR.Ident.t -> PulseBasicInterface.ValueHistory.t -> t -> t

val havoc_deref_field : PulseDomainInterface.PathContext.t -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> IR.Fieldname.t -> PulseBasicInterface.ValueHistory.t -> t -> t PulseDomainInterface.AccessResult.t

Havoc dereferenced field address.

val realloc_pvar : IR.Tenv.t -> PulseDomainInterface.PathContext.t -> IR.Pvar.t -> IR.Typ.t -> IBase.Location.t -> t -> t

val write_id : IR.Ident.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t

val read_id : IR.Ident.t -> t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) option

val write_field : PulseDomainInterface.PathContext.t -> IBase.Location.t -> ref: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> IR.Fieldname.t -> obj: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

write the edge ref --.field--> obj

val write_deref_field : PulseDomainInterface.PathContext.t -> IBase.Location.t -> ref: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> IR.Fieldname.t -> obj: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

write the edge ref --.field--> _ --*--> obj

val write_arr_index : PulseDomainInterface.PathContext.t -> IBase.Location.t -> ref: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> index:PulseBasicInterface.AbstractValue.t -> obj: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

write the edge ref[index]--> obj

val write_deref : PulseDomainInterface.PathContext.t -> IBase.Location.t -> ref: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> obj: (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

write the edge ref --*--> obj

type invalidation_access =

| MemoryAccess of {

pointer : PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t;

access : PulseDomainInterface.BaseMemory.Access.t;

hist_obj_default : PulseBasicInterface.ValueHistory.t;

}	(*
the value was read from the heap following the access edge at address pointer
*)

| StackAddress of IR.Var.t * PulseBasicInterface.ValueHistory.t	(*
the value was read from the stack
*)

| UntraceableAccess	(*
we don't know where the value came from; avoid using if possible
*)

the way that was used to get to the invalidated address in the state; this is used to record the invalidation point in its history in addition to inside the Invalid attribute

val invalidate : PulseDomainInterface.PathContext.t -> invalidation_access -> IBase.Location.t -> PulseBasicInterface.Invalidation.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

record that the address is invalid

val always_reachable : PulseBasicInterface.AbstractValue.t -> t -> t

val allocate : PulseBasicInterface.Attribute.allocator -> IBase.Location.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val java_resource_release : recursive:bool -> PulseBasicInterface.AbstractValue.t -> t -> t

releases the resource of the argument, and recursively calls itself on the delegated resource if recursive==true

val csharp_resource_release : recursive:bool -> PulseBasicInterface.AbstractValue.t -> t -> t

releases the resource of the argument, and recursively calls itself on the delegated resource if recursive==true

val add_dynamic_type : IR.Typ.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val add_dynamic_type_source_file : IR.Typ.t -> IBase.SourceFile.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val add_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> t

val is_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> bool

val remove_allocation_attr : PulseBasicInterface.AbstractValue.t -> t -> t

val invalidate_access : PulseDomainInterface.PathContext.t -> IBase.Location.t -> PulseBasicInterface.Invalidation.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseDomainInterface.BaseMemory.Access.t -> t -> t PulseDomainInterface.AccessResult.t

record that what the address points via the access to is invalid

val invalidate_deref_access : PulseDomainInterface.PathContext.t -> IBase.Location.t -> PulseBasicInterface.Invalidation.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseDomainInterface.BaseMemory.Access.t -> t -> t PulseDomainInterface.AccessResult.t

Like invalidate_access but invalidates dereferenced address.

val invalidate_array_elements : PulseDomainInterface.PathContext.t -> IBase.Location.t -> PulseBasicInterface.Invalidation.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> t PulseDomainInterface.AccessResult.t

record that all the array elements that address points to is invalid

val shallow_copy : PulseDomainInterface.PathContext.t -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t

returns the address of a new cell with the same edges as the original

val deep_copy : ?depth_max:int -> PulseDomainInterface.PathContext.t -> IBase.Location.t -> (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> t -> (t * (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t)) PulseDomainInterface.AccessResult.t

returns the address of a new cell with the copied edges from the original. The content is deeply copied up until max_depth edges deep in memory. The deepest copied value is then shallow copied. If no max_depth is specified, then there is no shallow copy and everything is deeply copied until there is no more edge to follow

val get_dynamic_type_unreachable_values : IR.Var.t list -> t -> (IR.Var.t * IR.Typ.t) list

Given a list of variables, computes the unreachable values if the variables were removed from the stack, then return the dynamic types of those values if they are available

val remove_vars : IR.Var.t list -> IBase.Location.t -> t -> t

val check_address_escape : IBase.Location.t -> IR.Procdesc.t -> PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.ValueHistory.t -> t -> t PulseDomainInterface.AccessResult.t

type call_kind = [

| `Closure of (IR.Exp.t * IR.Pvar.t * IR.Typ.t * IR.CapturedVar.capture_mode) list

| `Var of IR.Ident.t

| `ResolvedProcname

]

val get_captured_actuals : IR.Procname.t -> PulseDomainInterface.PathContext.t -> IBase.Location.t -> captured_formals:(IR.Var.t * IR.CapturedVar.capture_mode * IR.Typ.t) list -> call_kind:call_kind -> actuals: ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list -> t -> (t * ((PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list) PulseDomainInterface.AccessResult.t PulseBasicInterface.SatUnsat.t

val check_used_as_branch_cond : (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> pname_using_config:IR.Procname.t -> branch_location:IBase.Location.t -> location:IBase.Location.t -> PulseBasicInterface.Trace.t -> PulseDomainInterface.AbductiveDomain.t -> PulseDomainInterface.AbductiveDomain.t PulseDomainInterface.AccessResult.t

Check and report config usage issue on the abstract value that is used as branch condition. If it is not certain that tha abstract value is a config, it adds UsedAsBranchCond attribute.