Module PulseOperations.Import

For opening in other modules.

type access_mode =

| Read

| Write

| NoAccess	(*
The initialized-ness of the address is not checked when it evaluates a heap address without actual memory access, for example, when evaluating &x.f we need to check initialized-ness of x, not that of x.f.
*)

Imported types for ease of use and so we can write variants without the corresponding module prefix

type 'abductive_domain_t execution_domain_base_t = 'abductive_domain_t PulseDomainInterface.ExecutionDomain.base_t =

| ContinueProgram of 'abductive_domain_t

| ExceptionRaised of 'abductive_domain_t

| ExitProgram of PulseDomainInterface.AbductiveDomain.summary

| AbortProgram of PulseDomainInterface.AbductiveDomain.summary

| LatentAbortProgram of {

astate : PulseDomainInterface.AbductiveDomain.summary;

latent_issue : PulseDomainInterface.LatentIssue.t;

}

| LatentInvalidAccess of {

astate : PulseDomainInterface.AbductiveDomain.summary;

address : PulseDomainInterface.Decompiler.expr;

must_be_valid : PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option;

calling_context : (PulseBasicInterface.CallEvent.t * IBase.Location.t) list;

}

| ISLLatentMemoryError of PulseDomainInterface.AbductiveDomain.summary

type base_summary_error = PulseDomainInterface.AccessResult.summary_error =

| PotentialInvalidAccessSummary of {

astate : PulseDomainInterface.AbductiveDomain.summary;

address : PulseDomainInterface.Decompiler.expr;

must_be_valid : PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option;

}

| ReportableErrorSummary of {

astate : PulseDomainInterface.AbductiveDomain.summary;

diagnostic : PulseDomainInterface.Diagnostic.t;

}

| ISLErrorSummary of {

astate : PulseDomainInterface.AbductiveDomain.summary;

}

type base_error = PulseDomainInterface.AccessResult.error =

| PotentialInvalidAccess of {

astate : PulseDomainInterface.AbductiveDomain.t;

address : PulseDomainInterface.Decompiler.expr;

must_be_valid : PulseBasicInterface.Trace.t * PulseBasicInterface.Invalidation.must_be_valid_reason option;

}

| ReportableError of {

astate : PulseDomainInterface.AbductiveDomain.t;

diagnostic : PulseDomainInterface.Diagnostic.t;

}

| ISLError of {

astate : PulseDomainInterface.AbductiveDomain.t;

}

| Summary of base_summary_error

Monadic syntax

include module type of PulseResult.Let_syntax

include module type of Pulselib.PulseResult.Monad_infix

include module type of Pulselib.PulseResult.Type

for opening locally

type ('ok, 'err) pulse_result = ( 'ok, 'err ) Pulselib.PulseResult.t =

| Ok of 'ok

| Recoverable of 'ok * 'err list

| FatalError of 'err * 'err list

val (>>|) : ( 'ok, 'err ) Pulselib.PulseResult.t -> ( 'ok -> 'okk ) -> ( 'okk, 'err ) Pulselib.PulseResult.t

val (>>=) : ( 'ok, 'err ) Pulselib.PulseResult.t -> ( 'ok -> ( 'okk, 'err ) Pulselib.PulseResult.t ) -> ( 'okk, 'err ) Pulselib.PulseResult.t

val let+ : ( 'ok, 'err ) Pulselib.PulseResult.t -> ( 'ok -> 'okk ) -> ( 'okk, 'err ) Pulselib.PulseResult.t

val let* : ( 'ok, 'err ) Pulselib.PulseResult.t -> ( 'ok -> ( 'okk, 'err ) Pulselib.PulseResult.t ) -> ( 'okk, 'err ) Pulselib.PulseResult.t

val let<*> : 'a PulseDomainInterface.AccessResult.t -> ( 'a -> 'b PulseDomainInterface.AccessResult.t list ) -> 'b PulseDomainInterface.AccessResult.t list

monadic "bind" but not really that turns an AccessResult.t into a list of AccessResult.ts (not really because the first type is not an AccessResult.t list but just an AccessResult.t)

val let<+> : 'a PulseDomainInterface.AccessResult.t -> ( 'a -> 'abductive_domain_t ) -> 'abductive_domain_t execution_domain_base_t PulseDomainInterface.AccessResult.t list

monadic "map" but even less really that turns an AccessResult.t into an analysis result