Module Pulselib.PulseSpecializedCallGraph

val get_missed_captures : get_summary:(IR.Procname.t -> PulseSummary.t option) -> IR.SpecializedProcname.t list -> IR.SpecializedProcname.Set.t IR.Typ.Name.Map.t

traverses stored Pulse summaries and creates a map from missed types to the set of specialized procnames where the types are missing