Module PulseSpecializedCallGraph.JsonBuilder

type t

val make : unit -> t

Create a new builder.

val add : t -> IR.Procname.t -> PulseSummary.t option -> t

Add call graph edges for a caller based on its summary.

val finalize : t -> call_graph

Close the builder and return the resulting call graph.