Module Pulselib.PulseTimestamp

module F = Stdlib.Format

type t = private int

val compare : t -> t -> int

val equal : t -> t -> bool

val t0 : t

val pp : F.formatter -> t -> unit

val incr : t -> t

type trace = private t list

A trace of timestamps: a way of recording evolution of a timestamp over the course of analysis. In particular, this can permit lightweight comparison of histories across interprocedural calls.

val compare_trace : trace -> trace -> int

val equal_trace : trace -> trace -> bool

val trace0 : t -> trace

val pp_trace : F.formatter -> trace -> unit

val add_to_trace : trace -> t -> trace