Module Pulselib.PulseTopl

type value = PulseAbstractValue.t

type event =

| ArrayWrite of {

aw_array : value;

aw_index : value;

}

| Call of {

return : value option;

arguments : value list;

procname : IR.Procname.t;

}

type state

val compare_state : state -> state -> int

val equal_state : state -> state -> bool

val start : unit -> state

Return the initial state of Topl.automaton ().

val small_step : IBase.Location.t -> keep:PulseAbstractValue.Set.t -> get_dynamic_type:( value -> IR.Typ.t option ) -> path_condition:PulsePathCondition.t -> event -> state -> state

val large_step : call_location:IBase.Location.t -> callee_proc_name:IR.Procname.t -> substitution:(value * PulseValueHistory.t) PulseAbstractValue.Map.t -> keep:PulseAbstractValue.Set.t -> get_dynamic_type:( value -> IR.Typ.t option ) -> path_condition:PulsePathCondition.t -> callee_prepost:state -> state -> state

large_step ~substitution ~keep ~condition state ~callee_prepost updates state according to callee_prepost. The abstract values in condition and state are in one scope, and those in callee_prepost in another scope: the substitution maps from the callee scope to the condition&state scope.

val filter_for_summary : get_dynamic_type:( value -> IR.Typ.t option ) -> PulsePathCondition.t -> state -> state

Remove from state those parts that are inconsistent with the path condition. (We do a cheap check to not introduce inconsistent Topl states, but they may become inconsistent because the program path condition is updated later.)

val simplify : keep:PulseAbstractValue.Set.t -> get_dynamic_type:( value -> IR.Typ.t option ) -> path_condition:PulsePathCondition.t -> state -> state

Keep only a subset of abstract values. This is used for extracting summaries.

val report_errors : IR.Procdesc.t -> Absint.Errlog.t -> state -> unit

Calls Reporting.log_issue with error traces, if any.

val pp_state : Stdlib.Format.formatter -> state -> unit

module Debug : sig ... end