Module Pulselib__PulseAbstractValue.State

internal state of the module

Under the hood a "next fresh" reference counter is maintained to produce fresh t. The Constants module also remembers a mapping from certain constants to their corresponding t. Both of these should be per-procedure only so internal state bookkeeping has to be performed by the interprocedural analysis.

type t

val get : unit -> t

val set : t -> unit

val reset : unit -> unit