Module Pulselib__PulsePathContext

module F = Stdlib.Format

type t = {

conditions : Pulselib.PulseBasicInterface.ValueHistory.t list;
Each history represents a conditional that is affecting the path currently, with the most recent conditional first. The idea is to add these histories to the histories of all variables and memory locations modified while under the influence of these conditionals.

timestamp : Pulselib.PulseBasicInterface.Timestamp.t;
step number in an intra-procedural analysis

}

val compare : t -> t -> int

val equal : t -> t -> bool

val leq : lhs:t -> rhs:t -> bool

val equal_fast : t -> t -> bool

val initial : t

val with_context : t -> Pulselib.PulseBasicInterface.ValueHistory.t -> Pulselib.PulseBasicInterface.ValueHistory.t

val post_exec_instr : t -> t

call this after each step of the symbolic execution to update the path information

val pp : F.formatter -> t -> unit