Module Pulselib__PulseValueHistory

module F = Stdlib.Format

module CallEvent = Pulselib.PulseCallEvent

module Timestamp = Pulselib.PulseTimestamp

type event =

| Allocation of {

f : CallEvent.t;

location : IBase.Location.t;

timestamp : Timestamp.t;

}

| Assignment of IBase.Location.t * Timestamp.t

| Call of {

f : CallEvent.t;

location : IBase.Location.t;

in_call : t;

timestamp : Timestamp.t;

}

| Capture of {

captured_as : IR.Pvar.t;

mode : IR.CapturedVar.capture_mode;

location : IBase.Location.t;

timestamp : Timestamp.t;

}

| ConditionPassed of {

if_kind : IR.Sil.if_kind;

is_then_branch : bool;

location : IBase.Location.t;

timestamp : Timestamp.t;

}

| CppTemporaryCreated of IBase.Location.t * Timestamp.t

| FormalDeclared of IR.Pvar.t * IBase.Location.t * Timestamp.t

| Invalidated of Pulselib.PulseInvalidation.t * IBase.Location.t * Timestamp.t

| NilMessaging of IBase.Location.t * Timestamp.t

| Returned of IBase.Location.t * Timestamp.t

| StructFieldAddressCreated of IR.Fieldname.t IStdlib.RevList.t * IBase.Location.t * Timestamp.t

| VariableAccessed of IR.Pvar.t * IBase.Location.t * Timestamp.t

| VariableDeclared of IR.Pvar.t * IBase.Location.t * Timestamp.t

and t =

| Epoch
start of time

| Sequence of event * t
Sequence [event, hist] represents an event event occurring *after* hist. Invariant: the timestamp of event is greater than all the (local, i.e. not inside function calls) timestamps in hist.

| Branching of t list
several parallel histories

val compare_event : event -> event -> int

val compare : t -> t -> int

val equal_event : event -> event -> bool

val equal : t -> t -> bool

val yojson_of_event : event -> Ppx_yojson_conv_lib.Yojson.Safe.t

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val pp : F.formatter -> t -> unit

val pp_fields : F.formatter -> IR.Fieldname.t IStdlib.RevList.t -> unit

val singleton : event -> t

type iter_event =

| EnterCall of CallEvent.t * IBase.Location.t

| ReturnFromCall of CallEvent.t * IBase.Location.t

| Event of event

val iter : t -> f:(iter_event -> unit) -> unit

iterate on all events in reverse timestamp order, recursing into the histories inside call events. Timestamp order is the lexicographic order induced by projecting events onto their timestamps and appending timestamps within calls, e.g. the timestamp of the inner assignement in

Call {timestamp=10;
      in_call=[...,
               Call{timestamp=4;
                    in_call=[...,
                             Assignement (...,timestamp=3)
                            ]
                   }
              ]
     }

can be written 10.4.3 and the order is such that, e.g., 10.4.3 < 10.5, 10.5 being the timestamp of the event following the inner Call event in the example above.

val location_of_event : event -> IBase.Location.t

val add_to_errlog : nesting:int -> t -> Absint.Errlog.loc_trace_elem list -> Absint.Errlog.loc_trace_elem list

val get_first_event : t -> event option