Module PyBuiltin.Set

type t

This module keeps track of the builtins used by a code unit. Only the necessary Textual declarations are generated. Note that primitive wrappers are always generated (python_int, ...)

val supported_builtins : unit -> string list

val empty : t

val to_textual : t -> Textuallib.Textual.Module.decl list

Encode a set of builtin declarations into Textual declarations

val register : t -> builtin -> t

val get_type : builtin -> Textuallib.Textual.Typ.t