Module PythonFrontend.PyCommon

val type_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.TypeName.t

val proc_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.ProcName.t

val var_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.VarName.t

val node_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.NodeName.t

val field_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.FieldName.t

val qualified_procname : enclosing_class:Textuallib.Textual.TypeName.t -> Textuallib.Textual.ProcName.t -> Textuallib.Textual.qualified_procname

val builtin_name : string -> Textuallib.Textual.qualified_procname

Helper function to encode known builtin names correctly

val python_int : Textuallib.Textual.qualified_procname

Encoding of Python int type. Since Python integers are of arbitrary precision, they are not modeled directly with int.

val python_float : Textuallib.Textual.qualified_procname

Encoding of Python float type.

val python_string : Textuallib.Textual.qualified_procname

Encoding of Python str type.

val python_bytes : Textuallib.Textual.qualified_procname

Encoding of Python bytes type.

val python_bool : Textuallib.Textual.qualified_procname

Encoding of Python bool type.

val python_tuple : Textuallib.Textual.qualified_procname

Encoding of Python tuple type. It is the raw "untyped" one where every item is of type object.

val pyObject : Textuallib.Textual.Typ.t

object is the top type of Python. It helps us when no type information is available.

val pyCode : Textuallib.Textual.Typ.t

code is a builtin Python type to describe any code (function, class, ...) object

val pyClass : Textuallib.Textual.Typ.t

Like pyCode but for class creation

val pyIterItem : Textuallib.Textual.Typ.t

Wrapper type to implement next access on an iterator. Declared as pyIterItemStruct

val py_iter_item_has_item : Textuallib.Textual.qualified_fieldname

val py_iter_item_next_item : Textuallib.Textual.qualified_fieldname

val pyIterItemStruct : Textuallib.Textual.Struct.t

Textual structure with two fields. has_item indicates if something was read from the iterator or if it was empty. If has_item is true, next_item holds the value read from the iterator.

val pyMethod : Textuallib.Textual.Typ.t

Wrapper type to implement LOAD_METHOD. Declared as pyMethodStruct

val py_method_code : Textuallib.Textual.qualified_fieldname

val py_method_self : Textuallib.Textual.qualified_fieldname

val pyMethodStruct : Textuallib.Textual.Struct.t

Textual structure with three fields.

code stores the code of the method called, or an arbitrary callable the method was not found

self stores a reference to the instance where the method was found, or NULL if the method was not found

val pyInt : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type int

val pyString : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type str

val pyBytes : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type bytes

val pyBool : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type bool

val pyFloat : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type float

val pyNone : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type None

val mk_int : int64 -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal integers.

val mk_float : float -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal floats.

val mk_string : string -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal strings.

val mk_bytes : bytes -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal bytes.

val mk_bool : bool -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for boolean.

val unknown_global : string -> string

Wrap a variable name into a special enclosing class when a global's origin can't be found.

type annotated_name = {

name : string;

annotation : string;

}

Encoding of some type annotation like x: int

type method_info = {

name : string;

raw_qualified_name : string;

code : FFI.Constant.t;

signature : annotated_name list;

flags : int;

}

Method declaration info (name, signature, ...

val toplevel_function : string