Module PythonFrontend.PyEnv

module T = Textuallib.Textual

module DataStack : sig ... end

In Python, everything is an object, and the interpreter maintains a stack of references to such objects. Pushing and popping on the stack are always references to objets that leave in a heap. There is no need to model this heap, but the data stack is quite important.

type info = {

is_code : bool;
(*
is the entity a function / method ?
*)

is_class : bool;
(*
is the entity a class ?
*)

typ : T.Typ.t;
(*
Type annotation, if any (otherwise, object is used)
*)

}

Type information about various entities (toplevel declarations, temporaries, ...).

module SMap : IStdlib.IStd.Caml.Map.S with type key = string

module Symbol : sig ... end

module Import : sig ... end

type t

Global environment used during bytecode processing. Stores common global information like the toplevel symbols processed so far, or more local ones like the set of labels or variable ids currently used by a declaration.

module Label : sig ... end

val empty : t

val loc : t -> T.Location.t

Return the last recorded line information from the Python code-unit, if any.

val stack : t -> DataStack.t

Returns the DataStack.t for the current declaration

val globals : t -> Symbol.t SMap.t

Return the globals map

val get_used_builtins : t -> PyBuiltin.Set.t

Return a set of Builtin the we spotted in the code

val instructions : t -> T.Instr.t list

Returns the list of all instructions recorded for the current code unit

val label_of_offset : t -> int -> Label.info option

Check if the instruction is a possible jump location, and return the label information found there, if any.

val mk_fresh_ident : t -> info -> t * T.Ident.t

Generate a fresh temporary name

val get_ident_info : t -> T.Ident.t -> info option

Get back the information of a temporary

val mk_fresh_label : t -> t * string

Generate a fresh label name

val map : f:(t -> 'a -> t * 'b) -> env:t -> 'a list -> t * 'b list

Similar to List.map but an env is threaded along the way

val enter_proc : is_toplevel:bool -> module_name:string -> t -> t

Set the environment when entering a new code unit (like reset the instruction buffer, or id/label generators.

val enter_node : t -> t

Set the environment when entering a new node. Reset the instructions buffer.

val reset_stack : t -> t

Reset the stack field of a node

val update_last_line : t -> int option -> t

Update the last_line field of an env, if new information is availbe.

val push : t -> DataStack.cell -> t

Push a new DataStack.cell on the datastack

val pop : t -> (t * DataStack.cell) option

Pop a DataStack.cell from the datastack, if any is available

val peek : t -> DataStack.cell option

Peek a DataStack.cell from the datastack, if any is available

val push_instr : t -> T.Instr.t -> t

Record a new instruction for the current code unit

val register_label : offset:int -> Label.info -> t -> t

Register the fact that a new label info must be inserted before the instruction at offset

val process_label : offset:int -> Label.info -> t -> t

Mark the label info at offset as processed

val register_symbol : t -> global:bool -> string -> Symbol.t -> bool * t

Register a name (function, variable, ...). It might be a global symbol at the module level or in a local object. Also returns if the symbol was already bound

val lookup_symbol : t -> global:bool -> string -> Symbol.t option

Lookup information about a global/local symbol previously registered via register_symbol

val register_call : t -> string -> t

Register a function call. It enables us to deal correctly with builtin declaration.

val mk_builtin_call : t -> PyBuiltin.textual -> T.Exp.t list -> t * T.Ident.t * T.Typ.t

Wrapper to compute the Textual version of a call to a "textual" builtin * function (a builtin we introduced for modeling purpose)

val register_function : t -> string -> T.Location.t -> PyCommon.annotated_name list -> t

Register a function declaration. We keep track of them since they might shadow Python builtins or previous definitions

val register_method : t -> enclosing_class:string -> method_name:string -> PyCommon.annotated_name list -> t

Register a method declaration. We mostly keep track of their signatures

val lookup_signature : t -> T.enclosing_class -> T.ProcName.t -> PyCommon.annotated_name list option

Lookup the signature of a function / method

val register_class : t -> string -> t

Register a class declaration (based on LOAD_BUILD_CLASS)

val get_declared_classes : t -> string list

val register_import : t -> Import.t -> t

Register a import declaration (based on IMPORT_NAME)

val get_textual_imports : t -> T.Module.decl list

Get back the list of registered imports

val is_toplevel : t -> bool

Are we processing top level instructions, or something in a function/class ?

val module_name : t -> string

Returns the name of the current module