Module PyEnv.DataStack

In Python, everything is an object, and the interpreter maintains a stack of references to such objects. Pushing and popping on the stack are always references to objets that leave in a heap. There is no need to model this heap, but the data stack is quite important.

type cell =

| Const of int
(*
index in co_consts
*)

| Name of {

global : bool;

ndx : int;

}
(*
reference to a name, stored in co_names.
*)

| VarName of int
(*
reference to a local name, stored in co_varnames
*)

| Temp of T.Ident.t
(*
SSA variable
*)

| Code of {

fun_or_class : bool;

code_name : string;

code : FFI.Code.t;

}
(*
code Python object with its name. It can be a function, class, closure, ...
*)

| Map of (string * cell) list
(*
Light encoding of raw Python tuples/dicts. Only used for type annotations at the moment.
*)

| BuiltinBuildClass
(*
see Python's LOAD_BUILD_CLASS
*)

| Import of {

import_path : string;

symbols : string list;

}
(*
imported module path, with optional name of symbols
*)

| ImportCall of T.qualified_procname
(*
Static call to export definition
*)

val pp_cell : Ppx_show_runtime.Format.formatter -> cell -> unit

val show_cell : cell -> string

val as_code : FFI.Code.t -> cell -> FFI.Code.t option

type t = cell list