Module PyEnv.Symbol

type kind =

| Name of {

is_imported : bool;

typ : T.Typ.t;

}
(*
The identifier is a name, like a variable name, or any name that might have been imported
*)

| Builtin
(*
The identifier is a known builtin, like print or range
*)

| Code
(*
The identifier is a code object, like a function declaration
*)

| Class
(*
The identifier is a class name
*)

| ImportCall
(*
The identifier is an imported name that has been used in a call, so we can suppose it is a function
*)

| Import
(*
The identifier is the name of an imported module
*)

val pp_kind : Stdlib.Format.formatter -> kind -> unit

type t = {

id : Ident.t;

kind : kind;

loc : T.Location.t;

}

type key =

| Global of Ident.t

| Local of string

val pp_key : Stdlib.Format.formatter -> key -> unit

val pp : Stdlib.Format.formatter -> t -> unit