Module Quandary.TaintSpec

combination of a trace with functions for handling unknown code and converting to and from summaries

type action =

| Propagate_to_actual of int	(*
Propagate taint from all actuals to the actual with the given index
*)

| Propagate_to_receiver	(*
Propagate taint from all non-receiver actuals to the receiver actual
*)

| Propagate_to_return	(*
Propagate taint from all actuals to the return value
*)

module type S = sig ... end