Module type TaintSpec.S

module Trace : Absint.TaintTrace.S

module AccessTree : sig ... end

val handle_unknown_call : IR.Procname.t -> IR.Typ.t -> Absint.HilExp.t list -> IR.Tenv.t -> action list

return a summary for handling an unknown call at the given site with the given return type and actuals

val get_model : IR.Procname.t -> IR.Typ.t -> Absint.HilExp.t list -> IR.Tenv.t -> AccessTree.t -> action list option

returns a model that should be used for the given (procname, return type, actuals, summary) instead of using the summary for the procname

val is_taintable_type : IR.Typ.t -> bool

return true if the given typ can be tainted

val to_summary_access_tree : AccessTree.t -> QuandarySummary.AccessTree.t

val of_summary_access_tree : QuandarySummary.AccessTree.t -> AccessTree.t

val name : string