Module Textuallib

module CombinedLexer : sig ... end

module CombinedMenhir : sig ... end

module DoliAst : sig ... end

module DoliJavaAst : sig ... end

We can handle Java signatures like the one below

module DoliObjCAst : sig ... end

module DoliParser : sig ... end

module DoliToTextual : sig ... end

module LineMap : sig ... end

module Textual : sig ... end

module TextualBasicVerification : sig ... end

module TextualDecls : sig ... end

module TextualParser : sig ... end

module TextualSil : sig ... end

module TextualTransform : sig ... end

module TextualTypeVerification : sig ... end

This module provides basic type verification for Textual programs.