Module Textuallib.DoliToTextual

val program_to_textual_module : Textual.SourceFile.t -> Doli.doliProgram -> Textual.Module.t

program_to_textual_module takes a string representing a filename, and a doliProgram, and returns a module which contains declarations for all the doli rules from the program. Moreover, it stores the declarations in the global variable global_doli_matcher

val matcher : IR.Procname.t -> IR.Procname.t option

matcher procname searches in the declarations in the global variable global_doli_matcher for a doli-model that matches procname. If it finds one, then it returns the internal name of that doli-model.