Module Textual.ProcDecl

type t = {

qualified_name : qualified_procname;

formals_types : Typ.annotated list;

are_formal_types_fully_declared : bool;	(*
The list of formal argument types may not be fully specified. It is possible only for external function declarations when translating from Hack and is denoted with a special typ1, typ2, ... syntax. Functions defined within a textual module always have a fully declared list of formal parameters.
*)

result_type : Typ.annotated;

attributes : Attr.t list;

}

val formals_or_die : ?context:string -> t -> Typ.annotated list

val pp : F.formatter -> t -> unit

val of_unop : IR.Unop.t -> qualified_procname

val to_unop : qualified_procname -> IR.Unop.t option

val of_binop : IR.Binop.t -> qualified_procname

val to_binop : qualified_procname -> IR.Binop.t option

val is_cast_builtin : qualified_procname -> bool

val allocate_object_name : qualified_procname

val is_allocate_object_builtin : qualified_procname -> bool

val allocate_array_name : qualified_procname

val is_allocate_array_builtin : qualified_procname -> bool

val is_side_effect_free_sil_expr : qualified_procname -> bool

val is_not_regular_proc : qualified_procname -> bool