Module Textuallib.TextualOfSil

val from_c : IBase.SourceFile.t -> string -> IR.Tenv.t -> IR.Cfg.t -> unit

generate a .sil (named filename) file from the source file, containing all the functions and type declarations in the given cfg

val from_java : filename:string -> IR.Tenv.t -> IR.Cfg.t -> unit

generate a .sil file with name filename containing all the functions in the given cfg (Java)

val to_string : lang:Textual.Lang.t -> ?sil_source_file:IBase.SourceFile.t -> filename:string -> IR.Tenv.t -> IR.Cfg.t -> string

return textual SIL as a string