Module Textuallib.TextualSil

val proc_decl_to_sil : Textual.Lang.t -> Textual.ProcDecl.t -> IR.Procname.t

val module_to_sil : Textual.Module.t -> IR.Cfg.t * IR.Tenv.t

val from_java : filename:string -> IR.Tenv.t -> IR.Cfg.t -> unit

generate a .sil file with name filename containing all the functions in the given cfg

val dump_module : filename:string -> Textual.Module.t -> unit

generate a .sil file with name filename with all the content of the input module

val hack_dict_type_name : IR.Typ.name

val hack_mixed_type_name : IR.Typ.name

val hack_mixed_static_companion_type_name : IR.Typ.name

val hack_builtins_type_name : IR.Typ.name

val wildcard_sil_fieldname : Textual.Lang.t -> string -> IR.Fieldname.t