Module TreeSitterFrontend.TreeSitterCTranslator

val translate_xml : string -> file_name:string -> Textuallib.Textual.Module.t

Parse tree-sitter XML output for a C source file and produce a Textual module. file_name is the original source file path for source location mapping.

val translate_cst : TreeSitterFFI.cst_node -> file_name:string -> Textuallib.Textual.Module.t

Translate a CST node (from FFI or XML parsing) into a Textual module.