Module InferBase.ZipLib

val load : 'a Serialization.serializer -> string -> 'a option

load serializer path searches for the file at the given path in the zip libraries. If Config.infer_cache is set, already deserialized data will be saved there and path will be searched from the cache first.