Module InferIR__Objc_models

This module models special c struct types from the Apple's Core Foundation libraries for which there are particular rules for memory management.

val is_core_lib_type : InferIR.Typ.t -> bool

val is_malloc_model : InferIR.Typ.t -> InferIR.Typ.Procname.t -> bool