Module InferBase.ResultsDir

val assert_results_dir : string -> unit

Check that the results dir exists and sets up logging, the database, etc.

val remove_results_dir : unit -> unit

Recursively delete the results directory.

val create_results_dir : unit -> unit

Create the results dir and sets up logging, the database, etc.