Module ATDGenerated.InferCommand

type t =

| Analyze
analyze previously captured source files

| AnalyzeJson
analyze captured cfg and tenv json files

| Capture
capture compilation commands and translate source files into infer's intermediate language

| Compile
set up the infer environment then run the compilation commands without capturing the source files

| Debug
print information about internal structures

| Explore
explore infer reports

| Help
documentation about various aspects of infer

| Report
post-process infer results and reports

| ReportDiff
compute the difference of two infer reports

| Run
orchestrate the capture, analysis, and reporting of a compilation command

Main modes of operation for infer

val compare : t -> t -> int

val of_string : string -> t

val to_string : t -> string

val equal : t -> t -> bool

val all_commands : t list

val infer_exe_name : string

val to_exe_name : t -> string

val of_exe_name : string -> t option