Module ATDGenerated.Jsonbug_j

type transitive_callee_resolution = Jsonbug_t.transitive_callee_resolution

val equal_transitive_callee_resolution : transitive_callee_resolution -> transitive_callee_resolution -> bool

type transitive_callee_kind = Jsonbug_t.transitive_callee_kind

val equal_transitive_callee_kind : transitive_callee_kind -> transitive_callee_kind -> bool

type transitive_callee = Jsonbug_t.transitive_callee = {

callsite_filename : string;

callsite_absolute_position_in_file : int;

caller_name : string;

callsite_relative_position_in_caller : int;

kind : transitive_callee_kind;

resolution : transitive_callee_resolution;

}

val equal_transitive_callee : transitive_callee -> transitive_callee -> bool

type taint_extra = Jsonbug_t.taint_extra = {

taint_source : string option;

taint_sink : string option;

taint_policy_privacy_effect : string option;

tainted_expression : string option;

}

val equal_taint_extra : taint_extra -> taint_extra -> bool

type missed_capture = Jsonbug_t.missed_capture = {

class_name : string;

}

val equal_missed_capture : missed_capture -> missed_capture -> bool

type loc = Jsonbug_t.loc = {

file : string;

lnum : int;

cnum : int;

enum : int;

}

val equal_loc : loc -> loc -> bool

type json_trace_item = Jsonbug_t.json_trace_item = {

level : int;

filename : string;

line_number : int;

column_number : int;

description : string;

}

val equal_json_trace_item : json_trace_item -> json_trace_item -> bool

type config_usage_extra = Jsonbug_t.config_usage_extra = {

config_name : string;

function_name : string;

filename : string;

line_number : int;

}

val equal_config_usage_extra : config_usage_extra -> config_usage_extra -> bool

type extra = Jsonbug_t.extra = {

cost_polynomial : string option;

cost_degree : int option;

copy_type : string option;

config_usage_extra : config_usage_extra option;

taint_extra : taint_extra option;

transitive_callees : transitive_callee list;

transitive_missed_captures : missed_capture list;

}

val equal_extra : extra -> extra -> bool

type jsonbug = Jsonbug_t.jsonbug = {

bug_type : string;

qualifier : string;

suggestion : string option;

severity : string;

category : string;

line : int;

column : int;

procedure : string;

procedure_start_line : int;

file : string;

bug_trace : json_trace_item list;

bug_trace_length : int;

bug_trace_max_depth : int;

key : string;

node_key : string option;

hash : string;

dotty : string option;

infer_source_loc : loc option;

bug_type_hum : string;

traceview_id : int option;

censored_reason : string option;

access : string option;

extras : extra option;

}

val equal_jsonbug : jsonbug -> jsonbug -> bool

type report = Jsonbug_t.report

val equal_report : report -> report -> bool

val write_transitive_callee_resolution : Stdlib.Buffer.t -> transitive_callee_resolution -> unit

Output a JSON value of type transitive_callee_resolution.

val string_of_transitive_callee_resolution : ?len:int -> transitive_callee_resolution -> string

Serialize a value of type transitive_callee_resolution into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_transitive_callee_resolution : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> transitive_callee_resolution

Input JSON data of type transitive_callee_resolution.

val transitive_callee_resolution_of_string : string -> transitive_callee_resolution

Deserialize JSON data of type transitive_callee_resolution.

val write_transitive_callee_kind : Stdlib.Buffer.t -> transitive_callee_kind -> unit

Output a JSON value of type transitive_callee_kind.

val string_of_transitive_callee_kind : ?len:int -> transitive_callee_kind -> string

Serialize a value of type transitive_callee_kind into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_transitive_callee_kind : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> transitive_callee_kind

Input JSON data of type transitive_callee_kind.

val transitive_callee_kind_of_string : string -> transitive_callee_kind

Deserialize JSON data of type transitive_callee_kind.

val write_transitive_callee : Stdlib.Buffer.t -> transitive_callee -> unit

Output a JSON value of type transitive_callee.

val string_of_transitive_callee : ?len:int -> transitive_callee -> string

Serialize a value of type transitive_callee into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_transitive_callee : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> transitive_callee

Input JSON data of type transitive_callee.

val transitive_callee_of_string : string -> transitive_callee

Deserialize JSON data of type transitive_callee.

val write_taint_extra : Stdlib.Buffer.t -> taint_extra -> unit

Output a JSON value of type taint_extra.

val string_of_taint_extra : ?len:int -> taint_extra -> string

Serialize a value of type taint_extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_extra

Input JSON data of type taint_extra.

val taint_extra_of_string : string -> taint_extra

Deserialize JSON data of type taint_extra.

val write_missed_capture : Stdlib.Buffer.t -> missed_capture -> unit

Output a JSON value of type missed_capture.

val string_of_missed_capture : ?len:int -> missed_capture -> string

Serialize a value of type missed_capture into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_missed_capture : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> missed_capture

Input JSON data of type missed_capture.

val missed_capture_of_string : string -> missed_capture

Deserialize JSON data of type missed_capture.

val write_loc : Stdlib.Buffer.t -> loc -> unit

Output a JSON value of type loc.

val string_of_loc : ?len:int -> loc -> string

Serialize a value of type loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> loc

Input JSON data of type loc.

val loc_of_string : string -> loc

Deserialize JSON data of type loc.

val write_json_trace_item : Stdlib.Buffer.t -> json_trace_item -> unit

Output a JSON value of type json_trace_item.

val string_of_json_trace_item : ?len:int -> json_trace_item -> string

Serialize a value of type json_trace_item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_json_trace_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> json_trace_item

Input JSON data of type json_trace_item.

val json_trace_item_of_string : string -> json_trace_item

Deserialize JSON data of type json_trace_item.

val write_config_usage_extra : Stdlib.Buffer.t -> config_usage_extra -> unit

Output a JSON value of type config_usage_extra.

val string_of_config_usage_extra : ?len:int -> config_usage_extra -> string

Serialize a value of type config_usage_extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_config_usage_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> config_usage_extra

Input JSON data of type config_usage_extra.

val config_usage_extra_of_string : string -> config_usage_extra

Deserialize JSON data of type config_usage_extra.

val write_extra : Stdlib.Buffer.t -> extra -> unit

Output a JSON value of type extra.

val string_of_extra : ?len:int -> extra -> string

Serialize a value of type extra into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_extra : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> extra

Input JSON data of type extra.

val extra_of_string : string -> extra

Deserialize JSON data of type extra.

val write_jsonbug : Stdlib.Buffer.t -> jsonbug -> unit

Output a JSON value of type jsonbug.

val string_of_jsonbug : ?len:int -> jsonbug -> string

Serialize a value of type jsonbug into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_jsonbug : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> jsonbug

Input JSON data of type jsonbug.

val jsonbug_of_string : string -> jsonbug

Deserialize JSON data of type jsonbug.

val write_report : Stdlib.Buffer.t -> report -> unit

Output a JSON value of type report.

val string_of_report : ?len:int -> report -> string

Serialize a value of type report into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_report : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> report

Input JSON data of type report.

val report_of_string : string -> report

Deserialize JSON data of type report.