Module ATDGenerated.Pulse_config_j

type taint_target = Pulse_config_t.taint_target

type taint_flow = Pulse_config_t.taint_flow = {

source_kinds : string list;

sanitizer_kinds : string list;

sink_kinds : string list;

}

type taint_policy = Pulse_config_t.taint_policy = {

short_description : string;

taint_flows : taint_flow list;

privacy_effect : string option;

exclude_in : string list option;

exclude_matching : string list option;

}

type taint_policies = Pulse_config_t.taint_policies

type argument_constraint = Pulse_config_t.argument_constraint = {

index : int;

type_matches : string list;

}

type matcher = Pulse_config_t.matcher = {

field_regex : string option;

procedure : string option;

procedure_regex : string option;

class_name_regex : string option;

class_names : string list option;

class_with_annotation : string option;

field_names : string list option;

method_names : string list option;

method_return_type_names : string list option;

overrides_of_class_with_annotation : string option;

method_with_annotation : string option;

field_with_annotation : string option;

annotation_values : string list option;

block_passed_to : string option;

block_passed_to_regex : string option;

allocation : string option;

argument_constraints : argument_constraint list;

kinds : string list option;

taint_target : taint_target option;

exclude_from_regex_in : string list option;

exclude_from_regex_names : string list option;

sanitized_in : string list option;

}

type matchers = Pulse_config_t.matchers

type data_flow_kinds = Pulse_config_t.data_flow_kinds

val write_taint_target : Stdlib.Buffer.t -> taint_target -> unit

Output a JSON value of type taint_target.

val string_of_taint_target : ?len:int -> taint_target -> string

Serialize a value of type taint_target into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_target : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_target

Input JSON data of type taint_target.

val taint_target_of_string : string -> taint_target

Deserialize JSON data of type taint_target.

val write_taint_flow : Stdlib.Buffer.t -> taint_flow -> unit

Output a JSON value of type taint_flow.

val string_of_taint_flow : ?len:int -> taint_flow -> string

Serialize a value of type taint_flow into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_flow : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_flow

Input JSON data of type taint_flow.

val taint_flow_of_string : string -> taint_flow

Deserialize JSON data of type taint_flow.

val write_taint_policy : Stdlib.Buffer.t -> taint_policy -> unit

Output a JSON value of type taint_policy.

val string_of_taint_policy : ?len:int -> taint_policy -> string

Serialize a value of type taint_policy into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_policy : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_policy

Input JSON data of type taint_policy.

val taint_policy_of_string : string -> taint_policy

Deserialize JSON data of type taint_policy.

val write_taint_policies : Stdlib.Buffer.t -> taint_policies -> unit

Output a JSON value of type taint_policies.

val string_of_taint_policies : ?len:int -> taint_policies -> string

Serialize a value of type taint_policies into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_taint_policies : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> taint_policies

Input JSON data of type taint_policies.

val taint_policies_of_string : string -> taint_policies

Deserialize JSON data of type taint_policies.

val write_argument_constraint : Stdlib.Buffer.t -> argument_constraint -> unit

Output a JSON value of type argument_constraint.

val string_of_argument_constraint : ?len:int -> argument_constraint -> string

Serialize a value of type argument_constraint into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_argument_constraint : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> argument_constraint

Input JSON data of type argument_constraint.

val argument_constraint_of_string : string -> argument_constraint

Deserialize JSON data of type argument_constraint.

val write_matcher : Stdlib.Buffer.t -> matcher -> unit

Output a JSON value of type matcher.

val string_of_matcher : ?len:int -> matcher -> string

Serialize a value of type matcher into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_matcher : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> matcher

Input JSON data of type matcher.

val matcher_of_string : string -> matcher

Deserialize JSON data of type matcher.

val write_matchers : Stdlib.Buffer.t -> matchers -> unit

Output a JSON value of type matchers.

val string_of_matchers : ?len:int -> matchers -> string

Serialize a value of type matchers into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_matchers : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> matchers

Input JSON data of type matchers.

val matchers_of_string : string -> matchers

Deserialize JSON data of type matchers.

val write_data_flow_kinds : Stdlib.Buffer.t -> data_flow_kinds -> unit

Output a JSON value of type data_flow_kinds.

val string_of_data_flow_kinds : ?len:int -> data_flow_kinds -> string

Serialize a value of type data_flow_kinds into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_data_flow_kinds : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> data_flow_kinds

Input JSON data of type data_flow_kinds.

val data_flow_kinds_of_string : string -> data_flow_kinds

Deserialize JSON data of type data_flow_kinds.