Module ATDGenerated.Sarifbug_j

type region = Sarifbug_t.region = {

startLine : int;

startColumn : int;

}

type fileloc = Sarifbug_t.fileloc = {

uri : string;

uriBaseId : string;

}

type physicalloc = Sarifbug_t.physicalloc = {

artifactLocation : fileloc;

region : region;

}

type message = Sarifbug_t.message = {

text : string;

}

type trace_loc = Sarifbug_t.trace_loc = {

physicalLocation : physicalloc;

message : message;

}

type trace = Sarifbug_t.trace = {

nestingLevel : int;

location : trace_loc;

}

type threadflow = Sarifbug_t.threadflow = {

locations : trace list;

}

type loc = Sarifbug_t.loc = {

physicalLocation : physicalloc;

}

type fingerprints = Sarifbug_t.fingerprints = {

hashV1 : string;

key : string;

}

type codeflow = Sarifbug_t.codeflow = {

threadFlows : threadflow list;

}

type sarifbug = Sarifbug_t.sarifbug = {

message : message;

level : string;

ruleId : string;

codeFlows : codeflow list option;

locations : loc list;

fingerprints : fingerprints;

}

type rule = Sarifbug_t.rule = {

id : string;

shortDescription : message;

helpUri : string;

}

val write_region : Stdlib.Buffer.t -> region -> unit

Output a JSON value of type region.

val string_of_region : ?len:int -> region -> string

Serialize a value of type region into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_region : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> region

Input JSON data of type region.

val region_of_string : string -> region

Deserialize JSON data of type region.

val write_fileloc : Stdlib.Buffer.t -> fileloc -> unit

Output a JSON value of type fileloc.

val string_of_fileloc : ?len:int -> fileloc -> string

Serialize a value of type fileloc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_fileloc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> fileloc

Input JSON data of type fileloc.

val fileloc_of_string : string -> fileloc

Deserialize JSON data of type fileloc.

val write_physicalloc : Stdlib.Buffer.t -> physicalloc -> unit

Output a JSON value of type physicalloc.

val string_of_physicalloc : ?len:int -> physicalloc -> string

Serialize a value of type physicalloc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_physicalloc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> physicalloc

Input JSON data of type physicalloc.

val physicalloc_of_string : string -> physicalloc

Deserialize JSON data of type physicalloc.

val write_message : Stdlib.Buffer.t -> message -> unit

Output a JSON value of type message.

val string_of_message : ?len:int -> message -> string

Serialize a value of type message into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_message : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> message

Input JSON data of type message.

val message_of_string : string -> message

Deserialize JSON data of type message.

val write_trace_loc : Stdlib.Buffer.t -> trace_loc -> unit

Output a JSON value of type trace_loc.

val string_of_trace_loc : ?len:int -> trace_loc -> string

Serialize a value of type trace_loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_trace_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> trace_loc

Input JSON data of type trace_loc.

val trace_loc_of_string : string -> trace_loc

Deserialize JSON data of type trace_loc.

val write_trace : Stdlib.Buffer.t -> trace -> unit

Output a JSON value of type trace.

val string_of_trace : ?len:int -> trace -> string

Serialize a value of type trace into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_trace : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> trace

Input JSON data of type trace.

val trace_of_string : string -> trace

Deserialize JSON data of type trace.

val write_threadflow : Stdlib.Buffer.t -> threadflow -> unit

Output a JSON value of type threadflow.

val string_of_threadflow : ?len:int -> threadflow -> string

Serialize a value of type threadflow into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_threadflow : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> threadflow

Input JSON data of type threadflow.

val threadflow_of_string : string -> threadflow

Deserialize JSON data of type threadflow.

val write_loc : Stdlib.Buffer.t -> loc -> unit

Output a JSON value of type loc.

val string_of_loc : ?len:int -> loc -> string

Serialize a value of type loc into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_loc : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> loc

Input JSON data of type loc.

val loc_of_string : string -> loc

Deserialize JSON data of type loc.

val write_fingerprints : Stdlib.Buffer.t -> fingerprints -> unit

Output a JSON value of type fingerprints.

val string_of_fingerprints : ?len:int -> fingerprints -> string

Serialize a value of type fingerprints into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_fingerprints : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> fingerprints

Input JSON data of type fingerprints.

val fingerprints_of_string : string -> fingerprints

Deserialize JSON data of type fingerprints.

val write_codeflow : Stdlib.Buffer.t -> codeflow -> unit

Output a JSON value of type codeflow.

val string_of_codeflow : ?len:int -> codeflow -> string

Serialize a value of type codeflow into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_codeflow : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> codeflow

Input JSON data of type codeflow.

val codeflow_of_string : string -> codeflow

Deserialize JSON data of type codeflow.

val write_sarifbug : Stdlib.Buffer.t -> sarifbug -> unit

Output a JSON value of type sarifbug.

val string_of_sarifbug : ?len:int -> sarifbug -> string

Serialize a value of type sarifbug into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_sarifbug : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> sarifbug

Input JSON data of type sarifbug.

val sarifbug_of_string : string -> sarifbug

Deserialize JSON data of type sarifbug.

val write_rule : Stdlib.Buffer.t -> rule -> unit

Output a JSON value of type rule.

val string_of_rule : ?len:int -> rule -> string

Serialize a value of type rule into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_rule : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> rule

Input JSON data of type rule.

val rule_of_string : string -> rule

Deserialize JSON data of type rule.