Module AccessPath.Abs

type raw = t

type t =

| Abstracted of raw
(*
abstraction of heap reachable from an access path, e.g. x.f*
*)

| Exact of raw
(*
precise representation of an access path, e.g. x.f.g
*)

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val equal : t -> t -> bool

val get_footprint_index_base : base -> int option

return the formal index associated with the base of this access path if there is one, or None otherwise

val extract : t -> raw

extract a raw access path from its wrapper

val is_exact : t -> bool

return true if t is an exact representation of an access path, false if it's an abstraction

val pp : Stdlib.Format.formatter -> t -> unit