Module DataFlow.MakeDF

Functor to create an instance of a dataflow analysis.

Parameters

St : DFStateType

Signature

type t

type state = St.t

type transition =

| Dead_state

| Transition of state * state list * state list

val join : state list -> state -> state

val run : IR.Procdesc.t -> state -> IR.Procdesc.Node.t -> transition

Run the dataflow analysis on a procedure starting from the given state. Returns a function to lookup the results of the analysis on every node