Module Absint.Exe_env

Execution environments are a means to get a function's type environment and integer widths and cache those

type t

val mk : unit -> t

Create a new cache

val get_proc_tenv : t -> IR.Procname.t -> IR.Tenv.t

return the type environment associated with the procedure

val get_source_tenv : t -> IBase.SourceFile.t -> IR.Tenv.t option

return the type environment associated with the source file

val load_java_global_tenv : t -> IR.Tenv.t

Load Java type environment (if not done yet), and return it. Useful for accessing type info not related to any concrete function.

val get_integer_type_widths : t -> IR.Procname.t -> IR.IntegerWidths.t

return the integer type widths associated with the procedure