Module Absint.HilInstr

module F = Stdlib.Format

type call =

| Direct of IR.Procname.t

| Indirect of HilExp.t

type of a procedure call; either direct or via function pointer

val compare_call : call -> call -> int

val pp_call : F.formatter -> call -> unit

type t =

| Assign of HilExp.AccessExpression.t * HilExp.t * IBase.Location.t
LHS access expression, RHS expression

| Assume of HilExp.t * [ `Then | `Else ] * IR.Sil.if_kind * IBase.Location.t
Assumed expression, true_branch boolean, source of the assume (conditional, ternary, etc.)

| Call of AccessPath.base * call * HilExp.t list * IR.CallFlags.t * IBase.Location.t
Var to hold the return, call expression, formals

| Metadata of IR.Sil.instr_metadata
see Sil.instr_metadata

val compare : t -> t -> int

val pp : F.formatter -> t -> unit

type translation =

| Instr of t
HIL instruction to execute

| Bind of IR.Var.t * HilExp.AccessExpression.t
add binding to identifier map

Result of translating an SIL instruction

val of_sil : include_array_indexes:bool -> f_resolve_id:(IR.Var.t -> HilExp.AccessExpression.t option) -> IR.Sil.instr -> translation

convert an SIL instruction into an HIL instruction. The f_resolve_id function should map an SSA temporary variable to the access path it represents. Evaluating the HIL instruction should produce the same result as evaluating the SIL instruction and replacing the temporary variables using f_resolve_id.