Module PatternMatch.Java

val implements : string -> IR.Tenv.t -> string -> bool

Check whether class implements a given Java class

val implements_android : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class of Android

val implements_androidx_map : IR.Tenv.t -> string -> bool

Check whether class implements a AndroidX's Map

val implements_app_activity : IR.Tenv.t -> string -> bool

Check whether class implements a class of android.app.Activity

val implements_app_fragment : IR.Tenv.t -> string -> bool

Check whether class implements a class of androidx.fragment.app.Fragment

val implements_arrays : IR.Tenv.t -> string -> bool

Check whether class implements Java's Arrays

val implements_collection : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Collection

val implements_collections : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Collections

val implements_enumeration : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Enumeration

val implements_google : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class of Google

val implements_graphql_story : IR.Tenv.t -> string -> bool

Check whether class implements a class of com.facebook.graphql.model.GraphQLStory

val implements_infer_annotation : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class of Infer annotation

val implements_inject : string -> IR.Tenv.t -> string -> bool

Check whether class implements a Javax Inject

val implements_io : string -> IR.Tenv.t -> string -> bool

Check whether class implements a Java IO

val implements_iterable : IR.Tenv.t -> string -> bool

Check whether class implements Java's Iterable

val implements_iterator : IR.Tenv.t -> string -> bool

Check whether class implements Java's Iterator

val implements_jackson : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class from Jackson

val implements_lang : string -> IR.Tenv.t -> string -> bool

Check whether class implements a Java's lang

val implements_list : IR.Tenv.t -> string -> bool

Check whether class implements a Java's list

val implements_map : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Map

val implements_map_entry : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Map$Entry

val implements_math : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Math

val implements_nio : string -> IR.Tenv.t -> string -> bool

Check whether class implements a Java nio

val implements_number : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Number

val implements_org_json : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class from Json

val implements_pseudo_collection : IR.Tenv.t -> string -> bool

Check whether class implements a pseudo Collection with support for get() and size() methods

val implements_psi_element : IR.Tenv.t -> string -> bool

Check whether class implements a class of com.intellij.psi.PsiElement

val implements_queue : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Queue

val implements_set : IR.Tenv.t -> string -> bool

Check whether class implements a Java's Set

val implements_sparse_float_array : IR.Tenv.t -> string -> bool

Check whether class implements a class of com.facebook.litho.internal.SparseFloatArray

val implements_system : IR.Tenv.t -> string -> bool

Check whether class implements a Java's System

val implements_view_group : IR.Tenv.t -> string -> bool

Check whether class implements a class of android.view.ViewGroup

val implements_view_parent : IR.Tenv.t -> string -> bool

Check whether class implements a class of android.view.ViewParent

val implements_xmob_utils : string -> IR.Tenv.t -> string -> bool

Check whether class implements a class of xmod.utils

val is_throwable : IR.Tenv.t -> IR.Typ.Name.t -> bool

is_throwable tenv class_name checks if class_name is of type java.lang.Throwable

val is_enum : IR.Tenv.t -> IR.Typ.Name.t -> bool

Checks if the type is Java enum (extends java.lang.Enum)

val check_class_attributes : (IR.Annot.Item.t -> bool) -> IR.Tenv.t -> IR.Procname.t -> bool

tests whether any class attributes (e.g., @ThreadSafe) pass check of first argument, including supertypes

val check_current_class_attributes : (IR.Annot.Item.t -> bool) -> IR.Tenv.t -> IR.Procname.t -> bool

tests whether any class attributes (e.g., @ThreadSafe) pass check of first argument, for current class only

val find_superclasses_with_attributes : (IR.Annot.Item.t -> bool) -> IR.Tenv.t -> IR.Typ.Name.t -> IR.Typ.Name.t list

find superclasss with attributes (e.g., @ThreadSafe), including current class

val is_override_of_lang_object_equals : IR.Procname.t -> bool

Whether the method is an override of `java.lang.Object.equals(Object)` or `java.lang.Object.equals(Object)` itself

val method_is_initializer : IR.Tenv.t -> IR.ProcAttributes.t -> bool

Check if the method is one of the known initializer methods.