Module Absint.TaskSchedulerTypes

type target =

| Procname of {

proc_name : IR.Procname.t;

specialization : IR.Specialization.t option;

}

| File of IBase.SourceFile.t

type analysis_result =

| Ok
(*
Analysis finished normally.
*)

| RaceOn of {

dependency_filenames : string list;

}
(*
Analysis stopped when trying to access the summary of a callee and that callee is being analyzed by another worker. dependency_filenames are in the path to the callee's lock file.
*)