Module Absint.TaskSchedulerTypes

type target =

| Procname of IR.Procname.t

| File of IBase.SourceFile.t

| ProcUID of string
(*
matches primary key of procedures and specs tables; see Database.ml
*)

type analysis_result =

| Ok
(*
Analysis finished normally.
*)

| RaceOn of {

dependency_filename : string;

}
(*
Analysis stopped when trying to access the summary of a callee and that callee is being analyzed by another worker. dependency_filename is the path to the callee's lock file.
*)