Module Backend

module AnalysisDependencyGraph : sig ... end

module CallGraph : sig ... end

module CallGraphScheduler : sig ... end

module CallbackOfChecker : sig ... end

Conversions from checkers taking "functional" Absint.InterproceduralAnalysis.t et al. payloads to Callbacks.proc_callback_t and friends.

module Callbacks : sig ... end

Module to register and invoke checkers' callbacks.

module Devirtualizer : sig ... end

module FileScheduler : sig ... end

module GCStats : sig ... end

module InferAnalyze : sig ... end

Main module for the analysis after the capture phase

module Issue : sig ... end

module LineageTaint : sig ... end

module NodePrinter : sig ... end

Simplified html node printer for checkers

module Ondemand : sig ... end

Module for on-demand analysis.

module Payloads : sig ... end

module Preanal : sig ... end

module Printer : sig ... end

Printers for the analysis results

module ProcLocker : sig ... end

module Procedures : sig ... end

module ReactiveCapture : sig ... end

module RegisterCheckers : sig ... end

module ReplayScheduler : sig ... end

module RestartScheduler : sig ... end

module SourceFileGraph : sig ... end

module StarvationGlobalAnalysis : sig ... end

module Summary : sig ... end

Procedure summaries: the results of the capture and all the analysis for a single procedure, plus some statistics

module SyntacticCallGraph : sig ... end

module Tasks : sig ... end