Module Payloads.SQLite

val serialize : t -> old_pulse_payload:Sqlite3.Data.t option -> Sqlite3.Data.t list

serialize payloads, but gets an old Pulse's payload to merge, which was pre-existing in DB

val lazy_load : IBase.Database.analysis_table -> proc_uid:string -> t

load each payload lazily

val eager_load : Sqlite3.stmt -> first_column:int -> t

load all payload columns from the statement starting at first_column

val deserialize_payload_opt : Sqlite3.Data.t -> 'a IStdlib.IStd.Lazy.t option

use to deserialize the data from one column for a particular payload id; callers are responsible for casting the result to the appropriate payload option Lazy.t type