Module Backend.Printer

val curr_html_formatter : Stdlib.Format.formatter IStdlib.IStd.ref

Current html formatter

val node_finish_session : IR.Procdesc.Node.t -> unit

Finish a session, and perform delayed print actions if required

val node_start_session : pp_name:(Stdlib.Format.formatter -> unit) -> IR.Procdesc.Node.t -> int -> unit

Start a session, and create a new html fine for the node if it does not exist yet

val write_proc_html : IR.Procdesc.t -> unit

Write html file for the procedure.

val write_all_html_files : IBase.SourceFile.t -> unit

Group procedures used in the file by their corresponding source files, and create filename.ext.html for each such a file.