Module Backend.Summary

Procedure summaries: the results of the capture and all the analysis for a single procedure, plus some statistics

module Stats : sig ... end

type t = {

payloads : Payloads.t;

mutable sessions : int;	(*
Session number: how many nodes went through symbolic execution
*)

stats : Stats.t;

proc_name : IR.Procname.t;

err_log : Absint.Errlog.t;	(*
Those are issues that are detected for this procedure after per-procedure analysis. In addition to that there can be errors detected after file-level analysis (next stage after per-procedure analysis). This latter category of errors should NOT be written here, use IssueLog and its serialization capabilities instead.
*)

mutable dependencies : IR.Dependencies.t;	(*
Dynamically discovered analysis-time dependencies used to compute this summary
*)

}

summary of a procedure name

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val pp_html : IBase.SourceFile.t -> Stdlib.Format.formatter -> t -> unit

Print the summary in html format

val pp_text : Stdlib.Format.formatter -> t -> unit

Print the summary in text format

module OnDisk : sig ... end