Module Backend.Summary

module Stats : sig ... end

module Status : sig ... end

type t = {

payloads : Payloads.t;

mutable sessions : int;
Session number: how many nodes went through symbolic execution

stats : Stats.t;

status : Status.t;

proc_desc : IR.Procdesc.t;

err_log : Absint.Errlog.t;
Those are issues that are detected for this procedure after per-procedure analysis. In addition to that there can be errors detected after file-level analysis (next stage after per-procedure analysis). This latter category of errors should NOT be written here, use IssueLog and its serialization capabilities instead.

mutable callee_pnames : IR.Procname.Set.t;

}

summary of a procedure name

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val get_proc_name : t -> IR.Procname.t

Get the procedure name

val get_proc_desc : t -> IR.Procdesc.t

val get_err_log : t -> Absint.Errlog.t

val get_status : t -> Status.t

Return the status (active v.s. inactive) of a procedure summary

val pp_html : IBase.SourceFile.t -> Stdlib.Format.formatter -> t -> unit

Print the summary in html format

val pp_text : Stdlib.Format.formatter -> t -> unit

Print the summary in text format

module OnDisk : sig ... end