Module Summary.OnDisk

val clear_cache : unit -> unit

Remove all the elements from the cache of summaries

val get : lazy_payloads:bool -> IBase.AnalysisRequest.t -> IR.Procname.t -> t option

Return the summary option for the procedure name

val reset : IR.Procname.t -> IBase.AnalysisRequest.t -> t

Reset a summary rebuilding the dependents and preserving the proc attributes if present.

val store : IBase.AnalysisRequest.t -> t -> t

Save summary for the procedure into the spec database and return it. If the operation fails, store an empty summary and return that instead.

val add_errlog : IR.Procname.t -> Absint.Errlog.t -> unit

save additional errors in the summary of the given proc name; only use this if you need to report errors in a proc name that is *not* the current proc name (in which case just append to its own errlog), and if you cannot use a file-level analysis to do this either

val delete_all : procedures:IR.Procname.t list -> unit

Similar to delete, but delete all summaries for a list of procedures

val iter_specs : f:(t -> unit) -> unit

Iterates over all stored summaries

val iter_report_summaries_from_config : f: (IR.Procname.t -> IBase.Location.t -> Costlib.CostDomain.summary option -> Costlib.ConfigImpactAnalysis.Summary.t option -> Absint.Errlog.t -> unit) -> unit

Iterates over all analysis artefacts listed above, for each procedure

val get_count : unit -> int

Counts the summaries currently stored on disk.