Module Backend.SyntacticCallGraph

val make : IBase.SourceFile.t list -> (Absint.TaskSchedulerTypes.target, string) IBase.ProcessPool.TaskGenerator.t

task generator that works by

loading the syntactic call graph from the capture DB

restricting it to the reachable procs from the modified files

scheduling leaves only and removing them from the graph when analysed.