Module Backend__Summary.OnDisk

val clear_cache : unit -> unit

Remove all the elements from the cache of summaries

val remove_from_cache : IR.Procname.t -> unit

Remove an element from the cache of summaries. Contrast to reset which re-initializes a summary keeping the same Procdesc and updates the cache accordingly.

val get : IR.Procname.t -> t option

Return the summary option for the procedure name

val reset : IR.Procdesc.t -> t

Reset a summary rebuilding the dependents and preserving the proc attributes if present.

val specs_filename_of_procname : IR.Procname.t -> IBase.DB.filename

Return the path to the .specs file for the given procedure in the current results directory

val load_from_file : IBase.DB.filename -> t option

Load procedure summary from the given file

val proc_resolve_attributes : IR.Procname.t -> IR.ProcAttributes.t option

Try to find the attributes for a defined proc. First look at specs (to get attributes computed by analysis) then look at the attributes table. If no attributes can be found, return None.

val store : t -> unit

Save summary for the procedure into the spec database

val reset_all : filter:IR.Filtering.procedures_filter -> unit -> unit