Module Backend__Summary.OnDisk

val clear_cache : unit -> unit

Remove all the elements from the cache of summaries

val get : IR.Procname.t -> t option

Return the summary option for the procedure name

val reset : IR.Procdesc.t -> t

Reset a summary rebuilding the dependents and preserving the proc attributes if present.

val proc_resolve_attributes : IR.Procname.t -> IR.ProcAttributes.t option

Try to find the attributes for a defined proc. First look at specs (to get attributes computed by analysis) then look at the attributes table. If no attributes can be found, return None.

val store_analyzed : t -> unit

Save summary for the procedure into the spec database

val reset_all : filter:IR.Filtering.procedures_filter -> unit -> unit

val delete : IR.Procname.t -> unit

Delete the .specs file corresponding to the procname and remove its summary from the Summary cache

val iter_specs : f:(t -> unit) -> unit

Iterates over all stored summaries

val iter_report_summaries_from_config : f:(IR.Procname.t -> IBase.Location.t -> Costlib.CostDomain.summary option -> Costlib.ConfigImpactAnalysis.Summary.t option -> Absint.Errlog.t -> unit) -> unit

Iterates over all analysis artefacts listed above, for each procedure

val pp_specs_from_config : Stdlib.Format.formatter -> unit

pretty print all stored summaries