Module Biabduction.BiabductionReporting

val log_issue_using_state : IR.Procdesc.t -> Absint.Errlog.t -> exn -> unit

Add an issue to the given summary using biabduction state.

val log_issue_deprecated_using_state : IR.Procdesc.t -> Absint.Errlog.t -> ?⁠node:IR.Procdesc.Node.t -> ?⁠loc:IBase.Location.t -> ?⁠ltr:Absint.Errlog.loc_trace -> exn -> unit

Report an issue in the given procedure using biabduction state. DEPRECATED as it can create race conditions between checkers. Use log_error_using_state instead