Module Checkers.LineageShape

module FieldLabel : sig ... end

module FieldPath : sig ... end

A module to help manipulating lists of (nested) fields.

module VarPath : sig ... end

A variable path is a pair of a variable and a possibly empty list of subscripted fields. They are built from their in-program occurrences. They may semantically have sub-fields themselves: it is the job of the Cell module to determine the final graph nodes constructed from paths.

module Cell : sig ... end

Cells correspond to variable fields for which no subfield will be considered by the analysis, either because they semantically have none, or because the abstract domain decides that considering them would lead to too deep or too wide field structures.

module Summary : sig ... end

module StdModules : sig ... end

Can be safely opened to provide module definitions at once.

val checker : Summary.t Absint.InterproceduralAnalysis.t -> Summary.t option