Module ClangFrontend

module AstToRangeMap : sig ... end

module Ast_expressions : sig ... end

This module creates extra ast constructs that are needed for the translation

module CAddImplicitDeallocImpl : sig ... end

module CAddImplicitGettersSetters : sig ... end

module CArithmetic_trans : sig ... end

Utility module for translating unary, binary, and atomic operations and compound assignments

module CAst_utils : sig ... end

Functions for transformations of ast nodes

module CContext : sig ... end

Contains current class and current method to be translated as well as local variables, and the cfg, and tenv corresponding to the current file.

module CEnum_decl : sig ... end

Translate an enumeration declaration by adding it to the tenv and

module CField_decl : sig ... end

Utility module to retrieve fields of structs of classes

module CFrontend : sig ... end

module CFrontend_config : sig ... end

Module that contains constants and global state used in the frontend

module CFrontend_decl : sig ... end

Translate declarations *

module CFrontend_errors : sig ... end

module CGeneral_utils : sig ... end

General utility functions such as functions on lists

module CLocation : sig ... end

Module for function to retrieve the location (file, line, etc) of instructions

module CMethodProperties : sig ... end

Functions for extracting properties of functions or method declarations

module CMethodSignature : sig ... end

Define the signature of a method consisting of its name, its arguments, return type, location and whether its an instance method.

module CMethod_trans : sig ... end

Methods for creating a procdesc from a method or function declaration and for resolving a method call and finding the right callee

module CModule_type : sig ... end

module CScope : sig ... end

module CStructUtils : sig ... end

module CTrans : sig ... end

module CTrans_models : sig ... end

module CTrans_utils : sig ... end

module CType : sig ... end

Utility module for retrieving types

module CType_decl : sig ... end

module CType_to_sil_type : sig ... end

module CVar_decl : sig ... end

Process variable declarations by saving them as local or global variables.

module Capture : sig ... end

module ClangCommand : sig ... end

module ClangPointers : sig ... end

module ClangQuotes : sig ... end

module for escaping clang arguments on the command line and put them into files

module ClangWrapper : sig ... end

module Clang_ast_extend : sig ... end

module ObjcCategory_decl : sig ... end

In this module an ObjC category declaration or implementation is processed. The category

module ObjcInterface_decl : sig ... end

In this module an ObjC interface declaration is processed. The class is saved in the tenv as a struct with the corresponding fields, potential superclass and list of defined methods

module ObjcMethod_decl : sig ... end

module ObjcProperty_decl : sig ... end

Process properties by creating their getters and setters in the case that they need to be synthesized

module ObjcProtocol_decl : sig ... end

In this module an ObjC protocol declaration or implementation is processed. The protocol is saved in the tenv as a struct with the corresponding methods

module ProcessAST : sig ... end

module SwitchCase : sig ... end