Module type CModule_type.CTranslation

val instructions_trans : CContext.t -> ATDGenerated.Clang_ast_t.stmt -> CFrontend_config.instr_type list -> IR.Procdesc.Node.t -> is_destructor_wrapper:bool -> IR.Procdesc.Node.t list

It receives the context, a list of statements from clang ast, list of custom statments to be added before clang statements and the exit node and it returns a list of cfg nodes that represent the translation of the stmts into sil.