Module ClangFrontend.Capture

val run_clang : ClangCommand.t -> (IStdlib.IStd.In_channel.t -> 'a) -> 'a

val capture : ClangCommand.t -> unit

If the command is detected to correspond to a source file, translate it.

Given a compilation command, attach our ASTExporter clang plugin to the command and run it.

Our clang plugin emits the AST (Abstract Syntax Tree) as Biniou data that we deserialize. The AST format is described in ATDGenerated.Clang_ast_t (and its Biniou API is in ATDGenerated.Clang_ast_b).

If enabled, invoke translation to IR.Sil via CFrontend.