Module ClangFrontend.ClangQuotes

module for escaping clang arguments on the command line and put them into files

type style =

| EscapedDoubleQuotes
(*
the arguments should be enclosed in "double quotes" and are already escaped
*)

| SingleQuotes
(*
the arguments should be enclosed in 'single quotes' and have to be escaped
*)

| EscapedNoQuotes
(*
the arguments should not be enclosed in quotes and are already escaped
*)

quoting style of the arguments

val quote : style -> string -> string

val mk_arg_file : string -> style -> string list -> string