Module RacerDDomain.Attribute

type t =

| Nothing

| Functional
holds a value returned from a callee marked @Functional

| OnMainThread
boolean is true if the current procedure is running on the main thread

| LockHeld
boolean is true if a lock is currently held

| Synchronized
the object is a synchronized data structure