Module Costlib.BoundMap

module BasicCost = CostDomain.BasicCost

module Node = Absint.ProcCfg.DefaultNode

val lookup_upperbound : BasicCost.t Node.IdMap.t -> Node.id -> BasicCost.t

given a bound map and a node, lookup the number of times it can be executed

val compute_upperbound_map : IR.Procdesc.t -> BO.BufferOverrunAnalysis.invariant_map -> Checkers.Control.invariant_map -> Checkers.LoopInvariant.VarsInLoop.t IR.Procdesc.NodeMap.t -> BasicCost.t Node.IdMap.t

compute a map from each node to an upper bound on the number of times it can be executed.