Module Costlib.CostUtils

module CostModelEnv : sig ... end

type model = CostModelEnv.cost_model_env -> ret:(IR.Ident.t * IR.Typ.t) -> BO.BufferOverrunDomain.Mem.t -> CostDomain.BasicCost.t

val unit_cost_of : of_function:string -> IBase.Location.t -> CostDomain.BasicCost.t

Unit cost value

val cost_of_itv : itv:BO.Itv.t -> degree_kind:BO.Polynomials.DegreeKind.t -> of_function:string -> IBase.Location.t -> CostDomain.BasicCost.t

Translate interval to cost value

val string_len_range_itv : BO.BufferOverrunUtils.ModelEnv.model_env -> IR.Exp.t -> from:(IR.Exp.t * IR.IntegerWidths.t) option -> BO.BufferOverrunDomain.Mem.t -> BO.Itv.t

Given a string of length n and an optional starting index i (0 by default), return itv 0, n_u-i_l

module type S = sig ... end

Module type for getting length of various data types

module Array : S

module Collection : S

module Container : S

module CString : S

module NSCollection : S