Module IBase.DB

Filename

type filename

generic file name

val compare_filename : filename -> filename -> int

val filename_from_string : string -> filename

val filename_to_string : filename -> string

val filename_add_suffix : filename -> string -> filename

val file_exists : filename -> bool

val file_modified_time : ?⁠symlink:bool -> filename -> float

Return the time when a file was last modified. The file must exist.

Results Directory

module Results_dir : sig ... end

val append_crc_cutoff : ?⁠key:string -> string -> string * string

Append a crc to the string, using string_crc_hex32. Cut the string if it exceeds the cutoff limit. Use an optional key to compute the crc. Return a pair of the appended result and crc string.

val source_file_encoding : SourceFile.t -> string

string encoding of a source file (including path) as a single filename

Source Dirs

type source_dir

source directory: the directory inside the results dir corresponding to a source file

val compare_source_dir : source_dir -> source_dir -> int

val source_dir_get_internal_file : source_dir -> string -> filename

get the path to an internal file with the given extention (.tenv, ...)

val source_dir_from_source_file : SourceFile.t -> source_dir

get the source directory corresponding to a source file

val is_source_file : string -> bool

Check if a path is a Java, C, C++ or Objectve C source file according to the file extention