Module DB.Results_dir

type path = string list

path expressed as a list of strings

type path_kind =

| Abs_root
absolute path implicitly rooted at the root of the results dir

| Abs_source_dir of SourceFile.t
absolute path implicitly rooted at the source directory for the file

| Rel
relative path

kind of path: specifies how to interpret a path

val path_to_filename : path_kind -> path -> filename

convert a path to a filename

val init : ?⁠debug:bool -> SourceFile.t -> unit

Initialize the results directory

val create_file : path_kind -> path -> IStdlib.IStd.Unix.File_descr.t

create a file at the given path, creating any missing directories