Module Io_infer.Html

val close : (IStdlib.IStd.Unix.File_descr.t * Stdlib.Format.formatter) -> unit

Close an Html file

val create : IBase.SourceFile.t -> IBase.DB.Results_dir.path -> IStdlib.IStd.Unix.File_descr.t * Stdlib.Format.formatter

Create a new html file

val modified_during_analysis : IBase.SourceFile.t -> IBase.DB.Results_dir.path -> bool

Return true if the html file was modified since the beginning of the analysis

val node_filename : Procname.t -> int -> string

File name for the node, given the procedure name and node id

val open_out : IBase.SourceFile.t -> IBase.DB.Results_dir.path -> IStdlib.IStd.Unix.File_descr.t * Stdlib.Format.formatter

Open an Html file to append data

val pp_line_link : ?with_name:bool -> ?text:string option -> IBase.SourceFile.t -> IBase.DB.Results_dir.path -> Stdlib.Format.formatter -> int -> unit

Print an html link to the given line number of the current source file

val pp_hline : Stdlib.Format.formatter -> unit -> unit

Print a horizontal line

val pp_node_link : IBase.DB.Results_dir.path -> Procname.t -> description:string -> preds:int list -> succs:int list -> exn:int list -> isvisited:bool -> Stdlib.Format.formatter -> int -> unit

Print an html link to the given node. Usage: pp_node_link path_to_root ... fmt id. path_to_root is the path to the dir for the procedure in the spec db. id is the node identifier.

val pp_proc_link : IBase.DB.Results_dir.path -> Procname.t -> Stdlib.Format.formatter -> string -> unit

Print an html link to the given proc

val pp_session_link : ?with_name:bool -> ?proc_name:Procname.t -> IBase.SourceFile.t -> string list -> Stdlib.Format.formatter -> (int * int * int) -> unit

Print an html link given node id and session