Module Procname.Block

type block_type =

| InOuterScope of {

outer_scope : block_type;

block_index : int;

}	(*
a block nested in the scope of an outer one
*)

| SurroundingProc of {

class_name : Typ.name option;

name : string;

}	(*
tracks the name of the surrounding proc and an optional class name where the procedure is defined
*)

Type of Objective C block names.

type t = {

block_type : block_type;

parameters : Parameter.clang_parameter list;

}

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val make_in_outer_scope : block_type -> int -> Parameter.clang_parameter list -> t