Module Procname.Parameter

type clang_parameter = Typ.Name.t option

Type for parameters in clang procnames, Some name means the parameter is of type pointer to struct, with name being the name of the struct, None means the parameter is of some other type.

val compare_clang_parameter : clang_parameter -> clang_parameter -> int

val equal_clang_parameter : clang_parameter -> clang_parameter -> bool

type t =

| JavaParameter of Typ.t

| ClangParameter of clang_parameter

| CSharpParameter of Typ.t

| ErlangParameter

Type for parameters in procnames, for java and clang.

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val of_typ : Typ.t -> clang_parameter