Module Procname.Parameter

type clang_parameter = Typ.Name.t option

Type for parameters in clang procnames, Some name means the parameter is of type pointer to struct, with name being the name of the struct, None means the parameter is of some other type.

val compare_clang_parameter : clang_parameter -> clang_parameter -> int

val equal_clang_parameter : clang_parameter -> clang_parameter -> bool

type t =

| JavaParameter of Typ.t

| ClangParameter of clang_parameter

| CSharpParameter of Typ.t

Type for parameters in procnames, for java and clang.

val compare : t -> t -> int

val equal : t -> t -> bool

val of_typ : Typ.t -> clang_parameter