Module Typ.Name

type t = name

Named types.

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash_normalize : t -> t

val hash_normalize_opt : t option -> t option

val hash_normalize_list : t list -> t list

val compare_name : t -> t -> int

Similar to compare, but compares only names, except template arguments.

val equal : t -> t -> bool

Equality for typenames

val hash : t -> int

val to_string : t -> string

convert the typename to a string

val pp : Stdlib.Format.formatter -> t -> unit

val is_class : t -> bool

is_class name holds if name names CPP/Objc/Java class

val is_union : t -> bool

is_union name holds if name names C/CPP union

val is_same_type : t -> t -> bool

is_class name1 name2 holds if name1 and name2 name same kind of type

val name_without_templates : t -> string

name of the c++ typename without qualifier

val name : t -> string

name of the typename

val qual_name : t -> QualifiedCppName.t

qualified name of the type, may return nonsense for Java classes

val unqualified_name : t -> QualifiedCppName.t

val get_template_spec_info : t -> template_spec_info option

val is_objc_protocol : t -> bool

val is_objc_class : t -> bool

val is_objc_block : t -> bool

val is_hack_class : t -> bool

val is_python_class : t -> bool

module C : sig ... end

module CSharp : sig ... end

module Hack : sig ... end

module Java : sig ... end

module Cpp : sig ... end

module Objc : sig ... end

module Set : IStdlib.PrettyPrintable.PPSet with type elt = t

module Map : IStdlib.PrettyPrintable.PPMap with type key = t

module Hash : IStdlib.IStd.Caml.Hashtbl.S with type key = t