Module IStdlib.HashNormalizer

module type NormalizedT = sig ... end

a hashed type with a normalization function which respects equality

module type S = sig ... end

normalizer module which uses a hashtable to store normalized representatives

module Make (T : NormalizedT) : S with type t = T.t

module String : S with type t = string

normalizer for strings, lists are recursively normalized

module Int64 : S with type t = int64

val reset_all_normalizers : unit -> unit

reset hashtables in all normalizers made with Make

val register_reset : (unit -> unit) -> unit

register a function to run when reset_all_normalizers is run