Module IStdlib.HashNormalizer

module type NormalizedT = sig ... end

a hashed type with a normalization function which respects equality

module type S = sig ... end

normalizer module which uses a hashtable to store normalized representatives

module Make : functor (T : NormalizedT) -> S with type t = T.t

module StringNormalizer : S with type t = string

normalizer for strings