Module Integration

module Ant : sig ... end

module Buck : sig ... end

module BuckBuildReport : sig ... end

module BuckFlavors : sig ... end

module BuckJavaFlavor : sig ... end

module BxlCapture : sig ... end

module CaptureCompilationDatabase : sig ... end

module CaptureManipulation : sig ... end

module CaptureSILJson : sig ... end

Capture module for the json analysis in the capture phase

module Clang : sig ... end

module CompilationDatabase : sig ... end

module ConfigImpactIssuesTest : sig ... end

module ConfigImpactPostProcess : sig ... end

module CostIssuesTest : sig ... end

module DatalogFacts : sig ... end

module Differential : sig ... end

module DifferentialFilters : sig ... end

module Driver : sig ... end

entry points for top-level functionalities such as capture under various build systems, analysis, and reporting

module Erlang : sig ... end

module Gradle : sig ... end

module Hack : sig ... end

module Help : sig ... end

module InferCommandImplementation : sig ... end

implementations of infer commands

module IssuesTest : sig ... end

module Javac : sig ... end

module JsonReports : sig ... end

module Kotlinc : sig ... end

module Maven : sig ... end

module MergeCapture : sig ... end

module NdkBuild : sig ... end

module Python : sig ... end

module ReportDataFlows : sig ... end

module ReportDiff : sig ... end

module ReportLineage : sig ... end

module SarifReport : sig ... end

module StatsDiff : sig ... end

module TextReport : sig ... end

module TraceBugs : sig ... end

module XMLReport : sig ... end

module XcodeBuild : sig ... end