Module Integration.CompilationDatabase

type t

type compilation_data = {

directory : string;

executable : string;

escaped_arguments : string list;
argument list, where each argument is already escaped for the shell. This is because in some cases the argument list contains arguments that are actually themselves a list of arguments, for instance because the compilation database only contains a "command" entry.

}

val filter_compilation_data : t -> f:(IBase.SourceFile.t -> bool) -> (IBase.SourceFile.t * compilation_data) list

val from_json_files : [< `Escaped of string | `Raw of string ] list -> t