Module Integration.Javac

type compiler =

| Java

| Javac

val compare_compiler : compiler -> compiler -> int

val call_infer_javac_capture : javac_args:string list -> unit

perform a javac catpure given args to javac, this will shell out to 'infer capture -- javac

val capture : compiler -> prog:string -> args:string list -> unit

perform capture when given prog and args, this is the entrypoint for infer being called with 'capture -- javac'