Module JavaFrontend.JClasspath

type file_entry =

| Singleton of IBase.SourceFile.t

| Duplicate of (string list * IBase.SourceFile.t) list

map entry for source files with potential basename collision within the same compiler call

type t = {

classpath_channel : Javalib_pack.Javalib.class_path;

sources : file_entry IStdlib.IStd.String.Map.t;

classes : Javalib_pack.JBasics.ClassSet.t;

}

type source =

| FromVerboseOut of {

verbose_out_file : string;

}
(*
load the list of source files and the list of classes from the javac verbose file
*)

| FromArguments of {

path : string;

sources : string list;

}
(*
load the list of source files and the list of classes from Config.generated_classes
*)

val with_classpath : f:(t -> unit) -> source -> unit

load a class path, pass it to f and cleanup after f is done