Module AssignmentRule.ProvisionalViolation

Violation that will occur if the provisional annotation becomes real @Nullable

type t

val offending_annotations : t -> ProvisionalAnnotation.t list

Non-empty list of corresponding provisional annotations (adding any of those will lead to an issue)

val fix_annotation : t -> ProvisionalAnnotation.t option

If there is a place such as adding @Nullable will fix the issue, this is the one.

val from : violation -> t option

If the violation is provisional (so is not real but will become real when the annotation is added), create it.