Module Nullsafe.InheritanceRule

type violation

val compare_violation : violation -> violation -> int

type type_role =

| Param

| Ret

val check : type_role -> base:Nullability.t -> overridden:Nullability.t -> (unit, violation) IStdlib.IStd.result

See description of the rule in the header of the file. Note that formal fact of violation might or might not be reported to the user, depending on the mode. See to_reportable_violation

module ReportableViolation : sig ... end

Violation that needs to be reported to the user.