Module PulseAbductiveDomain.AddressAttributes

attribute operations like BaseAddressAttributes but that also take care of propagating facts to the precondition

val abduce_and_add : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attributes.t -> t -> t

add the attributes to both the current state and, if meaningful, the pre

val add_one : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attribute.t -> t -> t

add the attribute only to the post

val add_attrs : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attributes.t -> t -> t

val check_valid : PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> ( t, PulseBasicInterface.Invalidation.t * PulseBasicInterface.Trace.t ) IStdlib.IStd.result

val check_initialized : PathContext.t -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> ( t, unit ) IStdlib.IStd.result

val check_not_tainted : PathContext.t -> PulseBasicInterface.Taint.sink -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> ( t, PulseBasicInterface.Taint.source * PulseBasicInterface.ValueHistory.t ) IStdlib.IStd.result

val invalidate : (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseBasicInterface.Invalidation.t -> IBase.Location.t -> t -> t

val allocate : PulseBasicInterface.Attribute.allocator -> PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val java_resource_release : PulseBasicInterface.AbstractValue.t -> t -> t

val is_java_resource_released : PulseBasicInterface.AbstractValue.t -> t -> bool

val add_dynamic_type : IR.Typ.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val add_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> t

val is_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> bool

val remove_allocation_attr : PulseBasicInterface.AbstractValue.t -> t -> t

val get_allocation : PulseBasicInterface.AbstractValue.t -> t -> (PulseBasicInterface.Attribute.allocator * PulseBasicInterface.Trace.t) option

val get_closure_proc_name : PulseBasicInterface.AbstractValue.t -> t -> IR.Procname.t option

val get_copied_var : PulseBasicInterface.AbstractValue.t -> t -> IR.Var.t option

val get_source_origin_of_copy : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.AbstractValue.t option

val is_end_of_collection : PulseBasicInterface.AbstractValue.t -> t -> bool

val mark_as_end_of_collection : PulseBasicInterface.AbstractValue.t -> t -> t

val is_std_vector_reserved : PulseBasicInterface.AbstractValue.t -> t -> bool

val std_vector_reserve : PulseBasicInterface.AbstractValue.t -> t -> t

val add_unreachable_at : PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val find_opt : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attributes.t option

val check_valid_isl : PathContext.t -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> ?null_noop:bool -> t -> ( t, [> `ISLError of t | `InvalidAccess of PulseBasicInterface.AbstractValue.t * PulseBasicInterface.Invalidation.t * PulseBasicInterface.Trace.t * t ] ) IStdlib.IStd.result list