Module PulseAbductiveDomain.AddressAttributes

attribute operations like PulseBaseAddressAttributes but that also take care of propagating facts to the precondition

val abduce_attribute : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attribute.t -> t -> t

add the attribute to the pre, if the address is in pre

val abduce_and_add : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attributes.t -> t -> t

add the attributes to both the current state and, if meaningful, the pre

val add_one : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attribute.t -> t -> t

add the attribute only to the post

val add_attrs : PulseBasicInterface.AbstractValue.t -> PulseBasicInterface.Attributes.t -> t -> t

val check_valid : PathContext.t -> ?must_be_valid_reason:PulseBasicInterface.Invalidation.must_be_valid_reason -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> ( t, PulseBasicInterface.Invalidation.t * PulseBasicInterface.Trace.t ) IStdlib.IStd.result

val check_initialized : PathContext.t -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> ( t, unit ) IStdlib.IStd.result

val add_taint_sink : PathContext.t -> PulseBasicInterface.Taint.t -> PulseBasicInterface.Trace.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val invalidate : (PulseBasicInterface.AbstractValue.t * PulseBasicInterface.ValueHistory.t) -> PulseBasicInterface.Invalidation.t -> IBase.Location.t -> t -> t

val always_reachable : PulseBasicInterface.AbstractValue.t -> t -> t

val allocate : PulseBasicInterface.Attribute.allocator -> PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val java_resource_release : PulseBasicInterface.AbstractValue.t -> t -> t

val is_java_resource_released : PulseBasicInterface.AbstractValue.t -> t -> bool

val csharp_resource_release : PulseBasicInterface.AbstractValue.t -> t -> t

val is_csharp_resource_released : PulseBasicInterface.AbstractValue.t -> t -> bool

val add_dynamic_type : IR.Typ.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val add_dynamic_type_source_file : IR.Typ.t -> IBase.SourceFile.t -> PulseBasicInterface.AbstractValue.t -> t -> t

val add_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> t

val is_ref_counted : PulseBasicInterface.AbstractValue.t -> t -> bool

val remove_allocation_attr : PulseBasicInterface.AbstractValue.t -> t -> t

val remove_taint_attrs : PulseBasicInterface.AbstractValue.t -> t -> t

val get_dynamic_type : PulseBasicInterface.AbstractValue.t -> t -> IR.Typ.t option

val get_dynamic_type_source_file : PulseBasicInterface.AbstractValue.t -> t -> (IR.Typ.t * IBase.SourceFile.t option) option

val get_allocation : PulseBasicInterface.AbstractValue.t -> t -> (PulseBasicInterface.Attribute.allocator * PulseBasicInterface.Trace.t) option

val get_closure_proc_name : PulseBasicInterface.AbstractValue.t -> t -> IR.Procname.t option

val get_copied_into : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attribute.CopiedInto.t option

val get_copied_return : PulseBasicInterface.AbstractValue.t -> t -> (PulseBasicInterface.AbstractValue.t * bool * PulseBasicInterface.Attribute.CopyOrigin.t * IBase.Location.t) option

val remove_copied_return : PulseBasicInterface.AbstractValue.t -> t -> t

val get_source_origin_of_copy : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.AbstractValue.t option

val get_taint_sources_and_sanitizers : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attribute.TaintedSet.t * PulseBasicInterface.Attribute.TaintSanitizedSet.t

val get_propagate_taint_from : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attribute.taint_in list option

val is_end_of_collection : PulseBasicInterface.AbstractValue.t -> t -> bool

val mark_as_end_of_collection : PulseBasicInterface.AbstractValue.t -> t -> t

val is_std_vector_reserved : PulseBasicInterface.AbstractValue.t -> t -> bool

val std_vector_reserve : PulseBasicInterface.AbstractValue.t -> t -> t

val add_unreachable_at : PulseBasicInterface.AbstractValue.t -> IBase.Location.t -> t -> t

val add_copied_return : PulseBasicInterface.AbstractValue.t -> source:PulseBasicInterface.AbstractValue.t -> is_const_ref:bool -> PulseBasicInterface.Attribute.CopyOrigin.t -> IBase.Location.t -> t -> t

val get_config_usage : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attribute.ConfigUsage.t option

val get_const_string : PulseBasicInterface.AbstractValue.t -> t -> string option

val find_opt : PulseBasicInterface.AbstractValue.t -> t -> PulseBasicInterface.Attributes.t option