Module Pulselib.PulseAttribute

module F = Stdlib.Format

module CallEvent = PulseCallEvent

module Invalidation = PulseInvalidation

module Timestamp = PulseTimestamp

module Trace = PulseTrace

module ValueHistory = PulseValueHistory

type allocator =

| CMalloc

| CustomMalloc of IR.Procname.t

| CRealloc

| CustomRealloc of IR.Procname.t

| CppNew

| CppNewArray

val equal_allocator : allocator -> allocator -> bool

val pp_allocator : F.formatter -> allocator -> unit

type t =

| AddressOfCppTemporary of IR.Var.t * ValueHistory.t

| AddressOfStackVariable of IR.Var.t * IBase.Location.t * ValueHistory.t

| Allocated of allocator * Trace.t

| Closure of IR.Procname.t

| DynamicType of IR.Typ.t

| EndOfCollection

| Invalid of Invalidation.t * Trace.t

| ISLAbduced of Trace.t	(*
The allocation is abduced so as the analysis could run normally
*)

| MustBeInitialized of Timestamp.t * Trace.t

| MustBeValid of Timestamp.t * Trace.t * Invalidation.must_be_valid_reason option

| StdVectorReserve

| Uninitialized

| UnknownEffect of CallEvent.t * ValueHistory.t	(*
generated by calls to unknown functions to remember that a pointer has been passed to an unknown function and so everything reachable from it has potentially been affected in unknown ways
*)

| UnreachableAt of IBase.Location.t	(*
temporary marker to remember where a variable became unreachable; helps with accurately reporting leaks
*)

| WrittenTo of Trace.t

val compare : t -> t -> int

val pp : F.formatter -> t -> unit

val is_suitable_for_pre : t -> bool

val is_suitable_for_post : t -> bool

module Attributes : sig ... end