Module Pulselib.PulseBasicInterface

Basic Pulse modules that are safe to use in any module

module AbstractValue = PulseAbstractValue

module Access = PulseAccess

module AccessSet = PulseAccess.Set

module Attribute = PulseAttribute

module Attributes = PulseAttribute.Attributes

module CallEvent = PulseCallEvent

module Formula = PulseFormula

module Invalidation = PulseInvalidation

module SatUnsat = PulseSatUnsat

module SkippedCalls = PulseSkippedCalls

module TaintConfig = PulseTaintConfig

module TaintItem = PulseTaintItem

module Timestamp = PulseTimestamp

module Trace = PulseTrace

module TransitiveInfo = PulseTransitiveInfo

module ValueHistory = PulseValueHistory

module CellId = ValueHistory.CellId

module ValueOrigin = PulseValueOrigin

include module type of struct include SatUnsat.Types end

for opening to get Sat and Unsat in the namespace

for opening to get Sat and Unsat in the namespace

type nonrec 'a sat_unsat_t = 'a PulseSatUnsat.t =

| Unsat

| Sat of 'a

Enforce short form usage

include sig ... end

module PulseAbstractValue = PulseAbstractValue

module PulseAttribute = PulseAttribute

module PulseCallEvent = PulseCallEvent

module PulseInvalidation = PulseInvalidation

module PulseTimestamp = PulseTimestamp

module PulseSkippedCalls = PulseSkippedCalls

module PulseTaintConfig = PulseTaintConfig

module PulseTaintItem = PulseTaintItem

module PulseTrace = PulseTrace

module PulseValueHistory = PulseValueHistory

module PulseSatUnsat = PulseSatUnsat