Module Pulselib.PulseInterproc

type call_state

type contradiction = private

| Aliasing of {

addr_caller : Pulselib.PulseBasicInterface.AbstractValue.t;

addr_callee : Pulselib.PulseBasicInterface.AbstractValue.t;

addr_callee' : Pulselib.PulseBasicInterface.AbstractValue.t;

call_state : call_state;

}
(*
raised when the precondition and the current state disagree on the aliasing, i.e. some addresses callee_addr and callee_addr' that are distinct in the pre are aliased to a single address caller_addr in the caller's current state. Typically raised when calling foo(z,z) where the spec for foo(x,y) says that x and y are disjoint.
*)

| DynamicTypeNeeded of IR.Pvar.Set.t

| CapturedFormalActualLength of {

captured_formals : (IR.Var.t * IR.Typ.t) list;

captured_actuals : ((Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list;

}

| FormalActualLength of {

formals : (IR.Var.t * IR.Typ.t) list;

actuals : ((Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list;

}

| PathCondition

val is_aliasing_contradiction : contradiction -> bool

val is_dynamic_type_needed_contradiction : contradiction -> IR.Pvar.Set.t option

val apply_summary : Pulselib.PulseDomainInterface.PathContext.t -> IR.Procname.t -> IBase.Location.t -> callee_summary:Pulselib.PulseDomainInterface.AbductiveDomain.Summary.t -> captured_formals:(IR.Var.t * IR.Typ.t) list -> captured_actuals: ((Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list -> formals:(IR.Var.t * IR.Typ.t) list -> actuals: ((Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) * IR.Typ.t) list -> Pulselib.PulseDomainInterface.AbductiveDomain.t -> (Pulselib.PulseDomainInterface.AbductiveDomain.t * (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) option * (Pulselib.PulseBasicInterface.AbstractValue.t * Pulselib.PulseBasicInterface.ValueHistory.t) Pulselib.PulseBasicInterface.AbstractValue.Map.t) Pulselib.PulseDomainInterface.AccessResult.t Pulselib.PulseBasicInterface.SatUnsat.t * contradiction option

result of applying one pre/post pair of a callee's summary:

PulseSatUnsat.t.Unsat if that path in the callee is infeasible

otherwise, there can be an error detected

otherwise, the result is a new abstract state, an optional return value, and a substitution callee_abstract_value -> caller_abstract_value mapping callee's abstract values to what they became in the new (caller) state