Module Pulselib.PulseLatentIssue

module AbductiveDomain = PulseAbductiveDomain

module Diagnostic = PulseDiagnostic

A subset of PulseDiagnostic that can be "latent", i.e. there is a potential issue in the code but we want to delay reporting until we see the conditions for the bug manifest themselves in some calling context.

type t =

| AccessToInvalidAddress of Diagnostic.access_to_invalid_address

| ErlangError of Diagnostic.erlang_error

| ReadUninitializedValue of Diagnostic.read_uninitialized_value

val compare : t -> t -> int

val equal : t -> t -> bool

val yojson_of_t : t -> Ppx_yojson_conv_lib.Yojson.Safe.t

val to_diagnostic : t -> Diagnostic.t

val should_report : AbductiveDomain.summary -> Diagnostic.t -> [> `DelayReport of t | `ReportNow ]

val add_call : (PulseBasicInterface.CallEvent.t * IBase.Location.t) -> t -> t