Module Pulselib.PulseMutualRecursion

module F = Stdlib.Format

type t

the trace represents the start of a cycle but isn't a cycle yet, it's a sequence of calls that end where ondemand returned no summary due to mutual recursion; these proto-cycles get bubbled the cyclic call stack until the cycle is closed again

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val mk : IBase.Location.t -> IR.Procname.t -> t

a trace of length 1

val get_inner_call : t -> IR.Procname.t

the "initial", innermost call in the cycle

val get_outer_location : t -> IBase.Location.t

the location of the first call in the cycle

val add_call : IR.Procname.t -> IBase.Location.t -> t -> t

val iter_rotations : t -> f:(t -> unit) -> unit

val pp : F.formatter -> t -> unit

val get_error_message : t -> string

val to_errlog : t -> Absint.Errlog.loc_trace

module Set : IStdlib.PrettyPrintable.PPSet with type elt = t