Module PulseOperations.Import

For opening in other modules.

type access_mode =

| Read

| Write

| NoAccess
The initialized-ness of the address is not checked when it evaluates a heap address without actual memory access, for example, when evaluating &x.f we need to check initialized-ness of x, not that of x.f.

Imported types for ease of use and so we can write variants without the corresponding module prefix

type 'abductive_domain_t base_t = 'abductive_domain_t PulseDomainInterface.ExecutionDomain.base_t =

| ContinueProgram of 'abductive_domain_t

| ExitProgram of PulseDomainInterface.AbductiveDomain.summary

| AbortProgram of PulseDomainInterface.AbductiveDomain.summary

| LatentAbortProgram of {

astate : PulseDomainInterface.AbductiveDomain.summary;

latent_issue : PulseDomainInterface.LatentIssue.t;

}

| ISLLatentMemoryError of 'abductive_domain_t

type 'a access_result = 'a PulseReport.access_result

Monadic syntax

include module type of IStdlib.IResult.Let_syntax

include module type of IStdlib.IStd.Result.Monad_infix

val (>>=) : ('a, 'e) Core_kernel__Result.t -> ('a -> ('b, 'e) Core_kernel__Result.t) -> ('b, 'e) Core_kernel__Result.t

val (>>|) : ('a, 'e) Core_kernel__Result.t -> ('a -> 'b) -> ('b, 'e) Core_kernel__Result.t

val let+ : ('ok, 'err) IStdlib.IStd.result -> ('ok -> 'okk) -> ('okk, 'err) IStdlib.IStd.result

val let* : ('ok, 'err) IStdlib.IStd.result -> ('ok -> ('okk, 'err) IStdlib.IStd.result) -> ('okk, 'err) IStdlib.IStd.result

val let<*> : 'a access_result -> ('a -> 'b access_result list) -> 'b access_result list

monadic "bind" but not really that turns an access_result into a list of access_results (not really because the first type is not an access_result list but just an access_result)

val let<+> : 'a access_result -> ('a -> 'abductive_domain_t) -> 'abductive_domain_t base_t access_result list

monadic "map" but even less really that turns an access_result into an analysis result