Module Pulselib.PulseTimestamp

module F = Stdlib.Format

type t = private int

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val t0 : t

val pp : F.formatter -> t -> unit

val incr : t -> t

type trace = private t list

A trace of timestamps: a way of recording evolution of a timestamp over the course of analysis. In particular, this can permit lightweight comparison of histories across interprocedural calls.

val compare_trace : trace -> trace -> int

val equal_trace : trace -> trace -> bool

val trace0 : t -> trace

val pp_trace : F.formatter -> trace -> unit

val add_to_trace : trace -> t -> trace