Module PythonFrontend.PyCommon

module SMap : IStdlib.IStd.Caml.Map.S with type key = string

module SSet : IStdlib.IStd.Caml.Set.S with type elt = string

module IMap : IStdlib.IStd.Caml.Map.S with type key = int

val proc_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.ProcName.t

val var_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.VarName.t

val node_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.NodeName.t

val field_name : ?loc:Textuallib.Textual.Location.t -> string -> Textuallib.Textual.FieldName.t

val mk_type : string -> Textuallib.Textual.Typ.t

val qualified_procname : enclosing_class:Textuallib.Textual.TypeName.t -> Textuallib.Textual.ProcName.t -> Textuallib.Textual.QualifiedProcName.t

val builtin_name : string -> Textuallib.Textual.QualifiedProcName.t

Helper function to encode known builtin names correctly

val python_int : Textuallib.Textual.QualifiedProcName.t

Encoding of Python int type. Since Python integers are of arbitrary precision, they are not modeled directly with int.

val python_float : Textuallib.Textual.QualifiedProcName.t

Encoding of Python float type.

val python_string : Textuallib.Textual.QualifiedProcName.t

Encoding of Python str type.

val python_bytes : Textuallib.Textual.QualifiedProcName.t

Encoding of Python bytes type.

val python_bool : Textuallib.Textual.QualifiedProcName.t

Encoding of Python bool type.

val python_tuple : Textuallib.Textual.QualifiedProcName.t

Encoding of Python tuple type. It is the raw "untyped" one where every item is of type object.

val pyObject : Textuallib.Textual.Typ.t

object is the top type of Python. It helps us when no type information is available.

val pyCode : Textuallib.Textual.Typ.t

code is a builtin Python type to describe any code (function, class, ...) object

val pyClass : Textuallib.Textual.Typ.t

Like pyCode but for class creation

val pyIterItem : Textuallib.Textual.Typ.t

Wrapper type to implement next access on an iterator. Declared as pyIterItemStruct

val py_iter_item_has_item : Textuallib.Textual.qualified_fieldname

val py_iter_item_next_item : Textuallib.Textual.qualified_fieldname

val pyIterItemStruct : Textuallib.Textual.Struct.t

Textual structure with two fields. has_item indicates if something was read from the iterator or if it was empty. If has_item is true, next_item holds the value read from the iterator.

val pyInt : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type int

val pyString : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type str

val pyBytes : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type bytes

val pyBool : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type bool

val pyFloat : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type float

val pyNone : Textuallib.Textual.Typ.t

Textual encoding of the primitive Python type None

val pyList : Textuallib.Textual.Typ.t

Python's builtin list type

val pyMap : Textuallib.Textual.Typ.t

Python's builtin map type

val pySet : Textuallib.Textual.Typ.t

Python's builtin set type

val pyTuple : Textuallib.Textual.Typ.t

Python's builtin tuple type

val mk_int : Z.t -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal integers.

val mk_float : float -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal floats.

val mk_string : string -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal strings.

val get_string : Textuallib.Textual.Exp.t -> string option

Helper to get back a string built with mk_string

val get_tuple_as_list : Textuallib.Textual.Exp.t -> Textuallib.Textual.Exp.t list option

Helper to get back the inner elements of a python_tuple

val mk_bytes : bytes -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for literal bytes.

val mk_bool : bool -> Textuallib.Textual.Exp.t

Helper function to define typed Textual expression for boolean.

module Ident : sig ... end

type annotated_name = {

name : string;

annotation : Ident.t;

}

Encoding of some type annotation like x: int

val pp_annotated_name : Stdlib.Format.formatter -> annotated_name -> unit

type signature = annotated_name list

val pp_signature : Stdlib.Format.formatter -> signature -> unit

val toplevel_function : string

val static_method : string

val static_companion : string -> string

module ABC : sig ... end

val init__ : string

Name of the Python initialization method

val new__ : string

Name of the Python instantiation method

val exit : string

Name of the Python context exit method

val enter : string

Name of the Python context enter method

val return : string

Name of the binding name of return types of functions / methods

val class__ : string

Special name for __class__

val classcell : string

Special name for __classcell__

val annotations : string

Special name for __annotations__

val entry : string

Textual label name for entry points of function we synthesized (constructors, __init__, ...)

val self : string

Common name of the self argument in Python

module MakeFunctionFlags : sig ... end

Flags used by MAKE_FUNCTION