Module Textuallib.TextualSil

val proc_decl_to_sil : Textual.Lang.t -> Textual.ProcDecl.t -> IR.Procname.t

val module_to_sil : Textual.Module.t -> IR.Cfg.t * IR.Tenv.t

val from_java : filename:string -> IR.Tenv.t -> IR.Cfg.t -> unit

generate a .sil file with name filename containing all the functions in the given cfg

val dump_module : filename:string -> Textual.Module.t -> unit

generate a .sil file with name filename with all the content of the input module

val default_return_type : Textual.Lang.t option -> Textual.Location.t -> Textual.Typ.t

val hack_dict_type_name : IR.Typ.name

val hack_dict_iter_type_name : IR.Typ.name

val hack_vec_type_name : IR.Typ.name

val hack_vec_iter_type_name : IR.Typ.name

val hack_bool_type_name : IR.Typ.name

val hack_int_type_name : IR.Typ.name

val hack_float_type_name : IR.Typ.name

val hack_string_type_name : IR.Typ.name

val hack_splated_vec_type_name : IR.Typ.name

val hack_mixed_type_name : IR.Typ.name

val hack_awaitable_type_name : IR.Typ.name

val hack_mixed_static_companion_type_name : IR.Typ.name

val hack_builtins_type_name : IR.Typ.name

val hack_root_type_name : IR.Typ.name

val wildcard_sil_fieldname : Textual.Lang.t -> string -> IR.Fieldname.t

val textual_ext : string

val to_filename : string -> string